% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEPRECATED.R
\name{BinaryTransformation-deprecated}
\alias{BinaryTransformation-deprecated}
\title{BinaryTransformation}
\usage{
BinaryTransformation(...)
}
\arguments{
\item{...}{Additional arguments}
}
\description{
Deprecated function name for
  \code{\link{bm_BinaryTransformation}}
}
\seealso{
\code{\link{biomod2-deprecated}}
}
\keyword{internal}
