% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomod2_data.R
\docType{data}
\name{bioclim_future}
\alias{bioclim_future}
\title{Bioclimatic variables for SDM based on future condition}
\format{
A RasterStack with 5 layers:
\describe{
  \item{bio3}{Isothermality}
  \item{bio4}{Temperature Seasonality}
  \item{bio7}{Temperature Annual Range}
  \item{bio11}{Mean Temperature of Coldest Quarter}
  \item{bio12}{Annual Precipitation}
}
}
\usage{
bioclim_future
}
\description{
A RasterStack with 5 bioclimatic variables commonly used for SDM and
describing future climate based on  CMIP6 downscaled projections. Additional
information available at
\href{https://www.worldclim.org/data/cmip6/cmip6climate.html}{worldclim}
}
\keyword{datasets}
