% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomod2_classes_3.R
\docType{class}
\name{BIOMOD.ensemble.models.out}
\alias{BIOMOD.ensemble.models.out}
\title{\code{BIOMOD_EnsembleModeling()} output object class}
\description{
Class returned by \code{\link{BIOMOD_EnsembleModeling}}, and used by 
\code{\link{BIOMOD_LoadModels}}, \code{\link{BIOMOD_PresenceOnly}} and 
\code{\link{BIOMOD_EnsembleForecasting}}
}
\section{Slots}{

\describe{
\item{\code{sp.name}}{a \code{character} corresponding to the species name}

\item{\code{expl.var.names}}{a \code{vector} containing names of explanatory variables}

\item{\code{models.out.obj}}{a \code{\link{BIOMOD.stored.models.out}} object containing 
informations from \code{\link{BIOMOD_Modeling}} object}

\item{\code{eval.metric}}{a \code{vector} containing names of evaluation metrics}

\item{\code{eval.metric.quality.threshold}}{a \code{vector} of \code{numeric} values 
corresponding to the minimum scores (one for each \code{eval.metric}) below which single 
models has been excluded from the ensemble model building}

\item{\code{em.computed}}{a \code{vector} containing names of ensemble models}

\item{\code{em.by}}{a \code{character} corresponding to the way kept models have been combined to 
build the ensemble models, must be among \code{PA_dataset+repet}, \code{PA_dataset+algo}, 
\code{PA_dataset}, \code{algo}, \code{all}}

\item{\code{em.models}}{a \code{list} containing ensemble models}

\item{\code{modeling.id}}{a \code{character} corresponding to the name (ID) of the simulation set}

\item{\code{link}}{a \code{character} containing the file name of the saved object}
}}

\examples{

showClass("BIOMOD.ensemble.models.out")

## ------------------------------------------------------------------------
# species occurrences
myFile <- system.file("external/species/mammals_table.csv", package="biomod2")
DataSpecies <- read.csv(myFile, row.names = 1)
head(DataSpecies)

# the name of studied species
myRespName <- 'GuloGulo'

# the presence/absences data for our species
myResp <- as.numeric(DataSpecies[, myRespName])

# the XY coordinates of species data
myRespXY <- DataSpecies[, c("X_WGS84", "Y_WGS84")]


# Environmental variables extracted from BIOCLIM (bio_3, bio_4, bio_7, bio_11 & bio_12)
myFiles = paste0("external/bioclim/current/bio", c(3, 4, 7, 11, 12), ".grd")
myExpl = raster::stack(system.file(myFiles[1], package = "biomod2"),
                       system.file(myFiles[2], package = "biomod2"),
                       system.file(myFiles[3], package = "biomod2"),
                       system.file(myFiles[4], package = "biomod2"),
                       system.file(myFiles[5], package = "biomod2"))

# 1. Formating Data
myBiomodData <- BIOMOD_FormatingData(resp.var = myResp,
                                     expl.var = myExpl,
                                     resp.xy = myRespXY,
                                     resp.name = myRespName)

# 2. Defining Models Options using default options.
myBiomodOption <- BIOMOD_ModelingOptions()

# 3. Doing Modelisation
myBiomodModelOut <- BIOMOD_Modeling(myBiomodData,
                                    models = c('SRE','RF'),
                                    models.options = myBiomodOption,
                                    NbRunEval = 2,
                                    DataSplit = 80,
                                    VarImport = 3,
                                    models.eval.meth = c('TSS', 'ROC'),
                                    do.full.models = FALSE,
                                    modeling.id = 'test')
                    
# 4. Doing Ensemble Modeling
myBiomodEM <- BIOMOD_EnsembleModeling(modeling.output = myBiomodModelOut,
                                      chosen.models = 'all',
                                      em.by = 'all',
                                      eval.metric = c('TSS'),
                                      eval.metric.quality.threshold = c(0.7),
                                      models.eval.meth = c('TSS', 'ROC'),
                                      prob.mean = TRUE,
                                      prob.median = FALSE,
                                      prob.cv = FALSE,
                                      prob.ci = FALSE,
                                      prob.ci.alpha = 0.05,
                                      committee.averaging = FALSE,
                                      prob.mean.weight = TRUE,
                                      prob.mean.weight.decay = 'proportional')

# print summary
myBiomodEM


}
\seealso{
\code{\link{BIOMOD_EnsembleModeling}}, \code{\link{BIOMOD_LoadModels}}, 
\code{\link{BIOMOD_PresenceOnly}}, \code{\link{bm_VariablesImportance}}, 
\code{\link{bm_PlotEvalMean}}, \code{\link{bm_PlotEvalBoxplot}}, 
\code{\link{bm_PlotVarImpBoxplot}}, \code{\link{bm_PlotResponseCurves}}

Other Toolbox objects: 
\code{\link{BIOMOD.formated.data.PA}},
\code{\link{BIOMOD.formated.data}},
\code{\link{BIOMOD.model.options}},
\code{\link{BIOMOD.models.out}},
\code{\link{BIOMOD.projection.out}},
\code{\link{BIOMOD.stored.data}},
\code{\link{biomod2_ensemble_model}},
\code{\link{biomod2_model}}
}
\author{
Damien Georges
}
\concept{Toolbox objects}
