% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomod2_classes_1.R
\docType{class}
\name{BIOMOD.formated.data}
\alias{BIOMOD.formated.data}
\alias{BIOMOD.formated.data.PA}
\alias{BIOMOD.formated.data-class}
\alias{BIOMOD.formated.data-method}
\alias{plot}
\alias{show}
\title{BIOMOD.formated.data}
\description{
BIOMOD.formated.data
}
\author{
Damien Georges
}
