% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiomodClass.R
\docType{class}
\name{BIOMOD.Model.Options-class}
\alias{BIOMOD.Model.Options-class}
\alias{show,BIOMOD.Model.Options-method}
\alias{show,}
\alias{BIOMOD.Model.Options-method}
\title{BIOMOD_ModelingOptions() outputs objects class}
\usage{
\S4method{show}{BIOMOD.Model.Options}(object)
}
\description{
BIOMOD.Model.Options objects are created, used and returned
by BIOMOD functions. These objects will contains for each
model support within \pkg{biomod2}, a set of options that
users can change. Please refer to 
\code{\link[biomod2]{BIOMOD_ModelingOptions}} for further
 details. 
  
- output of: \code{\link[biomod2]{BIOMOD_ModelingOptions}}
- input of:  \code{\link[biomod2]{BIOMOD_Modeling}}
}
\examples{
showClass("BIOMOD.Model.Options")

}
\seealso{
\code{\link[biomod2]{BIOMOD_ModelingOptions}}
}
\author{
Damien Georges
}
