% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomod2_classes_1.R
\docType{class}
\name{BIOMOD.models.options}
\alias{BIOMOD.models.options}
\alias{BIOMOD.models.options-class}
\alias{show,BIOMOD.models.options-method}
\title{\code{BIOMOD_ModelingOptions()} output object class}
\usage{
\S4method{show}{BIOMOD.models.options}(object)
}
\arguments{
\item{object}{a \code{\link{BIOMOD.models.options}} object}
}
\description{
Class returned by \code{\link{BIOMOD_ModelingOptions}}, and used by 
\code{\link{BIOMOD_Tuning}} and \code{\link{BIOMOD_Modeling}}
}
\section{Slots}{

\describe{
\item{\code{GLM}}{a \code{list} containing GLM options}

\item{\code{GBM}}{a \code{list} containing GBM options}

\item{\code{GAM}}{a \code{list} containing GAM options}

\item{\code{CTA}}{a \code{list} containing CTA options}

\item{\code{ANN}}{a \code{list} containing ANN options}

\item{\code{SRE}}{a \code{list} containing SRE options}

\item{\code{FDA}}{a \code{list} containing FDA options}

\item{\code{MARS}}{a \code{list} containing MARS options}

\item{\code{RF}}{a \code{list} containing RF options}

\item{\code{MAXENT}}{a \code{list} containing MAXENT options}

\item{\code{MAXNET}}{a \code{list} containing MAXNET options}

\item{\code{XGBOOST}}{a \code{list} containing XGBOOST options}
}}

\examples{

showClass("BIOMOD.models.options")

## ----------------------------------------------------------------------- #
## default BIOMOD.models.options object
myBiomodOptions <- BIOMOD_ModelingOptions()

## print the object
myBiomodOptions


}
\seealso{
\code{\link{BIOMOD_ModelingOptions}}, \code{\link{BIOMOD_Tuning}}, 
\code{\link{BIOMOD_Modeling}}

Other Toolbox objects: 
\code{\link{BIOMOD.ensemble.models.out}},
\code{\link{BIOMOD.formated.data.PA}},
\code{\link{BIOMOD.formated.data}},
\code{\link{BIOMOD.models.out}},
\code{\link{BIOMOD.projection.out}},
\code{\link{BIOMOD.stored.data}},
\code{\link{biomod2_ensemble_model}},
\code{\link{biomod2_model}}
}
\author{
Damien Georges
}
\concept{Toolbox objects}
