% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomod2_classes_3.R
\docType{class}
\name{BIOMOD.models.out}
\alias{BIOMOD.models.out}
\alias{BIOMOD.stored.models.out}
\title{\code{BIOMOD_Modeling()} output object class}
\description{
Class returned by \code{\link{BIOMOD_Modeling}}, and used by 
\code{\link{BIOMOD_LoadModels}}, \code{\link{BIOMOD_PresenceOnly}}, 
\code{\link{BIOMOD_Projection}} and \code{\link{BIOMOD_EnsembleModeling}}
}
\section{Slots}{

\describe{
\item{\code{modeling.id}}{a \code{character} corresponding to the name (ID) of the simulation set}

\item{\code{sp.name}}{a \code{character} corresponding to the species name}

\item{\code{expl.var.names}}{a \code{vector} containing names of explanatory variables}

\item{\code{models.computed}}{a \code{vector} containing names of computed models}

\item{\code{models.failed}}{a \code{vector} containing names of failed models}

\item{\code{has.evaluation.data}}{a \code{logical} value defining whether evaluation data is given}

\item{\code{rescal.all.models}}{a \code{logical} value defining whether models have been rescaled or 
not}

\item{\code{models.evaluation}}{a \code{\link{BIOMOD.stored.array}} object containing models evaluation}

\item{\code{variables.importance}}{a \code{\link{BIOMOD.stored.array}} object containing variables 
importance}

\item{\code{models.prediction}}{a \code{\link{BIOMOD.stored.array}} object containing models 
predictions}

\item{\code{models.prediction.eval}}{a \code{\link{BIOMOD.stored.array}} object containing models 
predictions for evaluation data}

\item{\code{formated.input.data}}{a \code{\link{BIOMOD.stored.formated.data}} object containing 
informations from \code{\link{BIOMOD_FormatingData}} object}

\item{\code{calib.lines}}{a \code{\link{BIOMOD.stored.array}} object containing calibration lines}

\item{\code{models.options}}{a \code{\link{BIOMOD.stored.models.options}} object containing 
informations from \code{\link{BIOMOD_ModelingOptions}} object}

\item{\code{link}}{a \code{character} containing the file name of the saved object}

\item{\code{val}}{a \code{\link{BIOMOD.models.out}} object}
}}

\examples{

showClass("BIOMOD.models.out")
showClass("BIOMOD.stored.models.out")

## ------------------------------------------------------------------------
# species occurrences
myFile <- system.file("external/species/mammals_table.csv", package="biomod2")
DataSpecies <- read.csv(myFile, row.names = 1)
head(DataSpecies)

# the name of studied species
myRespName <- 'GuloGulo'

# the presence/absences data for our species
myResp <- as.numeric(DataSpecies[, myRespName])

# the XY coordinates of species data
myRespXY <- DataSpecies[, c("X_WGS84", "Y_WGS84")]


# Environmental variables extracted from BIOCLIM (bio_3, bio_4, bio_7, bio_11 & bio_12)
myFiles = paste0("external/bioclim/current/bio", c(3, 4, 7, 11, 12), ".grd")
myExpl = raster::stack(system.file(myFiles[1], package = "biomod2"),
                       system.file(myFiles[2], package = "biomod2"),
                       system.file(myFiles[3], package = "biomod2"),
                       system.file(myFiles[4], package = "biomod2"),
                       system.file(myFiles[5], package = "biomod2"))

# 1. Formating Data
myBiomodData <- BIOMOD_FormatingData(resp.var = myResp,
                                     expl.var = myExpl,
                                     resp.xy = myRespXY,
                                     resp.name = myRespName)

# 2. Defining Models Options using default options.
myBiomodOption <- BIOMOD_ModelingOptions()

# 3. Doing Modelisation
myBiomodModelOut <- BIOMOD_Modeling(myBiomodData,
                                    models = c('SRE','RF'),
                                    models.options = myBiomodOption,
                                    NbRunEval = 2,
                                    DataSplit = 80,
                                    VarImport = 3,
                                    models.eval.meth = c('TSS','ROC'),
                                    do.full.models = FALSE,
                                    modeling.id = 'test')

# print a summary of modeling stuff
myBiomodModelOut


}
\seealso{
\code{\link{BIOMOD_Modeling}}, \code{\link{BIOMOD_LoadModels}}, 
\code{\link{BIOMOD_PresenceOnly}}, \code{\link{BIOMOD_Projection}}, 
\code{\link{BIOMOD_EnsembleModeling}}, \code{\link{bm_VariablesImportance}}, 
\code{\link{bm_PlotEvalMean}}, \code{\link{bm_PlotEvalBoxplot}}, 
\code{\link{bm_PlotVarImpBoxplot}}, \code{\link{bm_PlotResponseCurves}}

Other Toolbox objects: 
\code{\link{BIOMOD.ensemble.models.out}},
\code{\link{BIOMOD.formated.data.PA}},
\code{\link{BIOMOD.formated.data}},
\code{\link{BIOMOD.models.options}},
\code{\link{BIOMOD.projection.out}},
\code{\link{BIOMOD.stored.data}},
\code{\link{biomod2_ensemble_model}},
\code{\link{biomod2_model}}
}
\author{
Damien Georges
}
\concept{Toolbox objects}
