% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomod2_classes_3.R
\docType{class}
\name{BIOMOD.projection.out}
\alias{BIOMOD.projection.out}
\title{\code{BIOMOD_Projection()} output object class}
\description{
Class returned by \code{\link{BIOMOD_Projection}}, and used by 
\code{\link{BIOMOD_EnsembleForecasting}}
}
\section{Slots}{

\describe{
\item{\code{proj.names}}{a \code{character} corresponding to the projection name}

\item{\code{sp.name}}{a \code{character} corresponding to the species name}

\item{\code{expl.var.names}}{a \code{vector} containing names of explanatory variables}

\item{\code{models.projected}}{a \code{vector} containing names of projected models}

\item{\code{models.scaled}}{a \code{vector} containing names of scaled models}

\item{\code{modeling.object}}{a \code{\link{BIOMOD.stored.data}} object}

\item{\code{modeling.object.id}}{a \code{character} corresponding to the name (ID) of the simulation 
set}

\item{\code{type}}{a \code{character} corresponding to the class of the \code{val} slot of the 
\code{proj} slot}

\item{\code{proj}}{a \code{\link{BIOMOD.stored.data}} object}

\item{\code{xy.coord}}{a 2-columns \code{matrix} or \code{data.frame} containing the corresponding 
\code{X} and \code{Y} coordinates used to project the species distribution model(s)}
}}

\examples{

showClass("BIOMOD.projection.out")

## ------------------------------------------------------------------------
# species occurrences
myFile <- system.file("external/species/mammals_table.csv", package="biomod2")
DataSpecies <- read.csv(myFile, row.names = 1)
head(DataSpecies)

# the name of studied species
myRespName <- 'GuloGulo'

# the presence/absences data for our species
myResp <- as.numeric(DataSpecies[, myRespName])

# the XY coordinates of species data
myRespXY <- DataSpecies[, c("X_WGS84", "Y_WGS84")]


# Environmental variables extracted from BIOCLIM (bio_3, bio_4, bio_7, bio_11 & bio_12)
myFiles = paste0("external/bioclim/current/bio", c(3, 4, 7, 11, 12), ".grd")
myExpl = raster::stack(system.file(myFiles[1], package = "biomod2"),
                       system.file(myFiles[2], package = "biomod2"),
                       system.file(myFiles[3], package = "biomod2"),
                       system.file(myFiles[4], package = "biomod2"),
                       system.file(myFiles[5], package = "biomod2"))

# 1. Formating Data
myBiomodData <- BIOMOD_FormatingData(resp.var = myResp,
                                     expl.var = myExpl,
                                     resp.xy = myRespXY,
                                     resp.name = myRespName)

# 2. Defining Models Options using default options.
myBiomodOption <- BIOMOD_ModelingOptions()

# 3. Doing Modelisation
myBiomodModelOut <- BIOMOD_Modeling(myBiomodData,
                                    models = c('SRE','RF'),
                                    models.options = myBiomodOption,
                                    NbRunEval = 2,
                                    DataSplit = 80,
                                    VarImport = 3,
                                    models.eval.meth = c('TSS','ROC'),
                                    do.full.models = FALSE,
                                    modeling.id = 'test')

# 4.1 Projecting on current environmental conditions
myBiomodProjection <- BIOMOD_Projection(modeling.output = myBiomodModelOut,
                                        new.env = myExpl,
                                        proj.name = 'current',
                                        chosen.models = 'all',
                                        binary.meth = 'TSS',
                                        compress = FALSE,
                                        build.clamping.mask = FALSE)

# print summary and plot projections
myBiomodProjection
plot(myBiomodProjection)


}
\seealso{
\code{\link{BIOMOD_Projection}}, \code{\link{BIOMOD_EnsembleForecasting}}
}
\author{
Damien Georges
}
