% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BIOMOD_RangeSize.R
\docType{class}
\name{BIOMOD.rangesize.out}
\alias{BIOMOD.rangesize.out}
\alias{BIOMOD.rangesize.out-class}
\title{\code{BIOMOD_RangeSize()} output object class}
\description{
Class returned by \code{\link{BIOMOD_RangeSize}}, and used by 
\code{\link{bm_PlotRangeSize}}
}
\section{Slots}{

\describe{
\item{\code{Compt.By.Models}}{a \code{data.frame} containing the summary of range change for each comparison}

\item{\code{Diff.By.Pixel}}{a \code{SpatRaster} or a \code{data.frame} containing a value for each point/pixel of each comparison}

\item{\code{loss.gain}}{a \code{SpatRaster} or a \code{data.frame} containg for each point/pixel 
a value indicating a loss, a gain or a stable sitatution}

\item{\code{data.type}}{a \code{character} corresponding to the data type}

\item{\code{coord}}{a 2-columns \code{matrix} or \code{data.frame} containing the corresponding 
\code{X} and \code{Y} coordinates used to project the species distribution model(s)}

\item{\code{row.names}}{A \code{vector} containing tags matching models names splitted by 
'_' character}
}}

\seealso{
Other Toolbox objects: 
\code{\link{BIOMOD.ensemble.models.out}},
\code{\link{BIOMOD.formated.data}},
\code{\link{BIOMOD.formated.data.PA}},
\code{\link{BIOMOD.models.options}},
\code{\link{BIOMOD.models.out}},
\code{\link{BIOMOD.options.dataset}},
\code{\link{BIOMOD.options.default}},
\code{\link{BIOMOD.projection.out}},
\code{\link{BIOMOD.stored.data}},
\code{\link{biomod2_ensemble_model}},
\code{\link{biomod2_model}}
}
\author{
Helene Blancheteau
}
\concept{Toolbox objects}
