% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomod2_classes_2.R
\docType{class}
\name{BIOMOD.stored.data}
\alias{BIOMOD.stored.data}
\alias{BIOMOD.stored.array}
\alias{BIOMOD.stored.data.frame}
\alias{BIOMOD.stored.raster.stack}
\alias{BIOMOD.stored.files}
\alias{BIOMOD.stored.formated.data}
\alias{BIOMOD.stored.models.options}
\title{\code{BIOMOD_EnsembleModeling()} output object class}
\description{
Classes used by \code{\link{BIOMOD_Modeling}} and 
\code{\link{BIOMOD_EnsembleModeling}} to build their output object (see 
\code{\link{BIOMOD.models.out}} objects)
}
\details{
\code{BIOMOD.stored.data} is the basic object containing the slots \code{inMemory} and 
\code{link}. \cr All listed classes below are derived from \code{BIOMOD.stored.data}, and 
contain a \code{val} slot of specific type :

\itemize{
  \item{\code{BIOMOD.stored.array} : }{\code{val} is an \code{array}}
  \item{\code{BIOMOD.stored.data.frame} : }{\code{val} is a \code{data.frame}}
  \item{\code{BIOMOD.stored.raster.stack} : }{\code{val} is a 
  \code{\link[raster]{RasterStack}}}
  \item{\code{BIOMOD.stored.files} : }{\code{val} is a \code{character}}
  \item{\code{BIOMOD.stored.formated.data} : }{\code{val} is a 
  \code{\link{BIOMOD.formated.data}} object}
  \item{\code{BIOMOD.stored.models.options} : }{\code{val} is a 
  \code{\link{BIOMOD.model.options}} object}
}
}
\section{Slots}{

\describe{
\item{\code{inMemory}}{a \code{logical} defining whether the \code{val} slot has been loaded in 
memory or not}

\item{\code{link}}{a \code{character} containing the file name of the saved \code{val} slot}

\item{\code{val}}{an object of type depending on the \code{BIOMOD.stored.[...]} class (see 
\href{BIOMOD.stored.data.html#details}{Details})}
}}

\seealso{
\code{\link{BIOMOD.formated.data}}, \code{\link{BIOMOD.model.options}}, 
\code{\link{BIOMOD_Modeling}}, \code{\link{BIOMOD_EnsembleModeling}}, 
\code{\link{BIOMOD_Projection}}, \code{\link{BIOMOD_EnsembleForecasting}}
}
\author{
Damien Georges
}
