\name{BIOMOD.stored.objects-methods}

\alias{load_stored_object}
\alias{load_stored_object,BIOMOD.stored.data-method}

\title{BIOMOD.stored.objects functions}

\description{ 
Functions to play with attributs of \code{BIOMOD.stored.objects} objects
}

\usage{
  load_stored_object(obj, ...)
}

\arguments{
  \item{obj}{ \code{BIOMOD.stored.objects} object }
  \item{...}{ extra arguments (see details) }
}

\value{
  returns the value stored within biomod2 storing object.
}

\author{Damien Georges}
