% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEPRECATED.R
\name{biomod2-deprecated}
\alias{biomod2-deprecated}
\alias{BIOMOD_CrossValidation}
\alias{BIOMOD_ModelingOptions}
\alias{BIOMOD_PresenceOnly}
\alias{BIOMOD_Tuning}
\alias{BinaryTransformation}
\alias{calculate.stat}
\alias{Find.Optim.Stat}
\alias{getStatOptimValue}
\alias{makeFormula}
\alias{models_scores_graph}
\alias{ProbDensFunc}
\alias{response.plot2}
\alias{sample.factor.levels}
\alias{sre}
\alias{.transform.outputs.list}
\alias{variables_importance}
\title{Deprecated functions in package \pkg{biomod2}.}
\usage{
BIOMOD_CrossValidation(...)

BIOMOD_ModelingOptions(...)

BIOMOD_PresenceOnly(...)

BIOMOD_Tuning(...)

BinaryTransformation(...)

calculate.stat(...)

Find.Optim.Stat(...)

getStatOptimValue(...)

makeFormula(...)

models_scores_graph(...)

ProbDensFunc(...)

response.plot2(...)

sample.factor.levels(...)

sre(...)

.transform.outputs.list(...)

variables_importance(...)
}
\description{
The functions listed below are deprecated and were removed
in the current version. When possible, alternative functions with similar
  functionality are mentioned. Help pages for deprecated functions are
  available at \code{help("<function>-deprecated")}.
}
\section{\code{BIOMOD_CrossValidation}}{

For \code{BIOMOD_CrossValidation} use \code{\link{bm_CrossValidation}}.
}

\section{\code{BIOMOD_ModelingOptions}}{

For \code{BIOMOD_ModelingOptions} use \code{\link{bm_ModelingOptions}}.
}

\section{\code{BIOMOD_PresenceOnly}}{

For \code{BIOMOD_PresenceOnly} use \code{\link{bm_FindOptimStat}}.
}

\section{\code{BIOMOD_Tuning}}{

For \code{BIOMOD_Tuning} use \code{\link{bm_Tuning}}.
}

\section{\code{BinaryTransformation}}{

For \code{BinaryTransformation} use \code{\link{bm_BinaryTransformation}}.
}

\section{\code{calculate.stat}}{

For \code{calculate.stat} use \code{\link{bm_CalculateStat}}.
}

\section{\code{Find.Optim.Stat}}{

For \code{Find.Optim.Stat} use \code{\link{bm_FindOptimStat}}.
}

\section{\code{getStatOptimValue}}{

For \code{getStatOptimValue} use \code{\link{get_optim_value}}.
}

\section{\code{makeFormula}}{

For \code{makeFormula} use \code{\link{bm_MakeFormula}}.
}

\section{\code{models_scores_graph}}{

For \code{models_scores_graph} use \code{\link{bm_PlotEvalMean}}.
}

\section{\code{ProbDensFunc}}{

For \code{ProbDensFunc} use \code{\link{bm_PlotRangeSize}}.
}

\section{\code{response.plot2}}{

For \code{response.plot2} use \code{\link{bm_PlotResponseCurves}}.
}

\section{\code{sample.factor.levels}}{

For \code{sample.factor.levels} use \code{\link{bm_SampleFactorLevels}}.
}

\section{\code{sre}}{

For \code{sre} use \code{\link{bm_SRE}}.
}

\section{\code{.transform.outputs.list}}{

For \code{.transform.outputs.list} use \code{.transform_outputs_list}.
}

\section{\code{variables_importance}}{

For \code{variables_importance} use \code{\link{bm_VariablesImportance}}.
}

\keyword{internal}
