\name{biomod2-package}
\alias{biomod2-package}
\alias{biomod2}

\docType{package}

\title{ Species Distribution Modeling within an Ensemble Forecasting framework 
}
\description{
\pkg{biomod2} is the updated object-oriented version of \pkg{BIOMOD} package.

\pkg{biomod2} offers the possibility to run 10 state-of-the-art modeling techniques to describe and model
the relationships between a given species and its environment. It is an attempt
to define the ecological niche of a particular species using environmental 
variables (temperature, precipitation, ...) with the potential use of making, 
for instance, future projections under climate and land use change scenarios.
Although it has been mostly developed for ecologists that aim to predict species distribution, 
\pkg{biomod2} can also be used to model any binomial data (for instance, gene, markers,
ecosystem...) in function of any explanatory variables. 

}
\details{
\tabular{ll}{
Package: \tab biomod2\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2012-07-01\cr
License: \tab GPL-2 \cr
}

}
\author{
Wilfried Thuiller and Damien Georges, with participation of Robin Engler 
<wilfried.thuiller@ujf-grenoble.fr>
}

\keyword{ package }
