\name{biomod2.inner-method}

\alias{Projection}
\alias{Projection,ANY,ANY,RasterStack-method}
\alias{Projection,ANY,ANY,data.frame-method}

\alias{.Projection.do.proj}
\alias{.Projection.do.proj,ANY,RasterStack-method}
\alias{.Projection.do.proj,ANY,data.frame-method}

\alias{disk.pseudo.abs.selection}
\alias{disk.pseudo.abs.selection,ANY,RasterStack-method}
\alias{disk.pseudo.abs.selection,ANY,SpatialPointsDataFrame-method}

\alias{random.pseudo.abs.selection}
\alias{random.pseudo.abs.selection,ANY,RasterStack-method}
\alias{random.pseudo.abs.selection,ANY,SpatialPointsDataFrame-method}

\alias{sre.pseudo.abs.selection}
\alias{sre.pseudo.abs.selection,ANY,RasterStack-method}
\alias{sre.pseudo.abs.selection,ANY,SpatialPointsDataFrame-method}

\alias{user.defined.pseudo.abs.selection}
\alias{user.defined.pseudo.abs.selection,ANY,RasterStack-method}
\alias{user.defined.pseudo.abs.selection,ANY,SpatialPointsDataFrame-method}

%\alias{.transform.outputs}
%\alias{.transform.outputs,array-method}
%\alias{.transform.outputs,list-method}

\alias{.Models.prepare.data}
\alias{.Models.prepare.data,BIOMOD.formated.data-method}
\alias{.Models.prepare.data,BIOMOD.formated.data.PA-method}


\title{ biomod2 internal functions }

\description{ 
biomod2 internal functions.
}


\details{
  All this functions are low level functions. Interested people can go into source code for further details.
}


\author{ Wilfried Thuiller, Damien Georges}

\seealso{ \code{\link{help}}}
