% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEPRECATED.R
\name{BIOMOD_ConvertOldRun-deprecated}
\alias{BIOMOD_ConvertOldRun-deprecated}
\title{BIOMOD_ConvertOldRun}
\usage{
BIOMOD_ConvertOldRun(...)
}
\arguments{
\item{...}{Additional arguments}
}
\description{
Deprecated function used in \pkg{biomod2} 3.5.1 to convert
results from older version into the current version
}
\seealso{
\code{\link{biomod2-deprecated}}
}
\keyword{internal}
