% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEPRECATED.R
\name{BIOMOD_CrossValidation-deprecated}
\alias{BIOMOD_CrossValidation-deprecated}
\title{BIOMOD_CrossValidation}
\usage{
BIOMOD_CrossValidation(...)
}
\arguments{
\item{...}{Additional arguments}
}
\description{
Deprecated function name for
  \code{\link{bm_CrossValidation}}
}
\seealso{
\code{\link{biomod2-deprecated}}
}
\keyword{internal}
