% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEPRECATED.R
\name{BIOMOD_cv-deprecated}
\alias{BIOMOD_cv-deprecated}
\title{BIOMOD_cv}
\usage{
BIOMOD_cv(...)
}
\arguments{
\item{...}{Additional arguments}
}
\description{
Deprecated function name for
  \code{\link{BIOMOD_CrossValidation}}
}
\seealso{
\code{\link{biomod2-deprecated}}
}
\keyword{internal}
