% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BIOMOD_EnsembleForecasting.R
\name{BIOMOD_EnsembleForecasting}
\alias{BIOMOD_EnsembleForecasting}
\title{Project ensemble species distribution models onto new environment}
\usage{
BIOMOD_EnsembleForecasting(
  EM.output,
  projection.output = NULL,
  proj.name = NULL,
  new.env = NULL,
  new.env.xy = NULL,
  chosen.models = "all",
  binary.meth = NULL,
  filtered.meth = NULL,
  compress = TRUE,
  ...
)
}
\arguments{
\item{EM.output}{a \code{\link{BIOMOD.EnsembleModeling.out}} object returned by the 
\code{\link{BIOMOD_EnsembleModeling}} function}

\item{projection.output}{a \code{\link{BIOMOD.projection.out}} object returned by the 
\code{\link{BIOMOD_Projection}} function}

\item{proj.name}{(\emph{optional, default} \code{NULL}) \cr 
If \code{projection.output = NULL}, a \code{character} corresponding to the name (ID) of the 
projection set (\emph{a new folder will be created within the simulation folder with this 
name})}

\item{new.env}{(\emph{optional, default} \code{NULL}) \cr 
If \code{projection.output = NULL}, a \code{matrix}, \code{data.frame} or 
\code{\link[raster:stack]{RasterStack}} object containing the new explanatory variables (in 
columns or layers, with names matching the variables names given to the 
\code{\link{BIOMOD_FormatingData}} function to build \code{EM.output}) that will be used to 
project the ensemble species distribution model(s)}

\item{new.env.xy}{(\emph{optional, default} \code{NULL}) \cr 
If \code{new.env} is a \code{matrix} or a \code{data.frame}, a 2-columns \code{matrix} 
containing the corresponding \code{X} and \code{Y} coordinates that will be used to project 
the ensemble species distribution model(s)}

\item{chosen.models}{a \code{vector} containing model names to be kept, must be either 
\code{all} or a sub-selection of model names}

\item{binary.meth}{(\emph{optional, default} \code{NULL}) \cr 
A \code{vector} containing evaluation metric names to be used to transform prediction values 
into binary values based on models evaluation scores obtained with the 
\code{\link{BIOMOD_Modeling}} function. Must be among \code{all} (same evaluation metrics than 
those of \code{modeling.output}) or \code{ROC}, \code{TSS}, \code{KAPPA}, \code{ACCURACY}, 
\code{BIAS}, \code{POD}, \code{FAR}, \code{POFD}, \code{SR}, \code{CSI}, \code{ETS}, 
\code{HK}, \code{HSS}, \code{OR}, \code{ORSS}}

\item{filtered.meth}{(\emph{optional, default} \code{NULL}) \cr 
A \code{vector} containing evaluation metric names to be used to transform prediction values 
into filtered values based on models evaluation scores obtained with the 
\code{\link{BIOMOD_Modeling}} function. Must be among \code{all} (same evaluation metrics than 
those of \code{modeling.output}) or \code{ROC}, \code{TSS}, \code{KAPPA}, \code{ACCURACY}, 
\code{BIAS}, \code{POD}, \code{FAR}, \code{POFD}, \code{SR}, \code{CSI}, \code{ETS}, 
\code{HK}, \code{HSS}, \code{OR}, \code{ORSS}}

\item{compress}{(\emph{optional, default} \code{TRUE}) \cr 
A \code{logical} or a \code{character} value defining whether and how objects should be 
compressed when saved on hard drive. Must be either \code{TRUE}, \code{FALSE}, \code{xz} or 
\code{gzip} (see Details)}

\item{...}{(\emph{optional, see Details})}
}
\value{
A \code{BIOMOD.projection.out} object containing models projections, or links to saved outputs.
Models projections are stored out of \R (for memory storage reasons) in \code{proj.name} folder 
created in the current working directory :
\enumerate{
  \item the output is a 4-dimensional array if \code{new.env} is a \code{matrix} or a 
  \code{data.frame}
  \item it is a \code{rasterStack} if \code{new.env} is a \code{rasterStack} (or several 
  \code{rasterLayer} objects, if \code{new.env} is too large)
  \item raw projections, as well as binary and filtered projections (if asked), are saved in 
  the \code{proj.name} folder
}
}
\description{
This function allows to project ensemble models built with the 
\code{\link[biomod2]{BIOMOD_EnsembleModeling}} function onto new environmental data (which can 
represent new areas, resolution or time scales for example).
}
\details{
###############################################################################################



If \code{chosen.models = 'all'}, projections are done for all evaluation and pseudo absences 
runs if applicable. These projections may be used later by the 
\code{\link{BIOMOD_EnsembleForecasting}} function.

If \code{build.clamping.mask = TRUE}, a raster file will be saved within the projection folder. 
This mask values will correspond to the number of variables in each pixel that are out of their 
calibration / training range, identifying locations where predictions are uncertain.

\code{...} can take the following values :
\itemize{
  \item{\code{on_0_1000} : }{a \code{logical} value defining whether \code{0 - 1} probabilities 
  are to be converted to \code{0 - 1000} scale to save memory on backup}
  \item{\code{do.stack} : }{a \code{logical} value defining whether all projections are to be 
  saved as one \code{RasterStack} object or several \code{RasterLayer} files (the default if 
  projections are too heavy to be all loaded at once in memory)}
  \item{\code{keep.in.memory} : }{a \code{logical} value defining whether all projections are 
  to be kept loaded at once in memory, or only links pointing to hard drive are to be returned}
  \item{\code{output.format} : }{a \code{character} value corresponding to the projections 
  saving format on hard drive, must be either \code{.grd}, \code{.img} or \code{.RData} (the 
  default if \code{new.env} is given as \code{matrix} or \code{data.frame})}
}
}
\examples{

# species occurrences
DataSpecies <- read.csv(system.file("external/species/mammals_table.csv", package="biomod2"), row.names = 1)
head(DataSpecies)

# the name of studied species
myRespName <- 'GuloGulo'

# the presence/absences data for our species
myResp <- as.numeric(DataSpecies[, myRespName])

# the XY coordinates of species data
myRespXY <- DataSpecies[, c("X_WGS84", "Y_WGS84")]


# Environmental variables extracted from BIOCLIM (bio_3, bio_4, bio_7, bio_11 & bio_12)
myFiles = paste0("external/bioclim/current/bio", c(3, 4, 7, 11, 12), ".grd")
myExpl = raster::stack(system.file(myFiles[1], package = "biomod2"),
                       system.file(myFiles[2], package = "biomod2"),
                       system.file(myFiles[3], package = "biomod2"),
                       system.file(myFiles[4], package = "biomod2"),
                       system.file(myFiles[5], package = "biomod2"))

# 1. Formating Data
myBiomodData <- BIOMOD_FormatingData(resp.var = myResp,
                                     expl.var = myExpl,
                                     resp.xy = myRespXY,
                                     resp.name = myRespName)

# 2. Defining Models Options using default options.
myBiomodOption <- BIOMOD_ModelingOptions()

# 3. Doing Modelisation
myBiomodModelOut <- BIOMOD_Modeling(myBiomodData,
                                    models = c('SRE','RF'),
                                    models.options = myBiomodOption,
                                    NbRunEval = 2,
                                    DataSplit = 80,
                                    VarImport = 0,
                                    models.eval.meth = c('TSS','ROC'),
                                    do.full.models = FALSE,
                                    modeling.id = 'test')
                                    
# 4. Doing Ensemble Modeling
myBiomodEM <- BIOMOD_EnsembleModeling(modeling.output = myBiomodModelOut,
                                      chosen.models = 'all',
                                      em.by = 'all',
                                      eval.metric = c('TSS'),
                                      eval.metric.quality.threshold = c(0.7),
                                      models.eval.meth = c('TSS', 'ROC'),
                                      prob.mean = TRUE,
                                      prob.median = FALSE,
                                      prob.cv = FALSE,
                                      prob.ci = FALSE,
                                      prob.ci.alpha = 0.05,
                                      committee.averaging = FALSE,
                                      prob.mean.weight = TRUE,
                                      prob.mean.weight.decay = 'proportional')

# 5. Projecting on current environmental conditions
myBiomodProjection <- BIOMOD_Projection(modeling.output = myBiomodModelOut,
                                        new.env = myExpl,
                                        proj.name = 'current',
                                        chosen.models = 'all',
                                        binary.meth = 'TSS',
                                        compress = FALSE,
                                        build.clamping.mask = FALSE)

# 6. Projecting Ensemble on current environmental conditions
myBiomodEMProjection <- BIOMOD_EnsembleForecasting(EM.output = myBiomodEM, 
                                                   projection.output = myBiomodProjection)


###############################################################################################
}
\seealso{
\code{\link{BIOMOD_FormatingData}}, \code{\link{BIOMOD_ModelingOptions}}, 
\code{\link{BIOMOD_Modeling}}, \code{\link{BIOMOD_EnsembleModeling}}
}
\author{
Wilfried Thuiller, Damien Georges, Robin Engler
}
\keyword{models,}
\keyword{projection}
