% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEPRECATED.R
\name{BIOMOD_ModelingOptions-deprecated}
\alias{BIOMOD_ModelingOptions-deprecated}
\title{BIOMOD_ModelingOptions}
\usage{
BIOMOD_ModelingOptions(...)
}
\arguments{
\item{...}{Additional arguments}
}
\description{
Deprecated function name for
  \code{\link{bm_ModelingOptions}}
}
\seealso{
\code{\link{biomod2-deprecated}}
}
\keyword{internal}
