% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BIOMOD_ModelingOptions.R
\name{BIOMOD_ModelingOptions}
\alias{BIOMOD_ModelingOptions}
\alias{bm_DefaultModelingOptions}
\title{Configure the modeling options for each selected model}
\usage{
BIOMOD_ModelingOptions(GLM = NULL,
                              GBM = NULL,
                              GAM = NULL,
                              CTA = NULL,
                              ANN = NULL,
                              SRE = NULL,
                              FDA = NULL,
                              MARS = NULL,
                              RF = NULL,
                              MAXENT.Phillips = NULL)
}
\arguments{
\item{GLM}{a \code{list} containing GLM options}

\item{GBM}{a \code{list} containing GBM options}

\item{GAM}{a \code{list} containing GAM options}

\item{CTA}{a \code{list} containing CTA options}

\item{ANN}{a \code{list} containing ANN options}

\item{SRE}{a \code{list} containing SRE options}

\item{FDA}{a \code{list} containing FDA options}

\item{MARS}{a \code{list} containing MARS options}

\item{RF}{a \code{list} containing RF options}

\item{MAXENT.Phillips}{a \code{list} containing MAXENT.Phillips options}
}
\description{
Parametrize and/or tune \pkg{biomod2}'s single models options.
}
\details{
###############################################################################################



This function allows advanced user to change some default parameters of \pkg{biomod2} inner
models. 10 single models are available within the package, and their options can be set
with this function through \code{list} objects.

The \code{\link{bm_DefaultModelingOptions}} function prints all default parameter values for
all available models. This output can be copied and pasted to be used as is (with wanted
changes) as function arguments (see Examples).

Below is the detailed list of all modifiable parameters for each available model.
}
\section{NA}{
//biodiversityinformatics.amnh.org/open_source/maxent/}{MAXENT.Phillips}}
\itemize{
  \item{\code{path_to_maxent.jar = getwd()}}{ : a \code{character} corresponding to \pkg{maxent.jar}
  file link}
  \item{\code{memory_allocated = 512}}{ : an \code{integer} corresponding to the amount of memory
  (in Mo) reserved for \code{java} to run \code{MAXENT.Phillips}, must be \code{64}, \code{128},
  \code{256}, \code{512}, \code{1024}... or \code{NULL} to use default \code{java} memory
  limitation parameter}
  \item{\code{background_data_dir}}{ : a \code{character} corresponding to directory path where
  explanatory variables are stored as \code{ASCII} files (raster format). \cr
  If specified \code{MAXENT.Phillips} will generate its own background data from explanatory
  variables rasters (as usually done in \code{MAXENT} studies). Otherwise \pkg{biomod2}
  pseudo-absences will be used (see \code{\link{BIOMOD_FormatingData}})}
  \item{\code{maximumbackground}}{ : an \code{integer} corresponding to the maximum number of
  background data to sample if the \code{background_data_dir} parameter has been set}
  \item{\code{maximumiterations = 200}}{ : an \code{integer} corresponding to the maximum number of
  iterations to do}
  \item{\code{visible = FALSE}}{ : a \code{logical} to make the \code{MAXENT} user interface available}
  \item{\code{linear = TRUE}}{ : a \code{logical} to allow linear features to be used}
  \item{\code{quadratic = TRUE}}{ : a \code{logical} to allow quadratic features to be used}
  \item{\code{product = TRUE}}{ : a \code{logical} to allow product features to be used}
  \item{\code{threshold = TRUE}}{ : a \code{logical} to allow threshold features to be used}
  \item{\code{hinge = TRUE}}{ : a \code{logical} to allow hinge features to be used}
  \item{\code{lq2lqptthreshold = 80}}{ : an \code{integer} corresponding to the number of samples
  at which product and threshold features start being used}
  \item{\code{l2lqthreshold = 10}}{ : an \code{integer} corresponding to the number of samples at
  which quadratic features start being used}
  \item{\code{hingethreshold = 15}}{ : an \code{integer} corresponding to the number of samples at
  which hinge features start being used}
  \item{\code{beta_threshold = -1.0}}{ : a \code{numeric} corresponding to the regularization
  parameter to be applied to all threshold features (\emph{negative value enables automatic setting})}
  \item{\code{beta_categorical = -1.0}}{ : a \code{numeric} corresponding to the regularization
  parameter to be applied to all categorical features (\emph{negative value enables automatic setting})}
  \item{\code{beta_lqp = -1.0}}{ : a \code{numeric} corresponding to the regularization parameter
  to be applied to all linear, quadratic and product features (\emph{negative value enables automatic setting})}
  \item{\code{beta_hinge = -1.0}}{ : a \code{numeric} corresponding to the regularization parameter
  to be applied to all hinge features (\emph{negative value enables automatic setting})}
  \item{\code{betamultiplier = 1}}{ : a \code{numeric} to multiply all automatic regularization
  parameters (\emph{higher number gives a more spread-out distribution})}
  \item{\code{defaultprevalence = 0.5}}{ : a \code{numeric} corresponding to the default prevalence
  of the species (\emph{probability of presence at ordinary occurrence points})}
}

% @section \bold{MAXENT.Tsuruoka (\code{\link[maxent]{maxent}})}
% \itemize{
%   \item{\code{l1_regularizer = 0.0}}{ : a \code{numeric} turning on L1 regularization and setting
%   the regularization parameter (\emph{a value of \code{0} will disable L1 regularization})}
%   \item{\code{l2_regularizer = 0.0}}{ : a \code{numeric} turning on L2 regularization and setting
%   the regularization parameter (\emph{a value of \code{0} will disable L2 regularization})}
%   \item{\code{use_sgd = FALSE}}{ : a \code{logical} to use SGD parameter estimation}
%   \item{\code{set_heldout = 0}}{ : an \code{integer} corresponding to the number of documents to
%   hold out (\emph{to test against and prevent overfitting, use carefully in case of dataset with
%   low number of occurrences})}
%   \item{\code{verbose = FALSE}}{ : a \code{logical} specifying whether to provide descriptive
%   output about the training process}
% }
}

\section{\bold{\href{https}{
//biodiversityinformatics.amnh.org/open_source/maxent/}{MAXENT.Phillips}}
\itemize{
  \item{\code{path_to_maxent.jar = getwd()}}{ : a \code{character} corresponding to \pkg{maxent.jar}
  file link}
  \item{\code{memory_allocated = 512}}{ : an \code{integer} corresponding to the amount of memory
  (in Mo) reserved for \code{java} to run \code{MAXENT.Phillips}, must be \code{64}, \code{128},
  \code{256}, \code{512}, \code{1024}... or \code{NULL} to use default \code{java} memory
  limitation parameter}
  \item{\code{background_data_dir}}{ : a \code{character} corresponding to directory path where
  explanatory variables are stored as \code{ASCII} files (raster format). \cr
  If specified \code{MAXENT.Phillips} will generate its own background data from explanatory
  variables rasters (as usually done in \code{MAXENT} studies). Otherwise \pkg{biomod2}
  pseudo-absences will be used (see \code{\link{BIOMOD_FormatingData}})}
  \item{\code{maximumbackground}}{ : an \code{integer} corresponding to the maximum number of
  background data to sample if the \code{background_data_dir} parameter has been set}
  \item{\code{maximumiterations = 200}}{ : an \code{integer} corresponding to the maximum number of
  iterations to do}
  \item{\code{visible = FALSE}}{ : a \code{logical} to make the \code{MAXENT} user interface available}
  \item{\code{linear = TRUE}}{ : a \code{logical} to allow linear features to be used}
  \item{\code{quadratic = TRUE}}{ : a \code{logical} to allow quadratic features to be used}
  \item{\code{product = TRUE}}{ : a \code{logical} to allow product features to be used}
  \item{\code{threshold = TRUE}}{ : a \code{logical} to allow threshold features to be used}
  \item{\code{hinge = TRUE}}{ : a \code{logical} to allow hinge features to be used}
  \item{\code{lq2lqptthreshold = 80}}{ : an \code{integer} corresponding to the number of samples
  at which product and threshold features start being used}
  \item{\code{l2lqthreshold = 10}}{ : an \code{integer} corresponding to the number of samples at
  which quadratic features start being used}
  \item{\code{hingethreshold = 15}}{ : an \code{integer} corresponding to the number of samples at
  which hinge features start being used}
  \item{\code{beta_threshold = -1.0}}{ : a \code{numeric} corresponding to the regularization
  parameter to be applied to all threshold features (\emph{negative value enables automatic setting})}
  \item{\code{beta_categorical = -1.0}}{ : a \code{numeric} corresponding to the regularization
  parameter to be applied to all categorical features (\emph{negative value enables automatic setting})}
  \item{\code{beta_lqp = -1.0}}{ : a \code{numeric} corresponding to the regularization parameter
  to be applied to all linear, quadratic and product features (\emph{negative value enables automatic setting})}
  \item{\code{beta_hinge = -1.0}}{ : a \code{numeric} corresponding to the regularization parameter
  to be applied to all hinge features (\emph{negative value enables automatic setting})}
  \item{\code{betamultiplier = 1}}{ : a \code{numeric} to multiply all automatic regularization
  parameters (\emph{higher number gives a more spread-out distribution})}
  \item{\code{defaultprevalence = 0.5}}{ : a \code{numeric} corresponding to the default prevalence
  of the species (\emph{probability of presence at ordinary occurrence points})}
}

% @section \bold{MAXENT.Tsuruoka (\code{\link[maxent]{maxent}})}
% \itemize{
%   \item{\code{l1_regularizer = 0.0}}{ : a \code{numeric} turning on L1 regularization and setting
%   the regularization parameter (\emph{a value of \code{0} will disable L1 regularization})}
%   \item{\code{l2_regularizer = 0.0}}{ : a \code{numeric} turning on L2 regularization and setting
%   the regularization parameter (\emph{a value of \code{0} will disable L2 regularization})}
%   \item{\code{use_sgd = FALSE}}{ : a \code{logical} to use SGD parameter estimation}
%   \item{\code{set_heldout = 0}}{ : an \code{integer} corresponding to the number of documents to
%   hold out (\emph{to test against and prevent overfitting, use carefully in case of dataset with
%   low number of occurrences})}
%   \item{\code{verbose = FALSE}}{ : a \code{logical} specifying whether to provide descriptive
%   output about the training process}
% }
}

\examples{

## default BIOMOD.model.option object
myBiomodOptions <- BIOMOD_ModelingOptions()

## print the object
myBiomodOptions

## you can copy a part of the print, change it and custom your options
## here we want to compute quadratic GLM and select best model with 'BIC' criterium
myBiomodOptions <- BIOMOD_ModelingOptions(
  GLM = list(type = 'quadratic',
             interaction.level = 0,
             myFormula = NULL,
             test = 'BIC',
             family = 'binomial',
             control = glm.control(epsilon = 1e-08,
                                   maxit = 1000,
                                   trace = FALSE)))

## check changes were done
myBiomodOptions

##' you can prefer to establish your own GLM formula
myBiomodOptions <- BIOMOD_ModelingOptions(
  GLM = list(myFormula = formula("Sp277 ~ bio3 +
                log(bio10) + poly(bio16,2) + bio19 + bio3:bio19")))

## check changes were done
myBiomodOptions

##' you also can directly print default parameters and then follow the same processus
bm_DefaultModelingOptions()

###############################################################################################
}
\author{
Damien Georges, Wilfried Thuiller
}
\keyword{models,}
\keyword{options}
