% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEPRECATED.R
\name{BIOMOD_presenceonly-deprecated}
\alias{BIOMOD_presenceonly-deprecated}
\title{BIOMOD_presenceonly}
\usage{
BIOMOD_presenceonly(...)
}
\arguments{
\item{...}{Additional arguments}
}
\description{
Deprecated function name for
  \code{\link{BIOMOD_PresenceOnly}}
}
\seealso{
\code{\link{biomod2-deprecated}}
}
\keyword{internal}
