% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BIOMOD_Report.R
\name{BIOMOD_Report}
\alias{BIOMOD_Report}
\alias{BIOMOD_Report.summary}
\alias{BIOMOD_Report.ODMAP}
\title{Produce summary outputs from a simulation folder}
\usage{
BIOMOD_Report(
  bm.out,
  strategy = "report",
  params.ODMAP = list(O.mod.objective = NULL, O.boundary = NULL, O.obs.type = NULL,
    O.pred.type = NULL, D.eco.level = NULL, D.samp.design = NULL)
)

BIOMOD_Report.summary(sp.name, bm.files, bm.mod, bm.ens, bm.form)

BIOMOD_Report.ODMAP(sp.name, bm.files, bm.mod, bm.ens, bm.form, params.ODMAP)
}
\arguments{
\item{bm.out}{a \code{\link{BIOMOD.formated.data}} or \code{\link{BIOMOD.formated.data.PA}} 
object returned by the \code{\link{BIOMOD_FormatingData}} function ; or a 
\code{\link{BIOMOD.models.out}} or \code{\link{BIOMOD.ensemble.models.out}} 
object that can be obtained with the \code{\link{BIOMOD_Modeling}} or 
\code{\link{BIOMOD_EnsembleModeling}} functions}

\item{strategy}{a \code{character} defining the type of summary file that will be produced, 
must be \code{report}, \code{ODMAP} or \code{code} (see Details)}

\item{params.ODMAP}{a \code{list} containing values of some ODMAP fields to be filled in 
from pre-existing choices (see Details)}
}
\value{
A ...
}
\description{
This function allows to produce summary report or ODMAP table from a 
\pkg{biomod2} simulation folder.
}
\details{
\bold{Concerning \code{ctrl.train} parameter :}

Set by default to : \cr
}
\examples{
library(terra)




}
\seealso{
\code{\link{ODMAP}}
}
\author{
Maya Gueguen
}
\concept{Primary functions}
