% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEPRECATED.R
\name{BIOMOD_tuning-deprecated}
\alias{BIOMOD_tuning-deprecated}
\title{BIOMOD_Tuning}
\usage{
BIOMOD_Tuning(...)
}
\arguments{
\item{...}{Additional arguments}
}
\description{
Deprecated function name for
  \code{\link{bm_Tuning}}
}
\seealso{
\code{\link{biomod2-deprecated}}
}
\keyword{internal}
