% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bm_BinaryTransformation.R
\name{bm_BinaryTransformation}
\alias{bm_BinaryTransformation}
\title{Convert probability values into binary values using a predefined threshold}
\usage{
bm_BinaryTransformation(data, threshold, doFiltering = FALSE)
}
\arguments{
\item{data}{a \code{vector}, a \code{matrix} or \code{data.frame}, a 
\code{\link[raster:raster]{raster}} or \code{\link[raster:stack]{RasterStack}} containing the 
data to be converted}

\item{threshold}{a \code{numeric} corresponding to the threshold used to convert the given 
data}

\item{doFiltering}{(\emph{optional, default} \code{FALSE}) \cr 
A \code{logical} value defining whether filtered data should be returned, or binary one 
(see \href{bm_BinaryTransformation.html#details}{Details})}
}
\value{
An object of the same class than \code{data} and containing either binary (\code{0} or  
\code{1}) values, or filtered values.
}
\description{
This internal \pkg{biomod2} function allows to convert probability (not necessary between 
\code{0} and \code{1}) values into binary presence-absence (\code{0} or \code{1}) values 
according to a predefined threshold (see \href{bm_BinaryTransformation.html#details}{Details}).
}
\details{
If \code{data} is a \code{vector} or \code{\link[raster:raster]{raster}}, \code{threshold} 
should be a single \code{numeric} value. \cr
If \code{data} is a \code{matrix}, \code{data.frame} or 
\code{\link[raster:stack]{RasterStack}}, \code{threshold} should be a \code{vector} containing 
as many values as the number of columns or layers contained in \code{data}. If only one 
\code{numeric} value is given, the same threshold will be applied to all columns or layers. 
\cr \cr

If \code{doFiltering = FALSE}, binary (\code{0} or \code{1}) values are returned. \cr 
If \code{doFiltering = TRUE}, values will be \emph{filtered} according to \code{threshold}, 
meaning that :
\itemize{
  \item \code{data < threshold} will return 0
  \item \code{data >= threshold } will return the actual values of \code{data} (not 
  transformed in \code{1})
}
}
\examples{

## generate 'probability' vector
xx <- rnorm(50, 10)

yy <- bm_BinaryTransformation(data = xx, threshold = 10)
yy_filt <- bm_BinaryTransformation(data = xx, threshold = 10, doFiltering = TRUE)

cbind(xx, yy, yy_filt)


}
\seealso{
\code{\link{BIOMOD_Projection}}, \code{\link{BIOMOD_EnsembleForecasting}}
}
\author{
Wilfried Thuiller, Damien Georges
}
\keyword{binary,}
\keyword{convert,}
\keyword{filter}
\keyword{threshold,}
