% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bm_FindOptimStat.R
\name{bm_CalculateStat}
\alias{bm_CalculateStat}
\title{Calculate evaluation metrics based on a misclassification table - only for binary predictions}
\usage{
bm_CalculateStat(Misc, stat='TSS')
}
\arguments{
\item{Misc}{a misclassification table}

\item{stat}{either 'TSS', 'KAPPA', 'ACCURACY', 'BIAS', 'POD', 'FAR', 'POFD', 'SR', 'CSI', 'ETS', 'HK', 'HSS', 'OR' or 'ORSS'}
}
\value{
The \code{stat} score for the \code{Misc} table.
}
\description{
\code{bm_CalculateStat} is used internally in \pkg{biomod2} to get scores, based on a misclassification table
of some referenced evaluation metrics.
}
\details{
Please refer to \code{\link[biomod2]{BIOMOD_Modeling}} to get more information about these metrics.
}
\examples{
  a <- sample(c(0,1),100, replace=TRUE)
  b <- sample(c(0,1),100, replace=TRUE)

  miscTab_aa <- table(a,a)
  miscTab_ab <- table(a,b)

  # perfect score
  bm_CalculateStat(miscTab_aa, stat='TSS')
  # random score
  bm_CalculateStat(miscTab_ab, stat='TSS')

}
\seealso{
\code{\link[biomod2]{BIOMOD_Modeling}}, \code{\link[biomod2]{getStatOptimValue}}, \code{\link[biomod2]{bm_FindOptimStat}}
}
\author{
Damien Georges
}
