% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bm_PlotEvalBoxplot.R
\name{bm_PlotEvalBoxplot}
\alias{bm_PlotEvalBoxplot}
\title{Plot boxplot of evaluation scores}
\usage{
bm_PlotEvalBoxplot(
  modeling.output,
  group.by = c("algo", "run"),
  plot = TRUE,
  ...
)
}
\arguments{
\item{modeling.output}{a \code{\link{BIOMOD.models.out}} or \code{BIOMOD.ensemble.models.out} 
object that can be obtained from \code{\link{BIOMOD_Modeling}} or 
\code{\link{BIOMOD_EnsembleModeling}} functions}

\item{group.by}{a 2-length \code{vector} containing the way kept models will be represented,
must be among \code{model}, \code{algo}, \code{run}, \code{dataset}}

\item{plot}{(\emph{optional, default} \code{TRUE}) \cr 
A \code{logical} value defining whether the plot is to be rendered or not}

\item{\ldots}{some additional arguments (see \href{bm_PlotEvalMean.html#details}{Details})}
}
\value{
A \code{ggplot} object representing boxplot of evaluation scores.
}
\description{
This function represents boxplot of evaluation scores of species distribution models, from 
\code{BIOMOD.models.out} or \code{BIOMOD.ensemble.models.out} objects that can be obtained 
from \code{\link{BIOMOD_Modeling}} or \code{\link{BIOMOD_EnsembleModeling}} functions. Scores 
are represented according to 2 grouping methods (see 
\href{bm_PlotEvalBoxplot.html#details}{Details}).
}
\details{
\code{...} can take the following values :

\itemize{
  \item{\code{main}}{ : a \code{character} corresponding to the graphic title}
  \item{\code{scales}}{ : a \code{character} corresponding to the \code{scales} argument of 
  the \code{\link[ggplot2]{facet_wrap}} function, must be either \code{fixed}, \code{free_x}, 
  \code{free_y} or \code{free}}
}
}
\examples{


}
\seealso{
\code{\link{BIOMOD_Modeling}}, \code{\link{BIOMOD_EnsembleModeling}}, 
\code{\link{get_evaluations}}
}
\author{
Damien Georges, Maya Gueguen
}
\keyword{boxplot}
\keyword{evaluation,}
\keyword{ggplot,}
