% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bm_PlotEvalMean.R
\name{bm_PlotEvalMean}
\alias{bm_PlotEvalMean}
\title{Plot mean evaluation scores}
\usage{
bm_PlotEvalMean(
  modeling.output,
  eval.metric = NULL,
  group.by = "algo",
  plot = TRUE,
  ...
)
}
\arguments{
\item{modeling.output}{a \code{\link{BIOMOD.models.out}} or \code{BIOMOD.ensemble.models.out} 
object that can be obtained from \code{\link{BIOMOD_Modeling}} or 
\code{\link{BIOMOD_EnsembleModeling}} functions}

\item{eval.metric}{a \code{vector} containing evaluation metric names to be used, must 
be among \code{ROC}, \code{TSS}, \code{KAPPA}, \code{ACCURACY}, \code{BIAS}, \code{POD}, 
\code{FAR}, \code{POFD}, \code{SR}, \code{CSI}, \code{ETS}, \code{HK}, \code{HSS}, \code{OR}, 
\code{ORSS}}

\item{group.by}{a \code{character} corresponding to the way kept models will be combined to 
compute mean and sd evaluation scores, must be among \code{model}, \code{algo}, \code{run}
\code{dataset}}

\item{plot}{(\emph{optional, default} \code{TRUE}) \cr 
A \code{logical} value defining whether the plot is to be rendered or not}

\item{\ldots}{some additional arguments (see \href{bm_PlotEvalMean.html#details}{Details})}
}
\value{
A \code{ggplot} object representing mean and standard deviation of evaluation scores 
according to 2 different evaluation methods.
}
\description{
This function represents mean evaluation scores of species distribution models, from 
\code{BIOMOD.models.out} or \code{BIOMOD.ensemble.models.out} objects that can be obtained 
from \code{\link{BIOMOD_Modeling}} or \code{\link{BIOMOD_EnsembleModeling}} functions. Scores 
are represented according to 2 different evaluation methods, and models can be grouped (see 
\href{bm_PlotEvalMean.html#details}{Details}).
}
\details{
\code{...} can take the following values :

\itemize{
  \item{\code{xlim}}{ : an \code{integer} corresponding to the x maximum limit to represent}
  \item{\code{ylim}}{ : an \code{integer} corresponding to the y maximum limit to represent}
  \item{\code{main}}{ : a \code{character} corresponding to the graphic title}
  \item{\code{col}}{ : a \code{vector} containing new color values}
}
}
\examples{


}
\seealso{
\code{\link{BIOMOD_Modeling}}, \code{\link{BIOMOD_EnsembleModeling}}, 
\code{\link{get_evaluations}}
}
\author{
Damien Georges, Maya Gueguen
}
\keyword{evaluation,}
\keyword{ggplot}
