% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bm_PlotRangeSize.R
\name{bm_PlotRangeSize}
\alias{bm_PlotRangeSize}
\title{Plot species range change}
\usage{
bm_PlotRangeSize(
  range.output,
  do.count = TRUE,
  do.perc = TRUE,
  do.maps = TRUE,
  do.mean = TRUE,
  plot = TRUE
)
}
\arguments{
\item{range.output}{an object that can be obtained from \code{\link{BIOMOD_RangeSize}} function}

\item{do.count}{(\emph{optional, default} \code{TRUE}) \cr 
A \code{logical} value defining whether the count plot is to be computed or not}

\item{do.perc}{(\emph{optional, default} \code{TRUE}) \cr 
A \code{logical} value defining whether the percentage plot is to be computed or not}

\item{do.maps}{(\emph{optional, default} \code{TRUE}) \cr 
A \code{logical} value defining whether the maps plot is to be computed or not}

\item{do.mean}{(\emph{optional, default} \code{TRUE}) \cr 
A \code{logical} value defining whether the mean maps plot is to be computed or not}

\item{plot}{(\emph{optional, default} \code{TRUE}) \cr 
A \code{logical} value defining whether the plots are to be rendered or not}
}
\value{
A \code{list} containing \code{ggplot} objects representing species range change.
}
\description{
This function represents species range change from object that can be obtained from 
\code{\link{BIOMOD_RangeSize}} function. Several graphics can be obtained, representing global 
counts or proportions of gains / losses, as well as spatial representations (see 
\href{bm_PlotRangeSize.html#details}{Details}).
}
\details{
4 plots can be obtained with this function :
\describe{
  \item{Count barplot}{representing absolute number of locations (pixels) lost, stable and 
  gained}
  \item{Percentage barplot}{representing percentage of locations (pixels) lost, stable, and 
  the corresponding Species Range Change (\code{PercGain - PercLoss})}
  \item{SRC models maps}{representing spatially locations (pixels) lost, stable and 
  gained for each single distribution model}
  \item{SRC community averaging maps}{representing spatially locations (pixels) lost, stable 
  and gained, taking the majoritary value across single distribution models (and representing 
  the percentage of models' agreement)}
}
\emph{Please see \code{\link{BIOMOD_RangeSize}} function for more details about the values.}
}
\examples{

# species occurrences
myFile <- system.file("external/species/mammals_table.csv", package="biomod2")
DataSpecies <- read.csv(myFile, row.names = 1)
head(DataSpecies)

# the name of studied species
myRespName <- 'GuloGulo'

# the presence/absences data for our species
myResp <- as.numeric(DataSpecies[, myRespName])

# the XY coordinates of species data
myRespXY <- DataSpecies[, c("X_WGS84", "Y_WGS84")]


# Environmental variables extracted from BIOCLIM (bio_3, bio_4, bio_7, bio_11 & bio_12)
myFiles = paste0("external/bioclim/current/bio", c(3, 4, 7, 11, 12), ".grd")
myExpl = raster::stack(system.file(myFiles[1], package = "biomod2"),
                       system.file(myFiles[2], package = "biomod2"),
                       system.file(myFiles[3], package = "biomod2"),
                       system.file(myFiles[4], package = "biomod2"),
                       system.file(myFiles[5], package = "biomod2"))

# 1. Formating Data
myBiomodData <- BIOMOD_FormatingData(resp.var = myResp,
                                     expl.var = myExpl,
                                     resp.xy = myRespXY,
                                     resp.name = myRespName)

# 2. Defining Models Options using default options.
myBiomodOption <- BIOMOD_ModelingOptions()

# 3. Doing Modelisation
myBiomodModelOut <- BIOMOD_Modeling(myBiomodData,
                                    models = c('SRE','RF'),
                                    models.options = myBiomodOption,
                                    NbRunEval = 2,
                                    DataSplit = 80,
                                    VarImport = 0,
                                    models.eval.meth = c('TSS','ROC'),
                                    do.full.models = FALSE,
                                    modeling.id = 'test')

# 4.1 Projecting on current environmental conditions
myBiomodProjection <- BIOMOD_Projection(modeling.output = myBiomodModelOut,
                                        new.env = myExpl,
                                        proj.name = 'current',
                                        chosen.models = 'all',
                                        binary.meth = 'TSS',
                                        compress = FALSE,
                                        build.clamping.mask = FALSE)

# 4.2 Projecting on future environmental conditions

# Environmental variables extracted from BIOCLIM (bio_3, bio_4, bio_7, bio_11 & bio_12)
myFiles = paste0("external/bioclim/future/bio", c(3, 4, 7, 11, 12), ".grd")
myExplFuture = raster::stack(system.file(myFiles[1], package = "biomod2"),
                             system.file(myFiles[2], package = "biomod2"),
                             system.file(myFiles[3], package = "biomod2"),
                             system.file(myFiles[4], package = "biomod2"),
                             system.file(myFiles[5], package = "biomod2"))

myBiomodProjectionFuture <- BIOMOD_Projection(modeling.output = myBiomodModelOut,
                                              new.env = myExplFuture,
                                              proj.name = 'future',
                                              chosen.models = 'all',
                                              binary.meth = 'TSS',
                                              compress = FALSE,
                                              build.clamping.mask = TRUE)

# 5. Detecting predicted species range change

# load binary projections
CurrentProj <- raster::stack("GuloGulo/proj_current/proj_current_GuloGulo_TSSbin.grd")
futurePred <- raster::stack("GuloGulo/proj_future/proj_future_GuloGulo_TSSbin.grd")

myBiomodRangeSize <- BIOMOD_RangeSize(CurrentProj = CurrentProj, FutureProj = futurePred)

# print summary and visualize changes
bm_PlotRangeSize(myBiomodRangeSize)



}
\seealso{
\code{\link{BIOMOD_RangeSize}}
}
\author{
Maya Gueguen
}
\keyword{change,}
\keyword{gain,}
\keyword{ggplot,}
\keyword{loss}
\keyword{projections,}
\keyword{range}
\keyword{species}
