% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bm_PlotResponseCurves.R
\name{bm_PlotResponseCurves}
\alias{bm_PlotResponseCurves}
\title{Plot response curves}
\usage{
bm_PlotResponseCurves(
  modeling.output,
  chosen.models = "all",
  new.env = get_formal_data(modeling.output, "expl.var"),
  show.variables = get_formal_data(modeling.output, "expl.var.names"),
  fixed.var.metric = "mean",
  do.bivariate = FALSE,
  plot = TRUE,
  ...
)
}
\arguments{
\item{modeling.output}{a \code{\link{BIOMOD.models.out}} or \code{BIOMOD.ensemble.models.out} 
object that can be obtained from \code{\link{BIOMOD_Modeling}} or 
\code{\link{BIOMOD_EnsembleModeling}} functions}

\item{chosen.models}{a \code{vector} containing model names to be kept, must be either 
\code{all} or a sub-selection of model names that can be obtained with the 
\code{\link{get_built_models}} function}

\item{new.env}{a \code{matrix}, \code{data.frame} or \code{\link[raster:stack]{RasterStack}} 
object containing the new explanatory variables (in columns or layers, with names matching the 
variables names given to the \code{\link{BIOMOD_FormatingData}} function to build 
\code{modeling.output}) that will be used to project the species distribution model(s)}

\item{show.variables}{a \code{vector} containing the names of the explanatory variables 
present into \code{new.env} parameter and to be plotted}

\item{fixed.var.metric}{a \code{character} corresponding to the statistic to be used to fix as 
constant the remaining variables other than the one used to predict response, must be either 
\code{mean}, \code{median}, \code{min}, \code{max}}

\item{do.bivariate}{(\emph{optional, default} \code{FALSE}) \cr 
A \code{logical} value defining whether the response curves are to be represented in 3 
dimensions (meaning 2 explanatory variables at a time) or not (meaning only 1)}

\item{plot}{(\emph{optional, default} \code{TRUE}) \cr 
A \code{logical} value defining whether the plot is to be rendered or not}

\item{\ldots}{some additional arguments (see 
\href{bm_PlotResponseCurves.html#details}{Details})}
}
\value{
A \code{ggplot} object representing response curves.
}
\description{
This function represents response curves of species distribution models, from 
\code{BIOMOD.models.out} or \code{BIOMOD.ensemble.models.out} objects that can be obtained 
from \code{\link{BIOMOD_Modeling}} or \code{\link{BIOMOD_EnsembleModeling}} functions.
Response curves can be represented in either 2 or 3 dimensions (meaning 1 or 2 explanatory 
variables at a time, see \href{bm_PlotResponseCurves.html#details}{Details}).
}
\details{
This function is an adaptation of the Evaluation Strip method proposed by Elith et al. (2005). 
To build the predicted response curves :
\itemize{
  \item \code{n-1} variables are set constant to a fixed value determined by the 
  \code{fixed.var.metric} parameter (in the case of categorical variable, the most represented 
  class is taken)
  \item the remaining variable is made to vary throughout its range given by the \code{new.env} 
  parameter 
  \item predicted values are computed with these \code{n-1} fixed variables, and this 
  studied variable varying 
}
If \code{do.bivariate = TRUE}, 2 variables are varying at the same time. \cr \cr

The response curves obtained show the sensibility of the model to the studied variable. \cr 
Note that this method does not account for interactions between variables. \cr \cr

\code{...} can take the following values :

\itemize{
  \item{\code{main}}{ : a \code{character} corresponding to the graphic title}
}
}
\examples{

# species occurrences
myFile <- system.file("external/species/mammals_table.csv", package="biomod2")
DataSpecies <- read.csv(myFile, row.names = 1)
head(DataSpecies)

# the name of studied species
myRespName <- 'GuloGulo'

# the presence/absences data for our species
myResp <- as.numeric(DataSpecies[, myRespName])

# the XY coordinates of species data
myRespXY <- DataSpecies[, c("X_WGS84", "Y_WGS84")]


# Environmental variables extracted from BIOCLIM (bio_3, bio_4, bio_7, bio_11 & bio_12)
myFiles = paste0("external/bioclim/current/bio", c(3, 4, 7, 11, 12), ".grd")
myExpl = raster::stack(system.file(myFiles[1], package = "biomod2"),
                       system.file(myFiles[2], package = "biomod2"),
                       system.file(myFiles[3], package = "biomod2"),
                       system.file(myFiles[4], package = "biomod2"),
                       system.file(myFiles[5], package = "biomod2"))

# 1. Formating Data
myBiomodData <- BIOMOD_FormatingData(resp.var = myResp,
                                     expl.var = myExpl,
                                     resp.xy = myRespXY,
                                     resp.name = myRespName)

# 2. Defining Models Options using default options.
myBiomodOption <- BIOMOD_ModelingOptions()

# 3. Doing Modelisation
myBiomodModelOut <- BIOMOD_Modeling(myBiomodData,
                                    models = c('SRE','RF'),
                                    models.options = myBiomodOption,
                                    NbRunEval = 2,
                                    DataSplit = 80,
                                    VarImport = 3,
                                    models.eval.meth = c('TSS','ROC'),
                                    do.full.models = FALSE,
                                    modeling.id = 'test')

# 4.1 Plotting 2D response curves
myRespPlot2D <- bm_PlotResponseCurves(myBiomodModelOut,
                                      chosen.models = get_built_models(myBiomodModelOut)[c(1:3, 12:14)],
                                      do.bivariate = FALSE,
                                      fixed.var.metric = 'median')
                                      
# 4.2 Plotting 3D response curves
myRespPlot3D <- bm_PlotResponseCurves(myBiomodModelOut,
                                      chosen.models = get_built_models(myBiomodModelOut)[3],
                                      do.bivariate = TRUE,
                                      fixed.var.metric = 'median')
                                      
                                      

}
\references{
\itemize{
  \item Elith, J., Ferrier, S., Huettmann, FALSE. and Leathwick, J. R. 2005. The evaluation 
  strip: A new and robust method for plotting predicted responses from species distribution 
  models. \emph{Ecological Modelling}, \bold{186}, 280-289.
}
}
\seealso{
\code{\link{BIOMOD_Modeling}}
}
\author{
Damien Georges, Maya Gueguen
}
\keyword{curve}
\keyword{ggplot,}
\keyword{response}
