% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bm_PlotVarImpBoxplot.R
\name{bm_PlotVarImpBoxplot}
\alias{bm_PlotVarImpBoxplot}
\title{Plot boxplot of variables importance}
\usage{
bm_PlotVarImpBoxplot(
  modeling.output,
  group.by = c("run", "expl.var", "algo"),
  plot = TRUE,
  ...
)
}
\arguments{
\item{modeling.output}{a \code{\link{BIOMOD.models.out}} or \code{BIOMOD.ensemble.models.out} 
object that can be obtained from \code{\link{BIOMOD_Modeling}} or 
\code{\link{BIOMOD_EnsembleModeling}} functions}

\item{group.by}{a 2-length \code{vector} containing the way kept models will be represented,
must be among \code{model}, \code{algo}, \code{run}, \code{dataset}, \code{expl.var}}

\item{plot}{(\emph{optional, default} \code{TRUE}) \cr 
A \code{logical} value defining whether the plot is to be rendered or not}

\item{\ldots}{some additional arguments (see \href{bm_PlotEvalMean.html#details}{Details})}
}
\value{
A \code{ggplot} object representing boxplot of variables importance.
}
\description{
This function represents boxplot of variables importance of species distribution 
models, from \code{\link{BIOMOD.models.out}} or \code{\link{BIOMOD.ensemble.models.out}} 
objects that can be obtained from \code{\link{BIOMOD_Modeling}} or 
\code{\link{BIOMOD_EnsembleModeling}} functions. Scores are represented according to 3 
grouping methods (see \href{bm_PlotVarImpBoxplot.html#details}{Details}).
}
\details{
\code{...} can take the following values :

\itemize{
  \item{\code{main}}{ : a \code{character} corresponding to the graphic title}
}
}
\examples{


}
\seealso{
\code{\link{BIOMOD.models.out}}, \code{\link{BIOMOD.ensemble.models.out}}, 
\code{\link{BIOMOD_Modeling}}, \code{\link{BIOMOD_EnsembleModeling}}, 
\code{\link{get_variables_importance}}

Other Secundary functions: 
\code{\link{bm_BinaryTransformation}()},
\code{\link{bm_CVnnet}()},
\code{\link{bm_FindOptimStat}()},
\code{\link{bm_MakeFormula}()},
\code{\link{bm_PlotEvalBoxplot}()},
\code{\link{bm_PlotEvalMean}()},
\code{\link{bm_PlotRangeSize}()},
\code{\link{bm_PlotResponseCurves}()},
\code{\link{bm_PseudoAbsences}()},
\code{\link{bm_Rescaler}()},
\code{\link{bm_RunModelsLoop}()},
\code{\link{bm_SRE}()},
\code{\link{bm_SampleBinaryVector}()},
\code{\link{bm_SampleFactorLevels}()},
\code{\link{bm_VariablesImportance}()}

Other Plot functions: 
\code{\link{bm_PlotEvalBoxplot}()},
\code{\link{bm_PlotEvalMean}()},
\code{\link{bm_PlotRangeSize}()},
\code{\link{bm_PlotResponseCurves}()}
}
\author{
Damien Georges, Maya Gueguen
}
\concept{Plot functions}
\concept{Secundary functions}
\keyword{boxplot}
\keyword{evaluation,}
\keyword{ggplot,}
