% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bm_Rescaler.R
\name{bm_Rescaler}
\alias{bm_Rescaler}
\title{Rescale values between \code{0} and \code{1} with a binomial GLM}
\usage{
bm_Rescaler(dataToRescale, name, doModel = FALSE, ref = NULL, weights = NULL)
}
\arguments{
\item{dataToRescale}{a \code{vector} containing predicted values to be rescaled}

\item{name}{a \code{character} corresponding to model name to be used, must be among \code{GLM}, 
\code{GBM}, \code{GAM}, \code{CTA}, \code{ANN}, \code{SRE}, \code{FDA}, \code{MARS}, 
\code{RF}, \code{MAXENT.Phillips}, \code{MAXENT.Phillips.2}}

\item{doModel}{(\emph{optional, default} \code{FALSE}) \cr 
A \code{logical} value defining whether the scaling model should be re-created or not (if 
\code{FALSE}, existing scaling model will be loaded)}

\item{ref}{(\emph{optional, default} \code{NULL}) \cr 
A \code{vector} containing observed data to create weights if none given}

\item{weights}{(\emph{optional, default} \code{NULL}) \cr 
A \code{vector} containing values to weight \code{dataToRescale} and influence prevalence}
}
\value{
A \code{vector} containing rescaled values between \code{0} and \code{1}.
}
\description{
This internal \pkg{biomod2} function allows the user to rescale predicted values 
from \code{0} to \code{1} with a binomial GLM (see \href{bm_Rescaler.html#details}{Details}).
}
\details{
\emph{Please refer to parameter \code{rescal.all.models} in \code{\link{BIOMOD_Modeling}} 
function to get more information about this procedure.}

Note that scaling model are saved within 
\file{resp.name/}\code{name}\file{/models/scaling_models/} folder.
}
\examples{

## generate a binary vector
a <- sample(c(0, 1), 100, replace = TRUE)

## biased drawing
BiasedDrawing <- function(x, m1 = 300, sd1 = 200, m2 = 700, sd2 = 200) {
  return(ifelse(x < 0.5, rnorm(1, m1, sd1), rnorm(1, m2, sd2)))
}
b <- sapply(a, BiasedDrawing)
b[which(b < 0)] <- 0
b[which(b > 1000)] <- 1000

b_rescaled <- bm_Rescaler(dataToRescale = b, name = "TEST", doModel = TRUE, ref = a)
plot(b, b_rescaled)
abline(a = 0, b = 0.001, lty = 2)


}
\seealso{
\code{\link{bm_RunModelsLoop}}, \code{\link{BIOMOD_Modeling}}

Other Secundary functions: 
\code{\link{bm_BinaryTransformation}()},
\code{\link{bm_CVnnet}()},
\code{\link{bm_FindOptimStat}()},
\code{\link{bm_MakeFormula}()},
\code{\link{bm_PlotEvalBoxplot}()},
\code{\link{bm_PlotEvalMean}()},
\code{\link{bm_PlotRangeSize}()},
\code{\link{bm_PlotResponseCurves}()},
\code{\link{bm_PlotVarImpBoxplot}()},
\code{\link{bm_PseudoAbsences}()},
\code{\link{bm_RunModelsLoop}()},
\code{\link{bm_SRE}()},
\code{\link{bm_SampleBinaryVector}()},
\code{\link{bm_SampleFactorLevels}()},
\code{\link{bm_VariablesImportance}()}
}
\author{
Damien Georges
}
\concept{Secundary functions}
\keyword{glm}
\keyword{models,}
\keyword{projection,}
\keyword{rescale,}
