% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bm_ModelAnalysis.R
\name{bm_return_info_analysis}
\alias{bm_return_info_analysis}
\title{Functions to get residuals and fitted values of a model}
\usage{
bm_return_info_analysis(mod)
}
\arguments{
\item{mod}{a \code{biomod2_model}}
}
\description{
Functions to get residuals and fitted values of a model
}
