% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bm_VariablesImportance.R
\name{bm_VariablesImportance}
\alias{bm_VariablesImportance}
\title{Variables' importance calculation}
\usage{
bm_VariablesImportance(model, data, method = "full_rand", nb_rand = 1, ...)
}
\arguments{
\item{model}{a \code{BIOMOD.models.out} object (coming either from 
\code{\link{BIOMOD_Modeling}} or \code{\link{BIOMOD_EnsembleModeling}} function) and for 
which variables importance is to be computed}

\item{data}{a \code{data.frame} containing the explanatory variables to use to perform 
variables importance calculation}

\item{method}{a \code{character} corresponding to the randomisation method, must be 
\code{full_rand} (\emph{only method available so far})}

\item{nb_rand}{an \code{integer} corresponding to the number of permutations to be done for 
each variable}
}
\value{
A \code{matrix} containing variables importance scores for each permutation run.
}
\description{
This internal \pkg{biomod2} function allows the user to compute a variable 
importance value for each variable involved in the given model.
}
\details{
For each variable to be evaluated :
\enumerate{
  \item shuffle the original variable
  \item compute model prediction with shuffled variable
  \item calculate Pearson's correlation between reference and shuffled predictions
  \item return score as \code{1 - cor}
}
The highest the value, the less reference and shuffled predictions are correlated, and the 
more influence the variable has on the model. A value of \code{0} assumes no influence of 
the variable on the model.

\emph{Note that this calculation does not account for variables' interactions.}

The same principle is used in \code{\link[randomForest]{randomForest}}.
}
\examples{

xx <- data.frame(a = sample(c(0, 1), 100, replace = TRUE),
                 b = rnorm(100),
                 c = 1:100)
  
mod <- glm(a ~ b + c, data = xx)

bm_VariablesImportance(model = mod, 
                       data = xx[, c('b', 'c')],
                       method = "full_rand",
                       nb_rand = 3)


}
\seealso{
\code{\link{bm_RunModelsLoop}}, \code{\link{BIOMOD_Modeling}}, 
\code{\link{BIOMOD_EnsembleModeling}}, \code{\link{bm_PlotVarImpBoxplot}}, 
\code{\link{get_variables_importance}}

Other Secundary functions: 
\code{\link{bm_BinaryTransformation}()},
\code{\link{bm_CVnnet}()},
\code{\link{bm_FindOptimStat}()},
\code{\link{bm_MakeFormula}()},
\code{\link{bm_PlotEvalBoxplot}()},
\code{\link{bm_PlotEvalMean}()},
\code{\link{bm_PlotRangeSize}()},
\code{\link{bm_PlotResponseCurves}()},
\code{\link{bm_PlotVarImpBoxplot}()},
\code{\link{bm_PseudoAbsences}()},
\code{\link{bm_Rescaler}()},
\code{\link{bm_RunModelsLoop}()},
\code{\link{bm_SRE}()},
\code{\link{bm_SampleBinaryVector}()},
\code{\link{bm_SampleFactorLevels}()}
}
\author{
Damien Georges
}
\concept{Secundary functions}
\keyword{Pearson}
\keyword{correlation}
\keyword{importance,}
\keyword{random,}
\keyword{shuffle,}
