% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEPRECATED.R
\name{calculate.stat-deprecated}
\alias{calculate.stat-deprecated}
\title{calculate.stat}
\usage{
calculate.stat(...)
}
\arguments{
\item{...}{Additional arguments}
}
\description{
Deprecated function name for
  \code{\link{bm_CalculateStat}}
}
\seealso{
\code{\link{biomod2-deprecated}}
}
\keyword{internal}
