% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomod2_internal.R
\name{check_duplicated_cells}
\alias{check_duplicated_cells}
\title{Check duplicated cells}
\usage{
check_duplicated_cells(env, xy, sp, filter.raster)
}
\arguments{
\item{env}{a \code{SpatRaster} with environmental data}

\item{xy}{a \code{data.frame} with coordinates}

\item{sp}{a \code{vector} with species occurrence data}

\item{filter.raster}{a \code{boolean}}
}
\value{
a \code{list}
}
\description{
Identify data that are contained in the same raster cells ; 
print warnings if need be and filter those data if asked for.
}
\keyword{internal}
