% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEPRECATED.R
\name{CustomIndexMaker-deprecated}
\alias{CustomIndexMaker-deprecated}
\title{CustomIndexMaker}
\usage{
CustomIndexMaker(...)
}
\arguments{
\item{...}{Additional arguments}
}
\description{
Deprecated function used in \pkg{biomod2} 3.5.1 to replace
  default html index file by a custom one if defined
}
\seealso{
\code{\link{CustomIndexMaker-deprecated}}
}
\keyword{internal}
