%\renewcommand{\usage}{\pkg{#1}}

\name{CustomIndexMaker}
\alias{CustomIndexMaker}

\title{ Replace default package Index help file by a custom one.}

\description{
This function replace default html index file by a custom one if defined
}
\usage{
CustomIndexMaker()
}

\details{
....
  
} %end detail section

\value{
  A \code{logical} specifying if operation succeed or not
}

\author{ Wilfried Thuiller, Damien Georges }

\examples{

## Automaticly done at buildinfg package state
# CustomIndexMaker()

}
                 
\keyword{ models }
\keyword{ datasets }
