% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomod2_data.R
\docType{data}
\name{DataSpecies}
\alias{DataSpecies}
\title{Presence-Absence data to build test SDM}
\format{
A \code{data.frame} object with 2488 rows and 10 variables:
\describe{
  \item{X_WGS84}{Longitude}
  \item{Y_WGS84}{Latitude}
  \item{ConnochaetesGnou}{Presence (1) or Absence (0) for black wildebeest}
  \item{GuloGulo}{Presence (1) or Absence (0) for wolverine}
  \item{PantheraOnca}{Presence (1) or Absence (0) for jaguar}
  \item{PteropusGiganteus}{Presence (1) or Absence (0) for indian flying fox}
  \item{TenrecEcaudatus}{Presence (1) or Absence (0) for tailless tenrec}
  \item{VulpesVulpes}{Presence (1) or Absence (0) for red fox}
}
}
\usage{
DataSpecies
}
\description{
A dataset covering all the continent with presence/absence data for 6 mammal
species. Presence/absence were derived from range maps downloaded at 
\href{https://www.iucnredlist.org/}{IUCN}.
}
\keyword{datasets}
