\name{biomod2.deprecated.functions}

\alias{getEM_needed_models}
\alias{getEM_needed_models,BIOMOD.EnsembleModeling.out-method}
\alias{getEMalgos}
\alias{getEMalgos,BIOMOD.EnsembleModeling.out-method}
\alias{getEMbuiltModels}
\alias{getEMbuiltModels,BIOMOD.EnsembleModeling.out-method}
\alias{getEMeval}
\alias{getEMeval,BIOMOD.EnsembleModeling.out-method}
\alias{getEMkeptModels}
\alias{getEMkeptModels,BIOMOD.EnsembleModeling.out-method}
\alias{getFormalModel}
\alias{getFormalModel,biomod2_model-method}
\alias{getModelsBuiltModels}
\alias{getModelsBuiltModels,BIOMOD.models.out-method}
\alias{getModelsEvaluations}
\alias{getModelsEvaluations,BIOMOD.models.out-method}
\alias{getModelsInputData}
\alias{getModelsInputData,BIOMOD.models.out-method}
\alias{getModelsOptions}
\alias{getModelsOptions,BIOMOD.models.out-method}
\alias{getModelsPrediction}
\alias{getModelsPrediction,BIOMOD.models.out-method}
\alias{getModelsPredictionEval}
\alias{getModelsPredictionEval,BIOMOD.models.out-method}
\alias{getModelsVarImport}
\alias{getModelsVarImport,BIOMOD.models.out-method}
\alias{getProjection}
\alias{getProjection,BIOMOD.projection.out-method}
\alias{getScalingModel}
\alias{getScalingModel,biomod2_model-method}



\title{ biomod2 deprecated functions }

\description{ 
  Set of functions that will be removed in a while. They have all an updated equivalent.
}

\usage{
getEM_needed_models(obj,...)
getEMalgos(obj,...)
getEMbuiltModels(obj,...)
getEMeval(obj,...)
getEMkeptModels(obj,...)
getFormalModel(obj,...)
getModelsBuiltModels(obj,...)
getModelsEvaluations(obj,...)
getModelsInputData(obj,...)
getModelsOptions(obj,...)
getModelsPrediction(obj,...)
getModelsPredictionEval(obj,...)
getModelsVarImport(obj,...)
getProjection(obj,...)
getScalingModel(obj,...)
}

\arguments{
  \item{obj}{ specific \code{biomod2} object }
  \item{...}{ extra arguments }
}


\author{ Wilfried Thuiller, Damien Georges}

\seealso{ \code{\link{help}}}

\keyword{ getteurs }
