% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bm_BinaryTransformation.R
\name{.bm_BinaryTransformation_singleLayer_SpatRaster}
\alias{.bm_BinaryTransformation_singleLayer_SpatRaster}
\title{Internal function for binary transformation.}
\usage{
.bm_BinaryTransformation_singleLayer_SpatRaster(
  data,
  threshold,
  do.filtering = FALSE
)
}
\description{
\code{.bm_BinaryTransformation_singleLayer_SpatRaster} 
performs binary transformation on a single layer 
\code{\link[raster:stack]{RasterLayer}}
}
\keyword{internal}
