% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomod2_internal.R
\name{.categorical2numeric}
\alias{.categorical2numeric}
\title{Transform categorical into numeric variables}
\usage{
.categorical2numeric(df, categorical_var)
}
\arguments{
\item{df}{data.frame to be transformed}

\item{categorical_var}{the names of categorical variables in df}
}
\value{
a data.frame without categorical variables
}
\description{
Internal function transform categorical variables in a 
data.frame into numeric variables. Mostly used with maxent which cannot 
read character
}
\keyword{internal}
