% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomod2_internal.R
\name{.get_categorical_names}
\alias{.get_categorical_names}
\title{Get categorical variable names}
\usage{
.get_categorical_names(df)
}
\arguments{
\item{df}{data.frame to be checked}
}
\value{
a vector with the name of categorical variables
}
\description{
Internal function to get categorical variables name from a data.frame.
}
\keyword{internal}
