% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomod2_internal.R
\name{.get_env_class}
\alias{.get_env_class}
\title{Get class of environmental data provided}
\usage{
.get_env_class(new.env)
}
\arguments{
\item{new.env}{object to identify}
}
\value{
a character
}
\description{
Get class of environmental data provided
}
\keyword{internal}
