% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomod2_internal.R
\name{.load_gam_namespace}
\alias{.load_gam_namespace}
\title{Load library for GAM models}
\usage{
.load_gam_namespace(model_subclass = c("GAM_mgcv", "BAM_mgcv", "GAM_gam"))
}
\arguments{
\item{model_subclass}{the subclass of GAM model}
}
\description{
This function loads library for either GAM and BAM from mgcv
  package or for GAM from gam package.
}
\keyword{internal}
