% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEPRECATED.R
\name{.transform.outputs.array-deprecated}
\alias{.transform.outputs.array-deprecated}
\title{.transform.outputs.array}
\usage{
.transform.outputs.array(...)
}
\arguments{
\item{...}{Additional arguments}
}
\description{
Deprecated function used in \pkg{biomod2} 3.5.1 to transform
outputs into array
}
\seealso{
\code{\link{biomod2-deprecated}}
}
\keyword{internal}
