% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BIOMOD_Modeling.R, R/transform.outputs.R
\name{.transform.outputs.array}
\alias{.transform.outputs.array}
\title{Reshape biomod2 objects}
\usage{
.transform.outputs.array(modOut, out = "evaluation")

.transform.outputs.array(modOut, out = "evaluation")
}
\arguments{
\item{modOut}{the object to transform given as a list}

\item{out}{character, the type of input object}

\item{modOut}{the object to transform given as a list}
}
\description{
This is an internal function (developper only)

This is an internal function (developper only)
}
