% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEPRECATED.R
\name{.transform.outputs.list-deprecated}
\alias{.transform.outputs.list-deprecated}
\title{.transform.outputs.list}
\usage{
.transform.outputs.list(...)
}
\arguments{
\item{...}{Additional arguments}
}
\description{
Deprecated function name for
  \code{.transform_outputs_list}
}
\seealso{
\code{\link{biomod2-deprecated}}
}
\keyword{internal}
