% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BIOMOD_Modeling.R
\name{.transform.outputs.list}
\alias{.transform.outputs.list}
\title{Reshape biomod2 objects}
\usage{
.transform.outputs.list(modOut, out = "evaluation", dim.names = NULL)
}
\arguments{
\item{modOut}{the object to transform given as a list}

\item{out}{character, the type of input object}

\item{dim.names}{character, if not `NULL` the resshaped object will be stored on the hard drive}
}
\value{
list, the extracted statistics
}
\description{
This is an internal function (developper only)
}
