% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomod2_internal.R
\name{.transform_model.as.col}
\alias{.transform_model.as.col}
\title{Transform predictions data.frame from long to wide with models as columns}
\usage{
.transform_model.as.col(df)
}
\arguments{
\item{df}{a long \code{data.frame}, generally a standard output of 
\code{\link{get_predictions}}}
}
\value{
a wide \code{data.frame}
}
\description{
This function is used internally in \code{\link{get_predictions}}
to ensure back-compatibility with former output of \code{\link{get_predictions}}
(i.e. for \pkg{biomod2} version < 4.2-2). It transform a long \code{data.frame}
into a wide \code{data.frame} with each column corresponding to a single model.

\emph{Note that the function is intended for internal use but have been 
made available for compatibility with \pkg{ecospat}}
}
\author{
Remi Patin
}
\keyword{internal}
