% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ecospat.BIOMOD.presenceonly.R
\name{ecospat.BIOMOD.presenceonly}
\alias{ecospat.BIOMOD.presenceonly}
\title{evaluate models with presences only metrics}
\usage{
ecospat.BIOMOD.presenceonly(modeling.output = NULL, EM.output = NULL,
  save.output = T)
}
\arguments{
\item{modeling.output}{"BIOMOD.models.out" object produced by a BIOMOD_Modeling run}

\item{EM.output}{a "BIOMOD.EnsembleModeling.out" returned by BIOMOD_EnsembleModeling}

\item{save.output}{logical. If TRUE (Default) the output is saved to the ".BIOMOD_DATA" folder}
}
\value{
data.frame containing evaluation scores for the evaluation metrics used for
the BIOMOD_Modeling function and additional Boyce index and MPA
}
\description{
This function enables to evaluate BIOMOD.models.out and
BIOMOD.EnsembleModeling.out object with presence-only evaluation methods
(Boyce index and Minimal Predicted Area MPA)
}
\details{
'em.by' of 'BIOMOD.EnsembleModeling' must be 'PA_dataset+repet' to have an
ensemble for each RUN of the 'NbRunEval' argument (BIOMOD_Modeling funtion)
for evaluation.
The Boyce index returns NA values for 'SRE' models because it is not possible
to be calculated with binary predictions. This is also the reason why there
are sometimes NA values for 'GLM' models if they don not converge.
}
\author{
Frank Breiner
}
\seealso{
ecospat.boyce, ecospat.mpa, BIOMOD_Modeling, BIOMOD_EnsembleModeling
}

