% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEPRECATED.R
\name{FilteringTransformation-deprecated}
\alias{FilteringTransformation-deprecated}
\title{FilteringTransformation}
\usage{
FilteringTransformation(...)
}
\arguments{
\item{...}{Additional arguments}
}
\description{
Deprecated function name for
  \code{\link{BinaryTransformation}}
}
\seealso{
\code{\link{biomod2-deprecated}}
}
\keyword{internal}
