% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEPRECATED.R
\name{Find.Optim.Stat-deprecated}
\alias{Find.Optim.Stat-deprecated}
\title{Find.Optim.Stat}
\usage{
Find.Optim.Stat(...)
}
\arguments{
\item{...}{Additional arguments}
}
\description{
Deprecated function name for
  \code{\link{bm_FindOptimStat}}
}
\seealso{
\code{\link{biomod2-deprecated}}
}
\keyword{internal}
