% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomod2_classes_3.R
\name{get_[...]}
\alias{get_[...]}
\alias{get_options}
\alias{get_calib_lines}
\alias{get_formal_data}
\alias{get_projected_models}
\alias{free}
\alias{get_predictions}
\alias{get_needed_models}
\alias{get_kept_models}
\alias{get_built_models}
\alias{get_evaluations}
\alias{get_variables_importance}
\title{Functions to extract informations from \pkg{biomod2} objects}
\usage{
## for signature 'BIOMOD.models.out' 
get_options(obj)
get_calib_lines(obj)
get_formal_data(obj)

## for signature 'BIOMOD.projection.out'
get_projected_models(obj)
free(obj)

## for signature 'BIOMOD.models.out', 'BIOMOD.projection.out' or 'BIOMOD.ensemble.models.out'
get_predictions(obj)

## for signature 'BIOMOD.ensemble.models.out'
get_needed_models(obj)
get_kept_models(obj)

## for signature 'BIOMOD.models.out' or 'BIOMOD.ensemble.models.out'
get_built_models(obj)
get_evaluations(obj)
get_variables_importance(obj)
}
\arguments{
\item{obj}{a \code{\link{BIOMOD.models.out}}, \code{\link{BIOMOD.projection.out}} or 
\code{\link{BIOMOD.ensemble.models.out}} object}
}
\description{
These functions allow the user to easily retrieve informations stored in the 
different \pkg{biomod2} objects from the different modeling steps, such as modeling options 
and formated data, models used or not, predictions, evaluations, variables importance.
}
\author{
Damien Georges
}
