% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEPRECATED.R
\name{getStatOptimValue-deprecated}
\alias{getStatOptimValue-deprecated}
\title{getStatOptimValue}
\usage{
getStatOptimValue(...)
}
\arguments{
\item{...}{Additional arguments}
}
\description{
Deprecated function name for
  \code{\link{get_optim_value}}
}
\seealso{
\code{\link{biomod2-deprecated}}
}
\keyword{internal}
