% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEPRECATED.R
\name{makeFormula-deprecated}
\alias{makeFormula-deprecated}
\title{makeFormula}
\usage{
makeFormula(...)
}
\arguments{
\item{...}{Additional arguments}
}
\description{
Deprecated function name for
  \code{\link{bm_MakeFormula}}
}
\seealso{
\code{\link{biomod2-deprecated}}
}
\keyword{internal}
