% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEPRECATED.R
\name{models_scores_graph-deprecated}
\alias{models_scores_graph-deprecated}
\title{models_scores_graph}
\usage{
models_scores_graph(...)
}
\arguments{
\item{...}{Additional arguments}
}
\description{
Deprecated function name for
  \code{\link{bm_PlotEvalMean}}
}
\seealso{
\code{\link{biomod2-deprecated}}
}
\keyword{internal}
