% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomod2_data.R
\docType{data}
\name{ModelsTable}
\alias{ModelsTable}
\title{Single models package and functions}
\format{
A data frame with 12 rows and 5 variables:
\describe{
  \item{model}{all single models that can be computed in \pkg{biomod2}}
  \item{type}{data type associated to the models}
  \item{package}{R package used}
  \item{func}{function used in the R package}
  \item{train}{function called by \pkg{caret} for the tuning}
}
}
\usage{
ModelsTable
}
\description{
A \code{data.frame} containing for each single model available in \pkg{biomod2} 
the package and functions to be called.
}
\keyword{datasets}
