% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomod2_data.R
\docType{data}
\name{ODMAP}
\alias{ODMAP}
\title{ODMAP empty table}
\format{
A \code{data.frame} object with 84 rows and 4 variables:
\describe{
  \item{section}{Overview, Data, Model, Assessment or Prediction step}
  \item{subsection}{corresponding field}
  \item{element}{corresponding field}
  \item{value}{to be filled with \code{\link{BIOMOD_Report}} function}
}
}
\usage{
ODMAP
}
\description{
A \code{data.frame} containing ODMAP (Overview, Data, Model, Assessment, Prediction) 
protocol components.
}
\keyword{datasets}
