% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomod2_data.R
\docType{data}
\name{OptionsBigboss}
\alias{OptionsBigboss}
\title{Bigboss pre-defined parameter values for single models}
\format{
A \code{\link{BIOMOD.models.options}} object
}
\usage{
OptionsBigboss
}
\description{
A \code{\link{BIOMOD.models.options}} object containing for each single model available in 
\pkg{biomod2} the parameter values pre-defined by \pkg{biomod2} team.
}
\keyword{datasets}
