% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomod2_classes_4.R
\name{predict.bm}
\alias{predict.bm}
\alias{predict}
\alias{.predict.ANN_biomod2_model.RasterStack}
\alias{.predict.ANN_biomod2_model.data.frame}
\alias{.predict.CTA_biomod2_model.RasterStack}
\alias{.predict.CTA_biomod2_model.data.frame}
\alias{.predict.FDA_biomod2_model.RasterStack}
\alias{.predict.FDA_biomod2_model.data.frame}
\alias{.predict.GAM_biomod2_model.RasterStack}
\alias{.predict.GAM_biomod2_model.data.frame}
\alias{.predict.GBM_biomod2_model.RasterStack}
\alias{.predict.GBM_biomod2_model.data.frame}
\alias{.predict.GLM_biomod2_model.RasterStack}
\alias{.predict.GLM_biomod2_model.data.frame}
\alias{.predict.MARS_biomod2_model.RasterStack}
\alias{.predict.MARS_biomod2_model.data.frame}
\alias{.predict.MAXENT.Phillips_biomod2_model.RasterStack}
\alias{.predict.MAXENT.Phillips_biomod2_model.data.frame}
\alias{.predict.MAXENT.Phillips.2_biomod2_model.RasterStack}
\alias{.predict.MAXENT.Phillips.2_biomod2_model.data.frame}
\alias{.predict.RF_biomod2_model.RasterStack}
\alias{.predict.RF_biomod2_model.data.frame}
\alias{.predict.SRE_biomod2_model.RasterStack}
\alias{.predict.SRE_biomod2_model.data.frame}
\title{Functions to get predictions from \code{\link{biomod2_model}} objects}
\usage{
## for signature 'biomod2_model'
predict(obj)
}
\arguments{
\item{obj}{a \code{\link{biomod2_model}} object}
}
\description{
This function allows the user to predict single models from 
\code{\link{biomod2_model}} on (new) explanatory variables.
}
\author{
Damien Georges
}
