% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEPRECATED.R
\name{Print_Default_ModelingOptions-deprecated}
\alias{Print_Default_ModelingOptions-deprecated}
\title{Print_Default_ModelingOptions}
\usage{
Print_Default_ModelingOptions(...)
}
\arguments{
\item{...}{Additional arguments}
}
\description{
Deprecated function name for
  \code{\link{bm_DefaultModelingOptions}}
}
\seealso{
\code{\link{biomod2-deprecated}}
}
\keyword{internal}
