% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEPRECATED.R
\name{Print_Default_ModelingOptions}
\alias{Print_Default_ModelingOptions}
\title{Print_Default_ModelingOptions}
\usage{
Print_Default_ModelingOptions(...)
}
\description{
Deprecated, use \code{bm_DefaultModelingOptions}
}
