% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEPRECATED.R
\name{ProbDensFunc-deprecated}
\alias{ProbDensFunc-deprecated}
\title{ProbDensFunc}
\usage{
ProbDensFunc(...)
}
\arguments{
\item{...}{Additional arguments}
}
\description{
Deprecated function name for
  \code{\link{bm_PlotRangeSize}}
}
\seealso{
\code{\link{biomod2-deprecated}}
}
\keyword{internal}
