% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomod2_internal.R
\name{rast.has.values}
\alias{rast.has.values}
\title{Check whether SpatRaster is an empty \code{rast()}}
\usage{
rast.has.values(x)
}
\arguments{
\item{x}{SpatRaster to be checked}
}
\value{
a boolean
}
\description{
Check whether SpatRaster is an empty \code{rast()}
}
\keyword{internal}
