% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEPRECATED.R
\name{response.plot-deprecated}
\alias{response.plot-deprecated}
\title{response.plot}
\usage{
response.plot(...)
}
\arguments{
\item{...}{Additional arguments}
}
\description{
Deprecated function name for
  \code{\link{bm_PlotResponseCurves}}
}
\seealso{
\code{\link{biomod2-deprecated}}
}
\keyword{internal}
