% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEPRECATED.R
\name{sample.factor.levels-deprecated}
\alias{sample.factor.levels-deprecated}
\title{sample.factor.levels}
\usage{
sample.factor.levels(...)
}
\arguments{
\item{...}{Additional arguments}
}
\description{
Deprecated function name for
  \code{\link{bm_SampleFactorLevels}}
}
\seealso{
\code{\link{biomod2-deprecated}}
}
\keyword{internal}
