% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomod2_classes_3.R
\name{setters}
\alias{setters}
\alias{set_new_dirname}
\alias{set_new_dirname,BIOMOD.models.out-method}
\alias{set_new_dirname,BIOMOD.ensemble.models.out-method}
\alias{set_new_dirname,BIOMOD.projection.out-method}
\title{Functions to change the place of the different biomod2 objects}
\usage{
set_new_dirname(obj, new.dir.name)

\S4method{set_new_dirname}{BIOMOD.models.out}(obj, new.dir.name)

\S4method{set_new_dirname}{BIOMOD.ensemble.models.out}(obj, new.dir.name)

\S4method{set_new_dirname}{BIOMOD.projection.out}(obj, new.dir.name)
}
\arguments{
\item{obj}{a \code{\link{BIOMOD.formated.data}}, \code{\link{BIOMOD.formated.data.PA}}, 
\code{\link{BIOMOD.models.out}}, \code{\link{BIOMOD.projection.out}} or 
\code{\link{BIOMOD.ensemble.models.out}} object}

\item{new.dir.name}{a \code{character} corresponding to the new folder path}
}
\description{
This function allow the user to change the folder where
the modelling of biomod2 have been done.
}
\seealso{
\code{\link{BIOMOD.models.out}}, \code{\link{BIOMOD.projection.out}}, 
\code{\link{BIOMOD.ensemble.models.out}}

Other Toolbox functions:
\code{\link{getters.bm}},
\code{\link{getters.out}},
\code{\link[=load_stored_object]{load_stored_object()}},
\code{\link{predict2.bm}},
\code{\link{predict2.em}}
}
\author{
Hélène Blancheteau
}
\concept{Toolbox functions}
