% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEPRECATED.R
\name{sre-deprecated}
\alias{sre-deprecated}
\title{sre}
\usage{
sre(...)
}
\arguments{
\item{...}{Additional arguments}
}
\description{
Deprecated function name for
  \code{\link{bm_SRE}}
}
\seealso{
\code{\link{biomod2-deprecated}}
}
\keyword{internal}
