% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomod2_data.R
\docType{data}
\name{TuningTable}
\alias{TuningTable}
\title{Tuning}
\format{
An object of class \code{data.frame} with 57 rows and 8 columns.
}
\usage{
TuningTable
}
\description{
Tuning
}
\keyword{datasets}
