% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEPRECATED.R
\name{update_objects-deprecated}
\alias{update_objects-deprecated}
\title{update_objects}
\usage{
update_objects(...)
}
\arguments{
\item{...}{Additional arguments}
}
\description{
Deprecated function used in \pkg{biomod2} 3.5.1 to update
  objects construct with a old version of biomod2 to a current one
}
\seealso{
\code{\link{biomod2-deprecated}}
}
\keyword{internal}
