% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_objects.R
\name{update_objects}
\alias{update_objects}
\title{update biomod2 objects}
\usage{
update_objects(obj, recursive = TRUE)
}
\arguments{
\item{obj}{a \code{biomod2} object you want to update (e.g. 
'BIOMOD.formated.data', 'biomod2_model' object)}

\item{recursive}{logical, if TRUE all objects on which updated object is based will be also updated}
}
\value{
the updated version of the  biomod2 object is return
}
\description{
This function is wrapper to update objects construct with a old
version of biomod2 to a current one
}
\details{
This function will add/change/delete all object slots that have
evolved between 2 versions of the package.
If required, objects stored on hard drive will also be updated.
}
\seealso{
\code{\link{variables_importance}}, 
\code{\link{full_suffling}}
}
\author{
Damien Georges
}
