% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEPRECATED.R
\name{variables_importance-deprecated}
\alias{variables_importance-deprecated}
\title{variables_importance}
\usage{
variables_importance(...)
}
\arguments{
\item{...}{Additional arguments}
}
\description{
Deprecated function name for
  \code{\link{bm_VariablesImportance}}
}
\seealso{
\code{\link{biomod2-deprecated}}
}
\keyword{internal}
