% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{zzz_bm}
\alias{zzz_bm}
\title{Dummy function to clean working directory after package checks}
\usage{
zzz_bm()
}
\value{
nothing returned
}
\description{
This is the last function that will be check and will remove residual files 
from other examples checking such as 'GuloGulo'. This is required to pass
CRAN checks.
}
\examples{

zzz_bm()
}
\keyword{internal}
