% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomod2_data.R
\docType{data}
\name{bioclim_future}
\alias{bioclim_future}
\title{Bioclimatic variables for SDM based on future condition}
\format{
A \code{\link[terra:rast]{SpatRaster}} with 5 layers:
\describe{
  \item{bio3}{Isothermality}
  \item{bio4}{Temperature Seasonality}
  \item{bio7}{Temperature Annual Range}
  \item{bio11}{Mean Temperature of Coldest Quarter}
  \item{bio12}{Annual Precipitation}
}
}
\usage{
bioclim_future
}
\description{
A \code{\link[terra:rast]{SpatRaster}} with 5 bioclimatic variables commonly
used for SDM and describing future climate based on old RCP scenarios at the
horizon 2080.
}
\keyword{datasets}
