% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEPRECATED.R
\name{biomod2-deprecated}
\alias{biomod2-deprecated}
\alias{BIOMOD_ConvertOldRun}
\alias{BIOMOD_cv}
\alias{BIOMOD_presenceonly}
\alias{BIOMOD_tuning}
\alias{BinaryTransformation}
\alias{calculate.stat}
\alias{CustomIndexMaker}
\alias{FilteringTransformation}
\alias{Find.Optim.Stat}
\alias{getStatOptimValue}
\alias{makeFormula}
\alias{models_scores_graph}
\alias{Print_Default_ModelingOptions}
\alias{ProbDensFunc}
\alias{response.plot}
\alias{response.plot2}
\alias{sample.factor.levels}
\alias{sre}
\alias{.transform.outputs.array}
\alias{.transform.outputs.list}
\alias{update_objects}
\alias{variables_importance}
\title{Deprecated functions in package \pkg{biomod2}.}
\usage{
BIOMOD_ConvertOldRun(...)

BIOMOD_cv(...)

BIOMOD_presenceonly(...)

BIOMOD_tuning(...)

BinaryTransformation(...)

calculate.stat(...)

CustomIndexMaker(...)

FilteringTransformation(...)

Find.Optim.Stat(...)

getStatOptimValue(...)

makeFormula(...)

models_scores_graph(...)

Print_Default_ModelingOptions(...)

ProbDensFunc(...)

response.plot(...)

response.plot2(...)

sample.factor.levels(...)

sre(...)

.transform.outputs.array(...)

.transform.outputs.list(...)

update_objects(...)

variables_importance(...)
}
\description{
The functions listed below are deprecated and were removed
in the current version. When possible, alternative functions with similar
  functionality are mentioned. Help pages for deprecated functions are
  available at \code{help("<function>-deprecated")}.
}
\section{\code{BIOMOD_ConvertOldRun}}{

\code{BIOMOD_ConvertOldRun} was removed in \pkg{biomod2} version 4.0.
}

\section{\code{BIOMOD_cv}}{

For \code{BIOMOD_cv} use \code{\link{BIOMOD_CrossValidation}}.
}

\section{\code{BIOMOD_presenceonly}}{

For \code{BIOMOD_presenceonly} use \code{\link{BIOMOD_PresenceOnly}}.
}

\section{\code{BIOMOD_tuning}}{

For \code{BIOMOD_tuning} use \code{\link{BIOMOD_Tuning}}.
}

\section{\code{BinaryTransformation}}{

For \code{BinaryTransformation} use \code{\link{bm_BinaryTransformation}}.
}

\section{\code{calculate.stat}}{

For \code{calculate.stat} use \code{\link{bm_CalculateStat}}.
}

\section{\code{CustomIndexMaker}}{

\code{CustomIndexMaker} was removed in \pkg{biomod2} version 4.0.
}

\section{\code{FilteringTransformation}}{

For \code{FilteringTransformation} use \code{\link{BinaryTransformation}}.
}

\section{\code{Find.Optim.Stat}}{

For \code{Find.Optim.Stat} use \code{\link{bm_FindOptimStat}}.
}

\section{\code{getStatOptimValue}}{

For \code{getStatOptimValue} use \code{\link{get_optim_value}}.
}

\section{\code{makeFormula}}{

For \code{makeFormula} use \code{\link{bm_MakeFormula}}.
}

\section{\code{models_scores_graph}}{

For \code{models_scores_graph} use \code{\link{bm_PlotEvalMean}}.
}

\section{\code{Print_Default_ModelingOptions}}{

For \code{Print_Default_ModelingOptions} use \code{\link{bm_DefaultModelingOptions}}.
}

\section{\code{ProbDensFunc}}{

For \code{ProbDensFunc} use \code{\link{bm_PlotRangeSize}}.
}

\section{\code{response.plot}}{

For \code{response.plot} use \code{\link{bm_PlotResponseCurves}}.
}

\section{\code{response.plot2}}{

For \code{response.plot2} use \code{\link{bm_PlotResponseCurves}}.
}

\section{\code{sample.factor.levels}}{

For \code{sample.factor.levels} use \code{\link{bm_SampleFactorLevels}}.
}

\section{\code{sre}}{

For \code{sre} use \code{\link{bm_SRE}}.
}

\section{\code{.transform.outputs.array}}{

\code{.transform.outputs.array} was removed in \pkg{biomod2} version 4.0.
}

\section{\code{.transform.outputs.list}}{

For \code{.transform.outputs.list} use \code{.transform_outputs_list}.
}

\section{\code{update_objects}}{

\code{update_objects} was removed in \pkg{biomod2} version 4.0.
}

\section{\code{variables_importance}}{

For \code{variables_importance} use \code{\link{bm_VariablesImportance}}.
}

\keyword{internal}
