% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BIOMOD_LoadModels.R
\name{BIOMOD_LoadModels}
\alias{BIOMOD_LoadModels}
\title{Load species distribution models built with \pkg{biomod2}}
\usage{
BIOMOD_LoadModels(bm.out, ...)
}
\arguments{
\item{bm.out}{a \code{BIOMOD.models.out} (or \code{BIOMOD.ensemble.models.out}) object that 
can be obtained from \code{\link{BIOMOD_Modeling}} (or \code{\link{BIOMOD_EnsembleModeling}}) 
function}

\item{...}{(\emph{optional, see \href{BIOMOD_LoadModels.html#details}{Details})})}
}
\value{
A \code{vector} containing the names of the loaded models.
}
\description{
This function loads individual models built with \code{\link{BIOMOD_Modeling}} 
or \code{\link{BIOMOD_EnsembleModeling}} functions.
}
\details{
This function might be of particular use to load models and make response plot analyses. \cr \cr

Running the function providing only \code{bm.out} argument will load all models built by the 
\code{\link{BIOMOD_Modeling}} or \code{\link{BIOMOD_EnsembleModeling}} function. But a 
subselection of models can be done using the following additional arguments :
\itemize{
  \item{\code{models} : }{a \code{vector} containing model names to be loaded, must be among 
  \code{GLM}, \code{GBM}, \code{GAM}, \code{CTA}, \code{ANN}, \code{SRE}, \code{FDA}, 
  \code{MARS}, \code{RF}, \code{MAXENT.Phillips}, \code{MAXENT.Phillips.2}}
  \item{\code{run.eval} : }{a \code{vector} containing repetition set to be loaded, must be 
  among \code{RUN1}, \code{RUN2}, \code{...}, \code{Full}}
  \item{\code{data.set} : }{a \code{vector} containing pseudo-absence set to be loaded, must 
  be among \code{PA1}, \code{PA2}, \code{...} \cr \cr}
  \item{\code{path} : }{a \code{character} corresponding to the location of the species folder 
  (if different from the current working directory) \cr \cr}
  \item{\code{full.name} : }{a \code{vector} containing model names to be kept, must be either 
  \code{all} or a sub-selection of model names \cr \cr}
  \item{\code{as} : }{a \code{character} to contain the loaded models}
}
}
\examples{

# species occurrences
myFile <- system.file("external/species/mammals_table.csv", package="biomod2")
DataSpecies <- read.csv(myFile, row.names = 1)
head(DataSpecies)

# the name of studied species
myRespName <- 'GuloGulo'

# the presence/absences data for our species
myResp <- as.numeric(DataSpecies[, myRespName])

# the XY coordinates of species data
myRespXY <- DataSpecies[, c("X_WGS84", "Y_WGS84")]


# Environmental variables extracted from BIOCLIM (bio_3, bio_4, bio_7, bio_11 & bio_12)
myFiles = paste0("external/bioclim/current/bio", c(3, 4, 7, 11, 12), ".grd")
myExpl = raster::stack(system.file(myFiles[1], package = "biomod2"),
                       system.file(myFiles[2], package = "biomod2"),
                       system.file(myFiles[3], package = "biomod2"),
                       system.file(myFiles[4], package = "biomod2"),
                       system.file(myFiles[5], package = "biomod2"))

# 1. Formatting Data
myBiomodData <- BIOMOD_FormatingData(resp.var = myResp,
                                     expl.var = myExpl,
                                     resp.xy = myRespXY,
                                     resp.name = myRespName)

# 2. Defining Models Options using default options.
myBiomodOption <- BIOMOD_ModelingOptions()

# 3. Doing Modelisation
myBiomodModelOut <- BIOMOD_Modeling(myBiomodData, 
                                    models = c('SRE', 'CTA', 'RF'), 
                                    models.options = myBiomodOption, 
                                    NbRunEval = 1, 
                                    DataSplit = 80, 
                                    Yweights = NULL, 
                                    VarImport = 0, 
                                    models.eval.meth = c('TSS','ROC'),
                                    SaveObj = TRUE,
                                    rescal.all.models = FALSE,
                                    do.full.models = FALSE)
                                    
# 4. Loading some models built
myLoadedModels <- BIOMOD_LoadModels(myBiomodModelOut, models = 'RF')
myLoadedModels


}
\seealso{
\code{\link{BIOMOD_Modeling}}, \code{\link{BIOMOD_EnsembleModeling}}

Other Main functions: 
\code{\link{BIOMOD_CrossValidation}()},
\code{\link{BIOMOD_EnsembleForecasting}()},
\code{\link{BIOMOD_EnsembleModeling}()},
\code{\link{BIOMOD_FormatingData}()},
\code{\link{BIOMOD_ModelingOptions}()},
\code{\link{BIOMOD_Modeling}()},
\code{\link{BIOMOD_PresenceOnly}()},
\code{\link{BIOMOD_Projection}()},
\code{\link{BIOMOD_RangeSize}()},
\code{\link{BIOMOD_Tuning}()}
}
\author{
Damien Georges
}
\concept{Main functions}
