% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BIOMOD_Modeling.R
\name{BIOMOD_Modeling}
\alias{BIOMOD_Modeling}
\title{Run a range of species distribution models}
\usage{
BIOMOD_Modeling(
  data,
  modeling.id = as.character(format(Sys.time(), "\%s")),
  models = c("GLM", "GBM", "GAM", "CTA", "ANN", "SRE", "FDA", "MARS", "RF",
    "MAXENT.Phillips", "MAXENT.Phillips.2"),
  models.options = NULL,
  NbRunEval = 1,
  DataSplit = 100,
  Yweights = NULL,
  Prevalence = NULL,
  VarImport = 0,
  models.eval.meth = c("KAPPA", "TSS", "ROC"),
  SaveObj = TRUE,
  rescal.all.models = FALSE,
  do.full.models = TRUE,
  ...
)
}
\arguments{
\item{data}{a \code{\link{BIOMOD.formated.data}} object returned by the 
\code{\link{BIOMOD_FormatingData}} function}

\item{modeling.id}{a \code{character} corresponding to the name (ID) of the simulation set 
(\emph{a random number by default})}

\item{models}{a \code{vector} containing model names to be computed, must be among \code{GLM}, 
\code{GBM}, \code{GAM}, \code{CTA}, \code{ANN}, \code{SRE}, \code{FDA}, \code{MARS}, 
\code{RF}, \code{MAXENT.Phillips}, \code{MAXENT.Phillips.2}}

\item{models.options}{a \code{\link{BIOMOD.models.options}} object returned by the 
\code{\link{BIOMOD_ModelingOptions}} function}

\item{NbRunEval}{an \code{integer} corresponding to the number of repetitions to be done for 
calibration/validation splitting \cr (\emph{if specified, \code{DataSplit} and 
\code{do.full.models} will be ignored})}

\item{DataSplit}{a \code{numeric} between \code{0} and \code{1} corresponding to the 
percentage of data used to calibrate the models (calibration/validation splitting)}

\item{Yweights}{(\emph{optional, default} \code{NULL}) \cr 
A \code{vector} of \code{numeric} values corresponding to observation weights (one per 
observation, see \href{BIOMOD_Modeling.html#details}{Details})}

\item{Prevalence}{(\emph{optional, default} \code{NULL}) \cr 
A \code{numeric} between \code{0} and \code{1} corresponding to the species prevalence to 
build '\emph{weighted response weights}' (see \href{BIOMOD_Modeling.html#details}{Details})}

\item{VarImport}{(\emph{optional, default} \code{NULL}) \cr 
An \code{integer} corresponding to the number of permutations to be done for each variable to 
estimate variable importance}

\item{models.eval.meth}{a \code{vector} containing evaluation metric names to be used, must 
be among \code{ROC}, \code{TSS}, \code{KAPPA}, \code{ACCURACY}, \code{BIAS}, \code{POD}, 
\code{FAR}, \code{POFD}, \code{SR}, \code{CSI}, \code{ETS}, \code{HK}, \code{HSS}, \code{OR}, 
\code{ORSS}}

\item{SaveObj}{(\emph{optional, default} \code{TRUE}) \cr 
A \code{logical} value defining whether all results and outputs must be saved on hard drive 
or not (\emph{! strongly recommended !})}

\item{rescal.all.models}{(\emph{optional, default} \code{FALSE}) \cr 
A \code{logical} value defining whether all models predictions must be scaled with a binomial 
GLM or not}

\item{do.full.models}{(\emph{optional, default} \code{TRUE}) \cr 
A \code{logical} value defining whether models calibrated and evaluated over the whole 
dataset must be computed or not}

\item{DataSplitTable}{(\emph{optional, default} \code{NULL}) \cr 
A \code{matrix} or \code{data.frame} defining for each repetition (in columns) which 
observation lines should be used for models calibration (\code{TRUE}) and validation 
(\code{FALSE}) (see \code{\link{BIOMOD_CrossValidation}})}
}
\value{
A \code{BIOMOD.models.out} object containing models outputs, or links to saved outputs. \cr
Models outputs are stored out of \R (for memory storage reasons) in 2 different folders 
created in the current working directory :
\enumerate{
  \item a \emph{models} folder, named after the \code{resp.name} argument of 
  \code{\link{BIOMOD_FormatingData}}, and containing all calibrated models for each 
  repetition and pseudo-absence run
  \item a \emph{hidden} folder, named \code{.BIOMOD_DATA}, and containing outputs related 
  files (original dataset, calibration lines, pseudo-absences selected, predictions, 
  variables importance, evaluation values...), that can be retrieved with 
  \code{\href{}{get_[...]}} or \code{\link{load}} functions, and used by other 
  \pkg{biomod2} functions, like \code{\link{BIOMOD_Projection}} or 
  \code{\link{BIOMOD_EnsembleModeling}}
}
}
\description{
This function allows to calibrate and evaluate a range of modeling techniques 
for a given species distribution. The dataset can be split up for independent calibration and 
validation, and the predictive power of the different models can be estimated using a range 
of evaluation metrics (see \href{BIOMOD_Modeling.html#details}{Details}).
}
\details{
\describe{
  \item{data}{If you have decided to add pseudo absences to your original dataset (see 
  \code{\link{BIOMOD_FormatingData}}), \cr \code{PA.nb.rep *(NbRunEval + 1)} models will be 
  created.}
  
  \item{models}{The set of models to be calibrated on the data. 10 modeling techniques 
  are currently available :
  \itemize{
    \item \code{GLM} : Generalized Linear Model (\code{\link[stats]{glm}})
    \item \code{GAM} : Generalized Additive Model (\code{\link[gam]{gam}}, \code{\link[mgcv]{gam}} 
    or \code{\link[mgcv]{bam}}) \cr 
    (see \code{\link{BIOMOD_ModelingOptions} for details on algorithm selection})
    \item \code{GBM} : Generalized Boosting Model, or usually called Boosted Regression Trees 
    (\code{\link[gbm]{gbm}})
    \item \code{CTA} : Classification Tree Analysis (\code{\link[rpart]{rpart}})
    \item \code{ANN} : Artificial Neural Network (\code{\link[nnet]{nnet}})
    \item \code{SRE} : Surface Range Envelop or usually called BIOCLIM
    \item \code{FDA} : Flexible Discriminant Analysis (\code{\link[mda]{fda}})
    \item \code{MARS} : Multiple Adaptive Regression Splines (\code{\link[earth]{earth}})
    \item \code{RF} : Random Forest (\code{\link[randomForest]{randomForest}})
    \item \code{MAXENT.Phillips} : Maximum Entropy 
    (\url{https://biodiversityinformatics.amnh.org/open_source/maxent})
    \item \code{MAXENT.Phillips.2} : Maximum Entropy (\code{\link[maxnet]{maxnet}})
  }}
  
  \item{NbRunEval & DataSplit}{
  \itemize{
    \item Most simple method in machine learning to calibrate and evaluate a model is to 
    split the original dataset in two, one to calibrate the model and the other one to 
    evaluate it. The \code{DataSplit} argument defines the percentage of data that will be 
    randomly selected and used for the calibration part, the remaining data constituting the 
    evaluation part. This process is repeated \code{NbRunEval} times, to be sure not to 
    include bias both in the modeling and evaluation parts.
    \item Other validation methods are also available to the user :
    \itemize{
      \item evaluation dataset can be directly given to the 
      \code{\link{BIOMOD_FormatingData}} function
      \item \code{DataSplitTable} argument can be used and obtained from the 
      \code{\link{BIOMOD_CrossValidation}} function
    }
  }}
  
  \item{Yweights & Prevalence}{More or less weight can be given to some specific observations.
  \itemize{
    \item If \code{Yweights = Prevalence = NULL}, each observation (presence or absence) will 
    have the same weight, no matter the total number of presences and absences.
    \item If \code{Prevalence = 0.5}, presences and absences will be weighted equally 
    (\emph{i.e. the weighted sum of presences equals the weighted sum of absences}). 
    \item If \code{Prevalence} is set below (\emph{above}) \code{0.5}, more weight will be 
    given to absences (\emph{presences}).
    \item If \code{Yweights} is defined, \code{Prevalence} argument will be ignored, and each 
    observation will have its own weight.
    \item If pseudo-absences have been generated (\code{PA.nb.rep > 0} in 
    \code{\link{BIOMOD_FormatingData}}), weights are by default calculated such that 
    \code{Prevalence = 0.5}. \emph{Automatically created \code{Yweights} will be \code{integer} 
    values to prevent some modeling issues.}
  }}

  \item{models.eval.meth}{
  \itemize{
    \item \code{ROC} : Relative Operating Characteristic
    \item \code{KAPPA} : Cohen's Kappa (Heidke skill score)
    \item \code{TSS} : True kill statistic (Hanssen and Kuipers discriminant, Peirce's skill 
    score)
    \item \code{FAR} : False alarm ratio
    \item \code{SR} : Success ratio
    \item \code{ACCURANCY} : Accuracy (fraction correct)
    \item \code{BIAS} : Bias score (frequency bias)
    \item \code{POD} : Probability of detection (hit rate)
    \item \code{CSI} : Critical success index (threat score)
    \item \code{ETS} : Equitable threat score (Gilbert skill score)
  }
  Optimal value of each method can be obtained with the \code{\link{get_optim_value}} 
  function. Several evaluation metrics can be selected. \emph{Please refer to the 
  \href{http://www.cawcr.gov.au/projects/verification/##'Methods_for_dichotomous_forecasts}{CAWRC website} 
  to get detailed description of each metric.}
  }
  
  \item{SaveObj}{\emph{If this argument is set to \code{FALSE}, it may prevent the evaluation 
  of the ensemble models (see \code{\link{BIOMOD_EnsembleModeling}}) in further steps. Strong 
  recommandation is to keep \code{SaveObj = TRUE}, even if it requires to have some free 
  space onto the hard drive.}
  }
  
  \item{rescal.all.models}{\bold{This parameter is quite experimental and it is recommended 
  not to use it. It may lead to reduction in projection scale amplitude.} Some categorical 
  models always have to be scaled (\code{FDA}, \code{ANN}), but it may be interesting to 
  scale all computed models to ensure comparable predictions (\code{0-1000} range). It might 
  be particularly useful when doing ensemble forecasting to remove the scale prediction effect 
  (\emph{the more extended projections are, the more they influence ensemble forecasting 
  results}).
  }
  
  \item{do.full.models}{Building models with all available information may be useful in some 
  particular cases (\emph{e.g. rare species with few presences points}). But calibration and 
  evaluation datasets will be the same, which might lead to over-optimistic evaluation scores.
  }
}
}
\examples{

# species occurrences
myFile <- system.file("external/species/mammals_table.csv", package="biomod2")
DataSpecies <- read.csv(myFile, row.names = 1)
head(DataSpecies)

# the name of studied species
myRespName <- 'GuloGulo'

# the presence/absences data for our species
myResp <- as.numeric(DataSpecies[, myRespName])

# the XY coordinates of species data
myRespXY <- DataSpecies[, c("X_WGS84", "Y_WGS84")]


# Environmental variables extracted from BIOCLIM (bio_3, bio_4, bio_7, bio_11 & bio_12)
myFiles = paste0("external/bioclim/current/bio", c(3, 4, 7, 11, 12), ".grd")
myExpl = raster::stack(system.file(myFiles[1], package = "biomod2"),
                       system.file(myFiles[2], package = "biomod2"),
                       system.file(myFiles[3], package = "biomod2"),
                       system.file(myFiles[4], package = "biomod2"),
                       system.file(myFiles[5], package = "biomod2"))

# 1. Formating Data
myBiomodData <- BIOMOD_FormatingData(resp.var = myResp,
                                     expl.var = myExpl,
                                     resp.xy = myRespXY,
                                     resp.name = myRespName)

# 2. Defining Models Options using default options.
myBiomodOption <- BIOMOD_ModelingOptions()

# 3. Doing Modelisation
myBiomodModelOut <- BIOMOD_Modeling(myBiomodData,
                                    models = c('SRE','RF'),
                                    models.options = myBiomodOption,
                                    NbRunEval = 2,
                                    DataSplit = 80,
                                    VarImport = 3,
                                    models.eval.meth = c('TSS','ROC'),
                                    do.full.models = FALSE,
                                    modeling.id = 'test')

# print a summary of modeling stuff
myBiomodModelOut



}
\seealso{
\code{\link{BIOMOD_FormatingData}}, \code{\link{BIOMOD_ModelingOptions}}, 
\code{\link{BIOMOD_CrossValidation}}, \code{ \link{bm_VariablesImportance}}, 
\code{\link{BIOMOD_Projection}}, \code{\link{BIOMOD_EnsembleModeling}}
}
\author{
Wilfried Thuiller, Damien Georges, Robin Engler
}
\keyword{models,}
\keyword{multivariate,}
\keyword{nonlinear,}
\keyword{nonparametric,}
\keyword{regression,}
\keyword{tree}
