% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEPRECATED.R
\name{BIOMOD_PresenceOnly-deprecated}
\alias{BIOMOD_PresenceOnly-deprecated}
\title{BIOMOD_PresenceOnly}
\usage{
BIOMOD_PresenceOnly(...)
}
\arguments{
\item{...}{Additional arguments}
}
\description{
Deprecated function name for
  \code{\link{bm_FindOptimStat}}
}
\seealso{
\code{\link{biomod2-deprecated}}
}
\keyword{internal}
