% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BIOMOD_RangeSize.R
\name{BIOMOD_RangeSize}
\alias{BIOMOD_RangeSize}
\title{Analyze the range size differences between projections of species distribution models}
\usage{
BIOMOD_RangeSize(CurrentProj, FutureProj, ...)
}
\arguments{
\item{CurrentProj}{an \code{array}, \code{data.frame}, \code{\link[raster:stack]{RasterLayer}} 
or \code{\link[raster:stack]{RasterStack}} object containing the initial binary projection(s) 
of the (ensemble) species distribution model(s)}

\item{FutureProj}{an \code{array}, \code{data.frame}, \code{\link[raster:stack]{RasterLayer}} 
or \code{\link[raster:stack]{RasterStack}} object containing the final binary projection(s) 
of the (ensemble) species distribution model(s)}
}
\value{
A \code{list} containing two objects :
\describe{
  \item{Compt.By.Species}{a \code{data.frame} containing the summary of range change for each 
  comparison
  \itemize{
    \item{\code{Loss} : }{number of pixels predicted to be lost}
    \item{\code{Stable0} : }{number of pixels not currently occupied and not predicted to be}
    \item{\code{Stable1} : }{number of pixels currently occupied and predicted to remain 
    occupied}
    \item{\code{Gain} : }{number of pixels predicted to be gained}
    \item{\code{PercLoss} : }{percentage of pixels currently occupied and predicted to be lost 
    (\code{Loss / (Loss + Stable1)})}
    \item{\code{PercGain} : }{percentage of pixels predicted to be gained compare to the 
    number of pixels currently occupied (\code{Gain / (Loss + Stable1)})}
    \item{\code{SpeciesRangeChange} : }{percentage of pixels predicted to change (loss or gain) 
    compare to the number of pixels currently occupied (\code{PercGain - PercLoss})}
    \item{\code{CurrentRangeSize} : }{number of pixels currently occupied}
    \item{\code{FutureRangeSize0Disp} : }{number of pixels predicted to be occupied, assuming 
    no migration}
    \item{\code{FutureRangeSize1Disp} : }{number of pixels predicted to be occupied, assuming 
    migration}
  }
  }
  \item{Diff.By.Pixel}{an object in the same form than the input data (\code{CurrentProj} and 
  \code{FutureProj}) and containing a value for each point/pixel of each comparison among :
  \itemize{
    \item \code{-2} : predicted to be lost
    \item \code{-1} : predicted to remain occupied
    \item \code{0} : predicted to remain unoccupied
    \item \code{1} : predicted to be gained
  }
  }
}
}
\description{
This function allows to calculate the absolute number of locations (pixels) lost, 
stable and gained, as well as the corresponding relative proportions, between two (or more) 
binary projections of (ensemble) species distribution models (\emph{which can represent new 
time scales or environmental scenarios for example}).
}
\details{
Note that \bold{this function is only relevant to compare binary projections, made on the 
same area with the same resolution}.

\code{Diff.By.Pixel} object is obtained by applying the simple following formula :
\deqn{FutureProj - 2 * CurrentProj}
}
\examples{

# species occurrences
myFile <- system.file("external/species/mammals_table.csv", package="biomod2")
DataSpecies <- read.csv(myFile, row.names = 1)
head(DataSpecies)

# the name of studied species
myRespName <- 'GuloGulo'

# the presence/absences data for our species
myResp <- as.numeric(DataSpecies[, myRespName])

# the XY coordinates of species data
myRespXY <- DataSpecies[, c("X_WGS84", "Y_WGS84")]


# Environmental variables extracted from BIOCLIM (bio_3, bio_4, bio_7, bio_11 & bio_12)
myFiles = paste0("external/bioclim/current/bio", c(3, 4, 7, 11, 12), ".grd")
myExpl = raster::stack(system.file(myFiles[1], package = "biomod2"),
                       system.file(myFiles[2], package = "biomod2"),
                       system.file(myFiles[3], package = "biomod2"),
                       system.file(myFiles[4], package = "biomod2"),
                       system.file(myFiles[5], package = "biomod2"))

# 1. Formating Data
myBiomodData <- BIOMOD_FormatingData(resp.var = myResp,
                                     expl.var = myExpl,
                                     resp.xy = myRespXY,
                                     resp.name = myRespName)

# 2. Defining Models Options using default options.
myBiomodOption <- BIOMOD_ModelingOptions()

# 3. Doing Modelisation
myBiomodModelOut <- BIOMOD_Modeling(myBiomodData,
                                    models = c('SRE','RF'),
                                    models.options = myBiomodOption,
                                    NbRunEval = 2,
                                    DataSplit = 80,
                                    VarImport = 0,
                                    models.eval.meth = c('TSS','ROC'),
                                    do.full.models = FALSE,
                                    modeling.id = 'test')

# 4.1 Projecting on current environmental conditions
myBiomodProjection <- BIOMOD_Projection(modeling.output = myBiomodModelOut,
                                        new.env = myExpl,
                                        proj.name = 'current',
                                        chosen.models = 'all',
                                        binary.meth = 'TSS',
                                        compress = FALSE,
                                        build.clamping.mask = FALSE)

# 4.2 Projecting on future environmental conditions

# Environmental variables extracted from BIOCLIM (bio_3, bio_4, bio_7, bio_11 & bio_12)
myFiles = paste0("external/bioclim/future/bio", c(3, 4, 7, 11, 12), ".grd")
myExplFuture = raster::stack(system.file(myFiles[1], package = "biomod2"),
                             system.file(myFiles[2], package = "biomod2"),
                             system.file(myFiles[3], package = "biomod2"),
                             system.file(myFiles[4], package = "biomod2"),
                             system.file(myFiles[5], package = "biomod2"))

myBiomodProjectionFuture <- BIOMOD_Projection(modeling.output = myBiomodModelOut,
                                              new.env = myExplFuture,
                                              proj.name = 'future',
                                              chosen.models = 'all',
                                              binary.meth = 'TSS',
                                              compress = FALSE,
                                              build.clamping.mask = TRUE)

# 5. Detecting predicted species range change

# load binary projections
CurrentProj <- raster::stack("GuloGulo/proj_current/proj_current_GuloGulo_TSSbin.grd")
futurePred <- raster::stack("GuloGulo/proj_future/proj_future_GuloGulo_TSSbin.grd")

myBiomodRangeSize <- BIOMOD_RangeSize(CurrentProj = CurrentProj, FutureProj = futurePred)

# print summary and visualize changes
myBiomodRangeSize$Compt.By.Models
plot(myBiomodRangeSize$Diff.By.Pixel)




}
\seealso{
\code{\link{BIOMOD_FormatingData}}, \code{\link{BIOMOD_ModelingOptions}}, 
\code{\link{BIOMOD_Projection}}, \code{\link{BIOMOD_EnsembleModeling}}, 
\code{\link{BIOMOD_EnsembleForecasting}}, \code{\link{level.plot}}
}
\author{
Wilfried Thuiller, Damien Georges, Bruno Lafourcade
}
\keyword{change,}
\keyword{gain,}
\keyword{loss}
\keyword{projections,}
\keyword{range}
\keyword{species}
