% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BIOMOD_Tuning.R
\name{BIOMOD_Tuning}
\alias{BIOMOD_Tuning}
\title{Tune models parameters}
\usage{
BIOMOD_Tuning(
  data,
  models = c("GLM", "GBM", "GAM", "CTA", "ANN", "SRE", "FDA", "MARS", "RF",
    "MAXENT.Phillips"),
  models.options = BIOMOD_ModelingOptions(),
  method.ANN = "avNNet",
  method.RF = "rf",
  method.MARS = "earth",
  method.GAM = "gam",
  method.GLM = "glmStepAIC",
  trControl = NULL,
  metric = "ROC",
  ctrl.CTA = NULL,
  ctrl.RF = NULL,
  ctrl.ANN = NULL,
  ctrl.MARS = NULL,
  ctrl.FDA = NULL,
  ctrl.GAM = NULL,
  ctrl.GBM = NULL,
  ctrl.GLM = NULL,
  tuneLength = 30,
  decay.tune.ANN = c(0.001, 0.01, 0.05, 0.1),
  size.tune.ANN = c(2, 4, 6, 8),
  maxit.ANN = 500,
  MaxNWts.ANN = 10 * (ncol(data@data.env.var) + 1) + 10 + 1,
  type.GLM = c("simple", "quadratic", "polynomial", "s_smoother"),
  interaction.GLM = c(0, 1),
  cvmethod.ME = "randomkfold",
  overlap.ME = FALSE,
  kfolds.ME = 10,
  n.bg.ME = 10000,
  env.ME = NULL,
  metric.ME = "ROC",
  clamp.ME = TRUE,
  parallel.ME = FALSE,
  numCores.ME = NULL,
  Yweights = NULL
)
}
\arguments{
\item{data}{a \code{\link{BIOMOD.formated.data}} object returned by the 
\code{\link{BIOMOD_FormatingData}} function}

\item{models}{a \code{vector} containing model names to be tuned, \cr 
must be among \code{GLM}, \code{GBM}, \code{GAM}, \code{CTA}, \code{ANN}, \code{SRE}, 
\code{FDA}, \code{MARS}, \code{RF}, \code{MAXENT.Phillips}}

\item{models.options}{a \code{\link{BIOMOD.models.options}} object returned by the 
\code{\link{BIOMOD_ModelingOptions}} function}

\item{method.ANN}{a \code{character} corresponding to the classification or regression model 
to use for \code{ANN}, \cr 
must be \code{avNNet} (see \url{http://topepo.github.io/caret/Neural_Network.html})}

\item{method.RF}{a \code{character} corresponding to the classification or regression model 
to use for \code{RF}, \cr 
must be \code{rf} (see \url{http://topepo.github.io/caret/Random_Forest.html})}

\item{method.MARS}{a \code{character} corresponding to the classification or regression model 
to use for \code{MARS}, \cr 
must be \code{earth} (see 
\url{http://topepo.github.io/caret/Multivariate_Adaptive_Regression_Splines.html})}

\item{method.GAM}{a \code{character} corresponding to the classification or regression model 
to use for \code{GAM}, \cr 
must be \code{gam} (see \url{http://topepo.github.io/caret/Generalized_Additive_Model.html})}

\item{method.GLM}{a \code{character} corresponding to the classification or regression model 
to use for \code{GLM}, \cr 
must be \code{glmStepAIC} (see 
\url{http://topepo.github.io/caret/Generalized_Linear_Model.html})}

\item{trControl}{global control parameters that can be obtained from 
\code{\link[caret]{trainControl}} function}

\item{metric}{a \code{character} corresponding to the evaluation metric used to select 
optimal models and tune parameters, must be either \code{ROC} or \code{TSS} 
(\emph{maximizing Sensitivity and Specificity})}

\item{ctrl.CTA}{control parameters for \code{CTA}}

\item{ctrl.RF}{control parameters for \code{RF}}

\item{ctrl.ANN}{control parameters for \code{ANN}}

\item{ctrl.MARS}{control parameters for \code{MARS}}

\item{ctrl.FDA}{control parameters for \code{FDA}}

\item{ctrl.GAM}{control parameters for \code{GAM}}

\item{ctrl.GBM}{control parameters for \code{GBM}}

\item{ctrl.GLM}{control parameters for \code{GLM}}

\item{tuneLength}{(see \code{\link[caret]{train}})}

\item{decay.tune.ANN}{a \code{vector} of weight decay parameters for \code{ANN}}

\item{size.tune.ANN}{a \code{vector} of size parameters (number of units in the hidden layer) 
for \code{ANN}}

\item{maxit.ANN}{an \code{integer} corresponding to the maximum number of iterations for 
\code{ANN}}

\item{MaxNWts.ANN}{an \code{integer} corresponding to the maximum allowable number of weights 
for \code{ANN}}

\item{type.GLM}{a \code{vector} of \code{character} corresponding to modeling types for 
\code{GLM}, \cr must be among \code{simple}, \code{quadratic}, \code{polynomial}, 
\code{s_smoother}}

\item{interaction.GLM}{a \code{vector} of interaction types, must be amond \code{0}, \code{1}}

\item{cvmethod.ME}{a \code{character} corresponding to the method used to partition data for 
\code{MAXENT.Phillips}, \cr must be \code{randomkfold}}

\item{overlap.ME}{(\emph{optional, default} \code{FALSE}) \cr 
A \code{logical} value defining whether to calculate pairwise metric of niche overlap or not 
(see \code{\link{calc.niche.overlap}})}

\item{kfolds.ME}{an \code{integer} corresponding to the number of bins for k-fold 
cross-validation for \code{MAXENT.Phillips}}

\item{n.bg.ME}{an \code{integer} corresponding to the number of background points used to run 
\code{MAXENT.Phillips}}

\item{env.ME}{a \code{RasterStack} object containing model predictor variables}

\item{metric.ME}{a \code{character} corresponding to the evaluation metric used to select 
optimal model and tune parameters for \code{MAXENT.Phillips}, must be either 
\code{auc.val.avg}, \code{auc.diff.avg}, \code{or.mtp.avg}, \code{or.10p.avg} or \code{AICc}}

\item{clamp.ME}{(\emph{optional, default} \code{TRUE}) \cr 
A \code{logical} value defining whether \emph{Features are constrained to remain within the 
range of values in the training data} (Elith et al. 2011) or not}

\item{parallel.ME}{(\emph{optional, default} \code{TRUE}) \cr 
A \code{logical} value defining whether to enable parallel computing for 
\code{MAXENT.Phillips} or not}

\item{numCores.ME}{an \code{integer} corresponding to the number of cores to be used to 
train \code{MAXENT.Phillips}}

\item{Yweights}{a \code{vector} of response points weights for models allowing case weights}
}
\value{
A \code{BIOMOD.models.options} object (see \code{\link{BIOMOD_ModelingOptions}}) with 
optimized parameters
}
\description{
Function to tune \pkg{biomod2} single models parameters
}
\details{
\itemize{
  \item \code{trControl} parameter is set by default to : \cr
  \code{caret::trainControl(method = 'cv', summaryFunction = caret::twoClassSummary,} \cr
  \code{classProbs = TRUE, returnData = FALSE)}.
  \item All control parameters for other models are set to \code{trControl} if unspecified.
  \item For more details on \code{MAXENT.Phillips} tuning, please refer to 
  \code{\link[ENMeval]{ENMevaluate}}.
  \item For more details on other models tuning, please refer to \code{\link[caret]{train}}.
}
}
\examples{

# species occurrences
myFile <- system.file("external/species/mammals_table.csv", package="biomod2")
DataSpecies <- read.csv(myFile, row.names = 1)
head(DataSpecies)

# the name of studied species
myRespName <- 'GuloGulo'

# the presence/absences data for our species
myResp <- as.numeric(DataSpecies[, myRespName])

# the XY coordinates of species data
myRespXY <- DataSpecies[, c("X_WGS84", "Y_WGS84")]


# Environmental variables extracted from BIOCLIM (bio_3, bio_4, bio_7, bio_11 & bio_12)
myFiles = paste0("external/bioclim/current/bio", c(3, 4, 7, 11, 12), ".grd")
myExpl = raster::stack(system.file(myFiles[1], package = "biomod2"),
                       system.file(myFiles[2], package = "biomod2"),
                       system.file(myFiles[3], package = "biomod2"),
                       system.file(myFiles[4], package = "biomod2"),
                       system.file(myFiles[5], package = "biomod2"))

# 1. Formating Data
myBiomodData <- BIOMOD_FormatingData(resp.var = myResp,
                                     expl.var = myExpl,
                                     resp.xy = myRespXY,
                                     resp.name = myRespName)

# 2. Defining Models Options using default options
### Duration for turing all models sequential with default settings 
### on 3.4 GHz processor: approx. 45 min tuning all models in parallel
### (on 8 cores) using foreach loops runs much faster: approx. 14 min

# library(doParallel)
# cl<-makeCluster(8)
# doParallel::registerDoParallel(cl) 

time.seq <- system.time(
  bm.tuning <- BIOMOD_Tuning(myBiomodData,
                             env.ME = myExpl,
                             n.bg.ME = ncell(myExpl))
)

# stopCluster(cl)

par(mfrow = c(1,3))
plot(bm.tuning$tune.CTA.rpart)
plot(bm.tuning$tune.CTA.rpart2)
plot(bm.tuning$tune.RF)

myBiomodModelOut <- BIOMOD_Modeling(myBiomodData, 
                                    models = c('RF', 'CTA'), 
                                    models.options = bm.tuning$models.options, 
                                    NbRunEval = 1, 
                                    DataSplit = 100, 
                                    VarImport = 0,
                                    models.eval.meth = c('ROC'),
                                    do.full.models = FALSE,
                                    modeling.id = 'test')


}
\references{
\itemize{
  \item Kuhn, Max. 2008. Building predictive models in R using the caret package. 
  \emph{Journal of Statistical Software} \bold{28}, 1-26.
  \item Kuhn, Max, and Kjell Johnson. 2013. Applied predictive modeling. New York: Springer.
  \item Muscarella, Robert, et al. 2014. ENMeval: An R package for conducting spatially 
  independent evaluations and estimating optimal model complexity for Maxent ecological 
  niche models. \emph{Methods in Ecology and Evolution}, \bold{5}, 1198-1205.
}
}
\seealso{
\code{\link[caret]{train}}, \code{\link[ENMeval]{ENMevaluate}}, 
\code{\link{BIOMOD_ModelingOptions}}, \code{\link{BIOMOD_Modeling}}

Other Main functions: 
\code{\link{BIOMOD_CrossValidation}()},
\code{\link{BIOMOD_EnsembleForecasting}()},
\code{\link{BIOMOD_EnsembleModeling}()},
\code{\link{BIOMOD_FormatingData}()},
\code{\link{BIOMOD_LoadModels}()},
\code{\link{BIOMOD_ModelingOptions}()},
\code{\link{BIOMOD_Modeling}()},
\code{\link{BIOMOD_PresenceOnly}()},
\code{\link{BIOMOD_Projection}()},
\code{\link{BIOMOD_RangeSize}()}
}
\author{
Frank Breiner
}
\concept{Main functions}
