% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bm_MakeFormula.R
\name{bm_MakeFormula}
\alias{bm_MakeFormula}
\title{Standardized formula maker}
\usage{
<<<<<<< HEAD
bm_MakeFormula(
  resp.name,
  expl.var,
  type = "simple",
  interaction.level = 0,
  ...
)
}
\arguments{
\item{resp.name}{a \code{character} corresponding to the response variable name}

\item{expl.var}{a \code{matrix} or \code{data.frame} containing the explanatory variables that 
will be used at the modeling step}

\item{type}{a \code{character} corresponding to the wanted type of formula, must be 
\code{simple}, \code{quadratic}, \code{polynomial} or \code{s_smoother}}

\item{interaction.level}{an \code{integer} corresponding to the interaction level depth 
between explanatory variables}

\item{\ldots}{some additional arguments (see \href{bm_MakeFormula.html#details}{Details})}
}
\value{
A \code{\link[stats]{formula}} class object that can be directly given to most of \R 
statistical models.
}
\description{
This internal \pkg{biomod2} function allows the user to create easily a 
standardized formula that can be used later by statistical models.
}
\details{
It is advised to give only a subset of \code{expl.var} table to avoid useless memory consuming. 
\cr If some explanatory variables are factorial, \code{expl.var} must be a \code{data.frame} 
whose corresponding columns are defined as \code{factor}. \cr \cr

\code{...} can take the following values :

\itemize{
  \item{\code{k}}{ : an \code{integer} corresponding to the smoothing parameter value of 
  \code{\link[mgcv]{s}} or \code{\link[gam]{s}} arguments (\emph{used only if 
=======
bm_MakeFormula(respName, explVar, type = "simple", interaction.level = 0, ...)
}
\arguments{
\item{respName}{a \code{character} defining the response variable name}

\item{explVar}{a \code{matrix} or \code{data.frame} corresponding to the explanatory
variables table that will be considered at modelling step}

\item{type}{a \code{character} defining the wanted type of formula, must be \code{simple},
\code{quadratic}, \code{polynomial} or \code{s_smoother}}

\item{interaction.level}{an \code{integer} corresponding to the interaction level depth
between explanatory variables}

\item{\ldots}{some additional arguments (see Details)}
}
\value{
A \code{link[stats]{formula}} class object that can be directly given to most of \R
statistical models.
}
\description{
This internal \pkg{biomod2} function allows the user to create easily a standardized formula
that can be used later by statistical models.
}
\details{
###############################################################################################



It is advised to give only a subset of \code{explVar} table to avoid useless memory consuming.
If some explanatory variables are factorial, \code{explVar} must be a \code{data.frame} whose
corresponding columns are defined as \code{factor}.

\code{...} argument available values are :

\itemize{
  \item{\code{k}}{ : an \code{integer} corresponding to the smoothing parameter value of
  \code{\link[mgcv]{s}} or \code{\link[gam]{s}} arguments (\emph{used only if
>>>>>>> devel
  \code{type = 's_smoother'}})}
}
}
\examples{

<<<<<<< HEAD
## Create simple simulated data
myResp.s <- sample(c(0, 1), 20, replace = TRUE)
myExpl.s <- data.frame(var1 = sample(c(0, 1), 100, replace = TRUE),
                       var2 = rnorm(100),
                       var3 = 1:100)

## Generate automatic formula
bm_MakeFormula(resp.name = 'myResp.s',
               expl.var = head(myExpl.s),
               type = 'quadratic',
               interaction.level = 0)


}
\seealso{
\code{\link[stats]{formula}}, \code{\link[mgcv]{s}}, \code{\link[gam]{s}}, 
\code{\link{BIOMOD_ModelingOptions}}, \code{\link{BIOMOD_Tuning}}, 
\code{\link{bm_RunModelsLoop}}

Other Secundary functions: 
\code{\link{bm_BinaryTransformation}()},
\code{\link{bm_CVnnet}()},
\code{\link{bm_FindOptimStat}()},
\code{\link{bm_PlotEvalBoxplot}()},
\code{\link{bm_PlotEvalMean}()},
\code{\link{bm_PlotRangeSize}()},
\code{\link{bm_PlotResponseCurves}()},
\code{\link{bm_PlotVarImpBoxplot}()},
\code{\link{bm_PseudoAbsences}()},
\code{\link{bm_RunModelsLoop}()},
\code{\link{bm_SRE}()},
\code{\link{bm_SampleBinaryVector}()},
\code{\link{bm_SampleFactorLevels}()},
\code{\link{bm_VariablesImportance}()}
=======
## Create simulated data
myResp <- sample(c(0, 1), 20, replace = TRUE)
myExpl <- matrix(runif(60),
                 ncol = 3,
                 dimnames = list(NULL, c('var1', 'var2', 'var3')))

## Create a formula
myFormula <- bm_MakeFormula(respName = 'myResp',
                           explVar = head(myExpl),
                           type = 'quadratic',
                           interaction.level = 0)

## Show formula created
myFormula

###############################################################################################
}
\seealso{
\code{\link[biomod2]{BIOMOD_ModelingOptions}},
\code{link[stats]{formula}}
>>>>>>> devel
}
\author{
Damien Georges
}
<<<<<<< HEAD
\concept{Secundary functions}
=======
>>>>>>> devel
\keyword{formula,}
\keyword{models,}
\keyword{options}
