% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bm_ModelAnalysis.R
\name{bm_ModelAnalysis}
\alias{bm_ModelAnalysis}
\title{Analyze the residuals of the single models}
\usage{
bm_ModelAnalysis(
  bm.mod,
  models.chosen = "all",
  do.plot = TRUE,
  color.by = "full.name"
)
}
\arguments{
\item{bm.mod}{a \code{BIOMOD.models.out}}

\item{models.chosen}{a \code{vector} containing model names to be kept, must be either
\code{all} or a sub-selection of model names that can be obtained with the
\code{\link{get_built_models}} function}

\item{do.plot}{a \code{logical}, print the plot or not}

\item{color.by}{a \code{character} between "full.name", "species", "PA", "RUN" or "algo"}
}
\value{
\enumerate{
\item a plot residuals ~ observations to detect outliers
\item a histogramm of the residuals
\item a qqplot of the residuals
\item plot residuals ~ fitted
\item a plot of the different Rsquared available (Rsquared, Rsquared_aj, Psquared)
}
}
\description{
This function return a pdf with differetn graph about the residuals of the single models.
}
\details{
Only yhe models coherent with a residuals analysis will be selected for some plots.
}
\author{
Hélène Blancheteau
}
\keyword{analyze}
\keyword{models}
\keyword{residuals}
