% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bm_ModelingOptions.R
\name{bm_ModelingOptions}
\alias{bm_ModelingOptions}
\title{Configure the modeling options for each selected model}
\usage{
bm_ModelingOptions(
  data.type,
  models = c("ANN", "CTA", "FDA", "GAM", "GBM", "GLM", "MARS", "MAXENT", "MAXNET", "RF",
    "SRE", "XGBOOST"),
  strategy,
  val.list = NULL,
  bm.format = NULL,
  calib.lines = NULL
)
}
\arguments{
\item{data.type}{a \code{character} corresponding to the data type to 
be used, must be either \code{binary}, \code{binary.PA}, \code{abundance}, 
\code{compositional}}

\item{models}{a \code{vector} containing model names to be computed, must be among 
\code{ANN}, \code{CTA}, \code{FDA}, \code{GAM}, \code{GBM}, \code{GLM}, \code{MARS}, 
\code{MAXENT}, \code{MAXNET}, \code{RF}, \code{SRE}, \code{XGBOOST}}

\item{strategy}{a \code{character} corresponding to the method to 
select models' parameters values, must be either \code{default}, 
\code{bigboss}, \code{user.defined}, \code{tuned}}

\item{val.list}{(\emph{optional, default} \code{NULL}) \cr
A \code{list} containing parameters values for some (all) models}

\item{bm.format}{(\emph{optional, default} \code{NULL}) \cr
A \code{\link{BIOMOD.formated.data}} or \code{\link{BIOMOD.formated.data.PA}} 
object returned by the \code{\link{BIOMOD_FormatingData}} function}

\item{calib.lines}{(\emph{optional, default} \code{NULL}) \cr
A \code{data.frame} object returned by \code{\link{get_calib_lines}} or 
\code{\link{bm_CrossValidation}} functions, to explore the distribution of calibration 
and validation datasets}
}
\value{
A \code{\link{BIOMOD.models.options}} of object that can be used to build 
species distribution model(s) with the \code{\link{BIOMOD_Modeling}} function.
}
\description{
Parameterize and/or tune \pkg{biomod2}'s single models options.
}
\details{
This function creates a \code{\link{BIOMOD.models.options}} object containing parameter values 
for each single models that can be run within \pkg{biomod2} through 
\code{\link{BIOMOD_Modeling}} function.

12 models are currently available, and are listed within \code{\link{ModelsTable}} dataset.

Different strategies are available to set those parameters, through the \code{strategy} 
argument :
\description{
  \item{default}{all parameters names and values are directly retrieve from functions to be 
  called through \code{\link{formalArgs}} and \code{\link{formals}} functions respectively}
  \item{bigboss}{default parameter values are updated with predefined}
  \item{user.defined}{}
  \item{tuned}{}
}
This function allows advanced user to change some default parameters of \pkg{biomod2} inner 
models. \cr 10 single models are available within the package, and their options can be set 
with this function through \code{list} objects.

The \code{bm_DefaultModelingOptions} function prints all default parameter values for 
all available models. \cr This output can be copied and pasted to be used as is (with wanted 
changes) as function arguments (see \href{https://biomodhub.github.io/biomod2/reference/bm_ModelingOptions.html#examples}{Examples}).

Below is the detailed list of all modifiable parameters for each available model.
}
\section{ANN }{
 (\code{\link[nnet]{nnet}})
}

\section{CTA }{
 (\code{\link[rpart]{rpart}})
}

\section{FDA }{
 (\code{\link[mda]{fda}})
}

\section{GAM }{
 (\code{\link[gam]{gam}} or \code{\link[mgcv]{gam}})
(see \code{\link[gam]{gam}}, \code{\link[mgcv]{gam}}, \code{\link[mgcv]{bam}})
}

\section{GBM }{
 (default \code{\link[gbm]{gbm}})
}

\section{GLM }{
 (\code{\link[stats]{glm}})
}

\section{MARS }{
 (\code{\link[earth]{earth}})
}

\section{MAXENT }{
 (\url{https://biodiversityinformatics.amnh.org/open_source/maxent/})
\itemize{
  \item{\code{path_to_maxent.jar = getwd()}}{ : a \code{character}
  corresponding to \pkg{maxent.jar} file link} 
  
  \item{\code{memory_allocated = 512}}{ : an \code{integer} corresponding to
  the amount of memory (in Mo) reserved for \code{java} to run
  \code{MAXENT}, must be \code{64}, \code{128}, \code{256},
  \code{512}, \code{1024}... or \code{NULL} to use default \code{java}
  memory limitation parameter}
  
  \item{\code{initial_heap_size = NULL}}{ : a \code{character} initial heap
  space (shared memory space) allocated to java. Argument transmitted to
  \code{-Xms} when calling java. Used in \code{\link{BIOMOD_Projection}} but
  not in \code{\link{BIOMOD_Modeling}}. Values can be \code{1024K},
  \code{4096M}, \code{10G} ... or \code{NULL} to use default \code{java}
  parameter}
  
  \item{\code{max_heap_size = NULL}}{ : a \code{character} initial heap
  space (shared memory space) allocated to java. Argument transmitted to
  \code{-Xmx} when calling java. Used in \code{\link{BIOMOD_Projection}} but
  not in \code{\link{BIOMOD_Modeling}}. Must be larger than
  \code{initial_heap_size}. Values can be \code{1024K}, \code{4096M},
  \code{10G} ... or \code{NULL} to use default \code{java} parameter}
  
  \item{\code{background_data_dir}}{ : a \code{character} corresponding to
  directory path where explanatory variables are stored as \code{ASCII}
  files (raster format). If specified, \code{MAXENT} will generate
  its own background data from explanatory variables rasters (as usually
  done in \code{MAXENT} studies). Otherwise \pkg{biomod2} pseudo-absences
  will be used (see \code{\link{BIOMOD_FormatingData}})}
  
  \item{\code{visible = FALSE}}{ : a \code{logical} to make the
  \code{MAXENT} user interface available}
  
  \item{\code{linear = TRUE}}{ : a \code{logical} to allow linear features
  to be used} \item{\code{quadratic = TRUE}}{ : a \code{logical} to allow
  quadratic features to be used} 
  
  \item{\code{product = TRUE}}{ : a \code{logical} to allow product features
  to be used}
  
  \item{\code{threshold = TRUE}}{ : a \code{logical} to allow threshold
  features to be used}
  
  \item{\code{hinge = TRUE}}{ : a \code{logical} to allow hinge features to
  be used} \item{\code{lq2lqptthreshold = 80}}{ : an \code{integer}
  corresponding to the number of samples at which product and threshold
  features start being used} 
  
  \item{\code{l2lqthreshold = 10}}{ : an
  \code{integer} corresponding to the number of samples at which quadratic
  features start being used} 
  
  \item{\code{hingethreshold = 15}}{ : an
  \code{integer} corresponding to the number of samples at which hinge
  features start being used}
  
  \item{\code{beta_threshold = -1.0}}{ : a
  \code{numeric} corresponding to the regularization parameter to be applied
  to all threshold features (\emph{negative value enables automatic
  setting})} 
  
  \item{\code{beta_categorical = -1.0}}{ : a \code{numeric}
  corresponding to the regularization parameter to be applied to all
  categorical features (\emph{negative value enables automatic setting})}
  
  \item{\code{beta_lqp = -1.0}}{ : a \code{numeric} corresponding to the
  regularization parameter to be applied to all linear, quadratic and
  product features (\emph{negative value enables automatic setting})}
  
  \item{\code{beta_hinge = -1.0}}{ : a \code{numeric} corresponding to the
  regularization parameter to be applied to all hinge features
  (\emph{negative value enables automatic setting})}
  
  \item{\code{betamultiplier = 1}}{ : a \code{numeric} to multiply all
  automatic regularization parameters \cr (\emph{higher number gives a more
  spread-out distribution})} 
  
  \item{\code{defaultprevalence = 0.5}}{ : a
  \code{numeric} corresponding to the default prevalence of the species \cr
  (\emph{probability of presence at ordinary occurrence points})}
}
}

\section{MAXNET }{
 (\code{\link[maxnet]{maxnet}})
}

\section{RF }{
 (\code{\link[randomForest]{randomForest}})
}

\section{SRE }{
 (\code{\link{bm_SRE}})
}

\section{XGBOOST }{
 (default \code{\link[xgboost]{xgboost}})
}

\examples{
library(terra)

# Load species occurrences (6 species available)
data(DataSpecies)
head(DataSpecies)

# Select the name of the studied species
myRespName <- 'GuloGulo'

# Get corresponding presence/absence data
myResp <- as.numeric(DataSpecies[, myRespName])

# Get corresponding XY coordinates
myRespXY <- DataSpecies[, c('X_WGS84', 'Y_WGS84')]

# Load environmental variables extracted from BIOCLIM (bio_3, bio_4, bio_7, bio_11 & bio_12)
data(bioclim_current)
myExpl <- terra::rast(bioclim_current)

\dontshow{
myExtent <- terra::ext(0,30,45,70)
myExpl <- terra::crop(myExpl, myExtent)
}

# ---------------------------------------------------------------#
# Print default modeling options
bm_DefaultModelingOptions()

# Create default modeling options
myBiomodOptions <- bm_ModelingOptions()
myBiomodOptions

# # Part (or totality) of the print can be copied and customized
# # Below is an example to compute quadratic GLM and select best model with 'BIC' criterium
# myBiomodOptions <- bm_ModelingOptions(
#   GLM = list(type = 'quadratic',
#              interaction.level = 0,
#              myFormula = NULL,
#              test = 'BIC',
#              family = 'binomial',
#              control = glm.control(epsilon = 1e-08,
#                                    maxit = 1000,
#                                    trace = FALSE)))
# myBiomodOptions
# 
# # It is also possible to give a specific GLM formula
# myForm <- 'Sp277 ~ bio3 + log(bio10) + poly(bio16, 2) + bio19 + bio3:bio19'
# myBiomodOptions <- bm_ModelingOptions(GLM = list(myFormula = formula(myForm)))
# myBiomodOptions


}
\seealso{
\code{\link{bm_Tuning}}, \code{\link{BIOMOD_Modeling}}

Other Secundary functions: 
\code{\link{bm_BinaryTransformation}()},
\code{\link{bm_CrossValidation}()},
\code{\link{bm_FindOptimStat}()},
\code{\link{bm_MakeFormula}()},
\code{\link{bm_PlotEvalBoxplot}()},
\code{\link{bm_PlotEvalMean}()},
\code{\link{bm_PlotRangeSize}()},
\code{\link{bm_PlotResponseCurves}()},
\code{\link{bm_PlotVarImpBoxplot}()},
\code{\link{bm_PseudoAbsences}()},
\code{\link{bm_RunModelsLoop}()},
\code{\link{bm_SRE}()},
\code{\link{bm_SampleBinaryVector}()},
\code{\link{bm_SampleFactorLevels}()},
\code{\link{bm_VariablesImportance}()}
}
\author{
Damien Georges, Wilfried Thuiller

Maya Gueguen
}
\concept{Secundary functions}
\keyword{models}
\keyword{options}
