% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bm_PseudoAbsences.R
\name{bm_PseudoAbsences}
\alias{bm_PseudoAbsences}
\title{Select pseudo-absences}
\usage{
bm_PseudoAbsences(
  sp,
  env,
  nb.repet = 1,
  strategy = "random",
  distMin = 0,
  distMax = NULL,
  nb.points = NULL,
  quant.SRE = 0,
  PA.table = NULL
)
}
\arguments{
\item{PA.table}{(\emph{optional, default} \code{NULL}) \cr}
}
\value{
A \code{list} containing the following elements :
\itemize{
  \item{\code{xy} : }{the coordinates of the species observations}
  \item{\code{sp} : }{the values of the species observations (\code{0}, \code{1} or \code{NA})}
  \item{\code{env} : }{the explanatory variables}
  \item{\code{pa.tab} : }{the corresponding table of selected pseudo-absences (indicated by 
  \code{TRUE} or \code{FALSE})}
}
}
\description{
This internal \pkg{biomod2} function allows to select pseudo-absences according to 4 different 
methods : \code{random}, \code{sre}, \code{disk} or \code{user.defined} (see 
\href{bm_PseudoAbsences/html#details}{Details}).
}
\details{
\bold{Concerning random selection :}

\bold{Concerning SRE selection :}

\bold{Concerning disk selection :}

\bold{Concerning user defined selection :}
}
\seealso{
\code{\link{BIOMOD_FormatingData}}
}
\author{
Wilfried Thuiller, Damien Georges
}
\keyword{SRE,}
\keyword{disk}
\keyword{pseudo-absence,}
\keyword{random,}
