% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bm_SampleBinaryVector.R
\name{bm_SampleBinaryVector}
\alias{bm_SampleBinaryVector}
\title{Sample binary vector}
\usage{
bm_SampleBinaryVector(ref, ratio, as.logi = FALSE)
}
\arguments{
\item{ref}{a \code{vector} containing binary values (either \code{0} or \code{1})}

\item{ratio}{a \code{numeric} between \code{0} and \code{1} corresponding to the proportion 
of \code{ref} values to sample}

\item{as.logi}{(\emph{optional, default} \code{FALSE}) \cr 
A \code{logical} value defining whether a \code{vector} of \code{TRUE/FALSE} values should 
be returned (if \code{as.logi = TRUE}) or a \code{vector} containing the indices of 
\code{ref} elements to keep}
}
\value{
A \code{list} containing to elements is returned :
\describe{
  \item{calibration}{IDs of elements selected for calibration}
  \item{evaluation}{IDs of elements selected for evaluation (complementary to the calibration 
  set)}
}
}
\description{
This internal \pkg{biomod2} function allows the user to sample a binary vector 
keeping the same proportion of \code{0} and \code{1} as the initial vector.
}
\examples{

## generate a binary vector
a <- sample(c(0,1), 100, replace = TRUE)

bm_SampleBinaryVector(ref = a, ratio = 0.7)


}
\seealso{
\code{\link{bm_CVnnet}}

Other Secundary functions: 
\code{\link{bm_BinaryTransformation}()},
\code{\link{bm_CVnnet}()},
\code{\link{bm_FindOptimStat}()},
\code{\link{bm_MakeFormula}()},
\code{\link{bm_PlotEvalBoxplot}()},
\code{\link{bm_PlotEvalMean}()},
\code{\link{bm_PlotRangeSize}()},
\code{\link{bm_PlotResponseCurves}()},
\code{\link{bm_PlotVarImpBoxplot}()},
\code{\link{bm_PseudoAbsences}()},
\code{\link{bm_Rescaler}()},
\code{\link{bm_RunModelsLoop}()},
\code{\link{bm_SRE}()},
\code{\link{bm_SampleFactorLevels}()},
\code{\link{bm_VariablesImportance}()}
}
\author{
Damien Georges
}
\concept{Secundary functions}
\keyword{formula,}
\keyword{models,}
\keyword{options}
