% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bm_Tuning.R
\name{bm_Tuning}
\alias{bm_Tuning}
\title{Tune models parameters}
\usage{
bm_Tuning(
  model,
  tuning.fun,
  do.formula = FALSE,
  do.stepAIC = FALSE,
  bm.options,
  bm.format,
  calib.lines = NULL,
  metric.eval = "TSS",
  metric.AIC = "AIC",
  weights = NULL,
  ctrl.train = NULL,
  params.train = list(ANN.size = c(2, 4, 6, 8), ANN.decay = c(0.001, 0.01, 0.05, 0.1),
    ANN.bag = FALSE, FDA.degree = 1:2, FDA.nprune = 2:38, GAM.select = c(TRUE, FALSE),
    GAM.method = c("GCV.Cp", "GACV.Cp", "REML", "P-REML", "ML", "P-ML"), GBM.n.trees =
    c(500, 1000, 2500), GBM.interaction.depth = seq(2, 8, by = 3), GBM.shrinkage =
    c(0.001, 0.01, 0.1), GBM.n.minobsinnode = 10, MARS.degree = 1:2, MARS.nprune =
    2:max(38, 2 * ncol(bm.format@data.env.var) + 1), RF.mtry = 1:min(10,
    ncol(bm.format@data.env.var)), SRE.quant = c(0, 
     0.0125, 0.025, 0.05, 0.1),
    XGBOOST.nrounds = 50, XGBOOST.max_depth = 1, XGBOOST.eta = c(0.3, 0.4), XGBOOST.gamma
    = 0, XGBOOST.colsample_bytree = c(0.6, 0.8), XGBOOST.min_child_weight = 1,
    XGBOOST.subsample = 0.5)
)
}
\arguments{
\item{model}{a \code{character} corresponding to the  algorithm to be tuned, must be either 
\code{ANN}, \code{CTA}, \code{FDA}, \code{GAM}, \code{GBM}, \code{GLM}, \code{MARS}, 
\code{MAXENT}, \code{MAXNET}, \code{RF}, \code{SRE}, \code{XGBOOST}}

\item{tuning.fun}{a \code{character} corresponding to the model function name to be called 
through \code{\link[caret]{train}} function for tuning parameters (see \code{\link{ModelsTable}} 
dataset)}

\item{do.formula}{(\emph{optional, default} \code{FALSE}) \cr
A \code{logical} value defining whether formula is to be optimized or not}

\item{do.stepAIC}{(\emph{optional, default} \code{FALSE}) \cr
A \code{logical} value defining whether variables selection is to be performed for 
\code{GLM} and \code{GAM} models or not}

\item{bm.options}{a \code{\link{BIOMOD.options.default}} or \code{\link{BIOMOD.options.dataset}} 
object returned by the \code{\link{bm_ModelingOptions}} function}

\item{bm.format}{a \code{\link{BIOMOD.formated.data}} or \code{\link{BIOMOD.formated.data.PA}} 
object returned by the \code{\link{BIOMOD_FormatingData}} function}

\item{calib.lines}{(\emph{optional, default} \code{NULL}) \cr
A \code{data.frame} object returned by \code{\link{get_calib_lines}} or 
\code{\link{bm_CrossValidation}} functions}

\item{metric.eval}{a \code{character} corresponding to the evaluation metric to be used, must 
be either \code{ROC}, \code{TSS}, \code{KAPPA} for \code{SRE} only, and \code{auc.val.avg}, 
\code{auc.diff.avg}, \code{or.mtp.avg}, \code{or.10p.avg}, \code{AICc} for \code{MAXENT} only}

\item{metric.AIC}{a \code{character} corresponding to the AIC metric to be used, must 
be either \code{AIC} or \code{BIC}}

\item{weights}{(\emph{optional, default} \code{NULL}) \cr 
A \code{vector} of \code{numeric} values corresponding to observation weights (one per 
observation, see Details)}

\item{params.train}{a \code{list} containing values of model parameters to be tested 
(see Details)}
}
\value{
A \code{\link{BIOMOD.models.options}} object (see \code{\link{bm_ModelingOptions}}) with 
optimized parameters
}
\description{
This internal \pkg{biomod2} function allows to tune single model parameters and 
select more efficient ones based on an evaluation metric.
}
\details{
\bold{Concerning \code{ctrl.train} parameter :}

Set by default to : \cr

\code{ctrl.train <- caret::trainControl(method = "repeatedcv", repeats = 3, number = 10,} \cr
\code{                                  summaryFunction = caret::twoClassSummary,} \cr
\code{                                  classProbs = TRUE, returnData = FALSE)} \cr \cr


\bold{Concerning \code{params.train} parameter :}

All elements of the \code{list} must have names matching \code{model.parameter_name} format, 
\code{parameter_name} being one of the parameter of the \code{tuning.fun} function called by 
\code{caret} package and that can be found through the \code{\link[caret]{getModelInfo}} 
function.

Currently, the available parameters to be tuned are the following :
\describe{
  \item{ANN}{\code{size}, \code{decay}, \code{bag}}
  \item{FDA}{\code{degree}, \code{nprune}}
  \item{GAM}{\code{select}, \code{method}}
  \item{GBM}{\code{n.trees}, \code{interaction.depth}, \code{shrinkage}, \code{n.minobsinnode}}
  \item{MARS}{\code{degree}, \code{nprune}}
  \item{RF}{\code{mtry}}
  \item{SRE}{\code{quant}}
  \item{XGBOOST}{\code{nrounds}, \code{max_depth}, \code{eta}, \code{gamma}, 
  \code{colsampl_bytree}, \code{min_child_weight}, \code{subsample}}
}

The \code{\link{expand.grid}} function is used to build a \code{matrix} containing all 
combinations of parameters to be tested.
}
\note{
\itemize{
  \item No tuning for \code{GLM} and \code{MAXNET}
  \item \code{MAXENT} is tuned through \code{\link[ENMeval]{ENMevaluate}} function
  \item \code{SRE} is tuned through \code{\link{bm_SRE}} function
  \item All other models are tuned through \code{\link[caret]{train}} function
  \item No optimization of formula for \code{MAXENT}, \code{MAXNET} and \code{SRE}
  \item Selection variables only for \code{GAM} and \code{GLM}
}
}
\examples{
library(terra)

# Load species occurrences (6 species available)
data(DataSpecies)
head(DataSpecies)

# Select the name of the studied species
myRespName <- 'GuloGulo'

# Get corresponding presence/absence data
myResp <- as.numeric(DataSpecies[, myRespName])

# Get corresponding XY coordinates
myRespXY <- DataSpecies[, c('X_WGS84', 'Y_WGS84')]

# Load environmental variables extracted from BIOCLIM (bio_3, bio_4, bio_7, bio_11 & bio_12)
data(bioclim_current)
myExpl <- terra::rast(bioclim_current)

\dontshow{
myExtent <- terra::ext(0,30,45,70)
myExpl <- terra::crop(myExpl, myExtent)
}

# --------------------------------------------------------------- #
# Format Data with true absences
myBiomodData <- BIOMOD_FormatingData(resp.var = myResp,
                                     expl.var = myExpl,
                                     resp.xy = myRespXY,
                                     resp.name = myRespName)


# --------------------------------------------------------------- #
### Duration for turing all models sequential with default settings 
### on 3.4 GHz processor: approx. 45 min tuning all models in parallel
### (on 8 cores) using foreach loops runs much faster: approx. 14 min

\dontrun{
# library(doParallel)
# cl <- makeCluster(8)
# doParallel::registerDoParallel(cl) 

time.seq <- system.time(
  bm.tuning <- bm_Tuning(bm.format = myBiomodData, ME.env = myExpl, ME.n.bg = ncell(myExpl))
)

# stopCluster(cl)

plot(bm.tuning$tune.CTA.rpart)
plot(bm.tuning$tune.CTA.rpart2)
plot(bm.tuning$tune.RF)
plot(bm.tuning$tune.ANN)
plot(bm.tuning$tune.MARS)
plot(bm.tuning$tune.FDA)
plot(bm.tuning$tune.GBM)
plot(bm.tuning$tune.GAM)

# Get tuned modeling options
myBiomodOptions <- bm.tuning$models.options
}


}
\seealso{
\code{\link[caret]{trainControl}}, \code{\link[caret]{train}}, 
\code{\link[ENMeval]{ENMevaluate}}, 
\code{\link{ModelsTable}}, \code{\link{BIOMOD.models.options}}, 
\code{\link{bm_ModelingOptions}}, \code{\link{BIOMOD_Modeling}}

Other Secundary functions: 
\code{\link{bm_BinaryTransformation}()},
\code{\link{bm_CrossValidation}()},
\code{\link{bm_FindOptimStat}()},
\code{\link{bm_MakeFormula}()},
\code{\link{bm_ModelingOptions}()},
\code{\link{bm_PlotEvalBoxplot}()},
\code{\link{bm_PlotEvalMean}()},
\code{\link{bm_PlotRangeSize}()},
\code{\link{bm_PlotResponseCurves}()},
\code{\link{bm_PlotVarImpBoxplot}()},
\code{\link{bm_PseudoAbsences}()},
\code{\link{bm_RunModelsLoop}()},
\code{\link{bm_SRE}()},
\code{\link{bm_SampleBinaryVector}()},
\code{\link{bm_SampleFactorLevels}()},
\code{\link{bm_VariablesImportance}()}
}
\author{
Frank Breiner, Maya Gueguen
}
\concept{Secundary functions}
