% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bm_Tuning.R
\name{bm_Tuning}
\alias{bm_Tuning}
\title{Tune models parameters}
\usage{
bm_Tuning(
  model,
  tuning.fun,
  do.formula = FALSE,
  do.stepAIC = FALSE,
  bm.options,
  bm.format,
  calib.lines = NULL,
  metric.eval = "TSS",
  metric.AIC = "AIC",
  weights = NULL,
  ctrl.train = NULL,
  params.train = list(ANN.size = c(2, 4, 6, 8), ANN.decay = c(0.001, 0.01, 0.05, 0.1),
    ANN.bag = FALSE, FDA.degree = 1:2, FDA.nprune = 2:38, GAM.select = c(TRUE, FALSE),
    GAM.method = c("GCV.Cp", "GACV.Cp", "REML", "P-REML", "ML", "P-ML"), GBM.n.trees =
    c(500, 1000, 2500), GBM.interaction.depth = seq(2, 8, by = 3), GBM.shrinkage =
    c(0.001, 0.01, 0.1), GBM.n.minobsinnode = 10, MARS.degree = 1:2, MARS.nprune =
    2:max(38, 2 * ncol(bm.format@data.env.var) + 1), RF.mtry = 1:min(10,
    ncol(bm.format@data.env.var)), SRE.quant = c(0, 
     0.0125, 0.025, 0.05, 0.1),
    XGBOOST.nrounds = 50, XGBOOST.max_depth = 1, XGBOOST.eta = c(0.3, 0.4), XGBOOST.gamma
    = 0, XGBOOST.colsample_bytree = c(0.6, 0.8), XGBOOST.min_child_weight = 1,
    XGBOOST.subsample = 0.5)
)
}
\arguments{
\item{model}{a \code{character} corresponding to the  algorithm to be tuned, must be either 
\code{ANN}, \code{CTA}, \code{FDA}, \code{GAM}, \code{GBM}, \code{GLM}, \code{MARS}, 
\code{MAXENT}, \code{MAXNET}, \code{RF}, \code{SRE}, \code{XGBOOST}}

\item{tuning.fun}{a \code{character} corresponding to the model function name 
to be called through \code{\link[caret]{train}} function for tuning parameters}

\item{do.formula}{(\emph{optional, default} \code{FALSE}) \cr
A \code{logical} value defining whether formula is to be optimized or not}

\item{do.stepAIC}{(\emph{optional, default} \code{FALSE}) \cr
A \code{logical} value defining whether selection variables is to be performed for 
\code{GLM} and \code{GAM} models or not}

\item{bm.options}{a \code{\link{BIOMOD.options.default}} or \code{\link{BIOMOD.options.dataset}} 
object returned by the \code{\link{bm_ModelingOptions}} function}

\item{bm.format}{a \code{\link{BIOMOD.formated.data}} or \code{\link{BIOMOD.formated.data.PA}} 
object returned by the \code{\link{BIOMOD_FormatingData}} function}

\item{calib.lines}{}

\item{metric.eval}{a \code{character} corresponding to the evaluation metric to be used, must 
be either \code{ROC}, \code{TSS}, \code{KAPPA} for \code{SRE} only, and \code{auc.val.avg}, 
\code{auc.diff.avg}, \code{or.mtp.avg}, \code{or.10p.avg}, \code{AICc} for \code{MAXENT} only}

\item{metric.AIC}{a \code{character} corresponding to the AIC metric to be used, must 
be either \code{AIC}, \code{BIC}}

\item{weights}{(\emph{optional, default} \code{NULL}) \cr 
A \code{vector} of \code{numeric} values corresponding to observation weights (one per 
observation, see Details)}

\item{params.train}{a \code{list} containing values of model parameters to be tested 
(see Details)}
}
\value{
A \code{\link{BIOMOD.models.options}} object (see \code{\link{bm_ModelingOptions}}) with 
optimized parameters
}
\description{
This internal \pkg{biomod2} function allows to tune single model parameters and 
select more efficient ones based on an evaluation metric (see Details).
}
\details{
\itemize{
  \item \code{ctrl.train} parameter is set by default to : \cr
  \code{caret::trainControl(method = 'cv', summaryFunction = caret::twoClassSummary,} \cr
  \code{classProbs = TRUE, returnData = FALSE)}.
  \item All control parameters for other models are set to \code{ctrl.train} if unspecified.
  \item For more details on \code{MAXENT} tuning, please refer to 
  \code{\link[ENMeval]{ENMevaluate}}.
  \item For more details on other models tuning, please refer to \code{\link[caret]{train}}.
}
}
\examples{
library(terra)

# Load species occurrences (6 species available)
data(DataSpecies)
head(DataSpecies)

# Select the name of the studied species
myRespName <- 'GuloGulo'

# Get corresponding presence/absence data
myResp <- as.numeric(DataSpecies[, myRespName])

# Get corresponding XY coordinates
myRespXY <- DataSpecies[, c('X_WGS84', 'Y_WGS84')]

# Load environmental variables extracted from BIOCLIM (bio_3, bio_4, bio_7, bio_11 & bio_12)
data(bioclim_current)
myExpl <- terra::rast(bioclim_current)

\dontshow{
myExtent <- terra::ext(0,30,45,70)
myExpl <- terra::crop(myExpl, myExtent)
}

# --------------------------------------------------------------- #
# Format Data with true absences
myBiomodData <- BIOMOD_FormatingData(resp.var = myResp,
                                     expl.var = myExpl,
                                     resp.xy = myRespXY,
                                     resp.name = myRespName)


# --------------------------------------------------------------- #
### Duration for turing all models sequential with default settings 
### on 3.4 GHz processor: approx. 45 min tuning all models in parallel
### (on 8 cores) using foreach loops runs much faster: approx. 14 min

\dontrun{
# library(doParallel)
# cl <- makeCluster(8)
# doParallel::registerDoParallel(cl) 

time.seq <- system.time(
  bm.tuning <- bm_Tuning(bm.format = myBiomodData, ME.env = myExpl, ME.n.bg = ncell(myExpl))
)

# stopCluster(cl)

plot(bm.tuning$tune.CTA.rpart)
plot(bm.tuning$tune.CTA.rpart2)
plot(bm.tuning$tune.RF)
plot(bm.tuning$tune.ANN)
plot(bm.tuning$tune.MARS)
plot(bm.tuning$tune.FDA)
plot(bm.tuning$tune.GBM)
plot(bm.tuning$tune.GAM)

# Get tuned modeling options
myBiomodOptions <- bm.tuning$models.options
}


}
\references{
\itemize{
  \item Kuhn, Max. 2008. Building predictive models in R using the caret package. 
  \emph{Journal of Statistical Software} \bold{28}, 1-26.
  \item Kuhn, Max, and Kjell Johnson. 2013. Applied predictive modeling. New York: Springer.
  \item Muscarella, Robert, et al. 2014. ENMeval: An R package for conducting spatially 
  independent evaluations and estimating optimal model complexity for Maxent ecological 
  niche models. \emph{Methods in Ecology and Evolution}, \bold{5}, 1198-1205.
}
}
\seealso{
\code{\link[caret]{trainControl}}, \code{\link[caret]{train}}, 
\code{\link[ENMeval]{calc.niche.overlap}}, \code{\link[ENMeval]{ENMevaluate}}, 
\code{\link{bm_ModelingOptions}}, \code{\link{BIOMOD_Modeling}}

Other Main functions: 
\code{\link{BIOMOD_EnsembleForecasting}()},
\code{\link{BIOMOD_EnsembleModeling}()},
\code{\link{BIOMOD_FormatingData}()},
\code{\link{BIOMOD_LoadModels}()},
\code{\link{BIOMOD_Modeling}()},
\code{\link{BIOMOD_Projection}()},
\code{\link{BIOMOD_RangeSize}()}
}
\author{
Frank Breiner
}
\concept{Main functions}
