% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BIOMOD_Modeling.R
\name{.transform.outputs.array}
\alias{.transform.outputs.array}
\title{Reshape biomod2 objects}
\usage{
.transform.outputs.array(modOut, out = "evaluation")
}
\arguments{
\item{modOut}{the model object to transform given as a list}

\item{out}{character, the type of output to be transformed}
}
\value{
extracted statistics of interest from the model object
  as `array`.
}
\description{
This is an internal function (developper only)
}
