% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomod2_classes_4.R
\name{getters.bm}
\alias{getters.bm}
\alias{get_formal_model}
\alias{get_scaling_model}
\title{Functions to extract informations from \code{\link{biomod2_model}} objects}
\usage{
## for signature 'biomod2_model' 
get_formal_model(obj)
get_scaling_model(obj)
}
\arguments{
\item{obj}{a \code{\link{biomod2_model}} object}
}
\description{
These functions allow the user to easily retrieve single models (formal or scaled) 
from \code{\link{biomod2_model}} objects from the modeling step.
}
\author{
Damien Georges
}
