% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomod2_classes_4.R
\name{getters.bm}
\alias{getters.bm}
\alias{get_formal_model}
\alias{get_scaling_model}
\alias{get_formal_model,biomod2_model-method}
\alias{get_scaling_model,biomod2_model-method}
\title{Functions to extract informations from \code{\link{biomod2_model}} objects}
\usage{
\S4method{get_formal_model}{biomod2_model}(object)

\S4method{get_scaling_model}{biomod2_model}(object)
}
\arguments{
\item{object}{a \code{\link{biomod2_model}} object}
}
\value{
\describe{
  \item{\code{get_formal_model}}{an object from the \code{model} slot of a 
  \code{\link{biomod2_model}} object}
  \item{\code{get_scaling_model}}{an object from the \code{scaling_model} slot of a 
  \code{\link{biomod2_model}} object}
}
}
\description{
These functions allow the user to easily retrieve single models (formal or scaled) 
from \code{\link{biomod2_model}} objects from the modeling step.
}
\seealso{
\code{\link{biomod2_model}}

Other Toolbox functions: 
\code{\link{getters.out}},
\code{\link{load_stored_object}()},
\code{\link{predict.bm}},
\code{\link{predict.em}},
\code{\link{predict2.bm}},
\code{\link{predict2.em}}
}
\author{
Damien Georges
}
\concept{Toolbox functions}
