% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomod2_classes_3.R
\name{getters.out}
\alias{getters.out}
\alias{get_options}
\alias{get_calib_lines}
\alias{get_formal_data}
\alias{get_projected_models}
\alias{free}
\alias{get_predictions}
\alias{get_needed_models}
\alias{get_kept_models}
\alias{get_built_models}
\alias{get_evaluations}
\alias{get_variables_importance}
\alias{get_options,BIOMOD.models.out-method}
\alias{get_calib_lines,BIOMOD.models.out-method}
\alias{get_formal_data,BIOMOD.models.out-method}
\alias{get_predictions,BIOMOD.models.out-method}
\alias{get_built_models,BIOMOD.models.out-method}
\alias{get_evaluations,BIOMOD.models.out-method}
\alias{get_variables_importance,BIOMOD.models.out-method}
\alias{get_projected_models,BIOMOD.projection.out-method}
\alias{free,BIOMOD.projection.out-method}
\alias{get_predictions,BIOMOD.projection.out-method}
\alias{get_needed_models,BIOMOD.ensemble.models.out-method}
\alias{get_kept_models,BIOMOD.ensemble.models.out-method}
\alias{get_predictions,BIOMOD.ensemble.models.out-method}
\alias{get_built_models,BIOMOD.ensemble.models.out-method}
\alias{get_evaluations,BIOMOD.ensemble.models.out-method}
\alias{get_variables_importance,BIOMOD.ensemble.models.out-method}
\title{Functions to extract informations from \code{\link{BIOMOD.models.out}}, 
\code{\link{BIOMOD.projection.out}} or \code{\link{BIOMOD.ensemble.models.out}} objects}
\usage{
\S4method{get_options}{BIOMOD.models.out}(obj)

\S4method{get_calib_lines}{BIOMOD.models.out}(obj, as.data.frame = FALSE, ...)

\S4method{get_formal_data}{BIOMOD.models.out}(obj, subinfo = NULL)

\S4method{get_predictions}{BIOMOD.models.out}(obj, as.data.frame = FALSE, evaluation = FALSE)

\S4method{get_built_models}{BIOMOD.models.out}(obj, ...)

\S4method{get_evaluations}{BIOMOD.models.out}(obj, ...)

\S4method{get_variables_importance}{BIOMOD.models.out}(obj, ...)

\S4method{get_projected_models}{BIOMOD.projection.out}(obj)

\S4method{free}{BIOMOD.projection.out}(obj)

\S4method{get_predictions}{BIOMOD.projection.out}(
  obj,
  as.data.frame = FALSE,
  full.name = NULL,
  model = NULL,
  run.eval = NULL,
  data.set = NULL
)

\S4method{get_needed_models}{BIOMOD.ensemble.models.out}(obj, selected.models = "all", ...)

\S4method{get_kept_models}{BIOMOD.ensemble.models.out}(obj, model, ...)

\S4method{get_predictions}{BIOMOD.ensemble.models.out}(obj, ...)

\S4method{get_built_models}{BIOMOD.ensemble.models.out}(obj, ...)

\S4method{get_evaluations}{BIOMOD.ensemble.models.out}(obj, ...)

\S4method{get_variables_importance}{BIOMOD.ensemble.models.out}(obj, ...)
}
\arguments{
\item{obj}{a \code{\link{BIOMOD.models.out}}, \code{\link{BIOMOD.projection.out}} or 
\code{\link{BIOMOD.ensemble.models.out}} object}
}
\description{
These functions allow the user to easily retrieve informations stored in the 
different \pkg{biomod2} objects from the different modeling steps, such as modeling options 
and formated data, models used or not, predictions, evaluations, variables importance.
}
\author{
Damien Georges
}
