% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomod2_classes_2.R
\name{load_stored_object}
\alias{load_stored_object}
\title{Functions to load \code{\link{BIOMOD.stored.data}} objects}
\usage{
load_stored_object(obj, ...)
}
\arguments{
\item{obj}{a \code{\link{BIOMOD.stored.data}} object}
}
\description{
This functions allow the user to load \code{\link{BIOMOD.stored.data}} objects 
into memory.
}
\author{
Damien Georges
}
