% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomod2_classes_2.R
\name{load_stored_object}
\alias{load_stored_object}
\alias{load_stored_object,BIOMOD.stored.data-method}
\title{Functions to load \code{\link{BIOMOD.stored.data}} objects}
\usage{
load_stored_object(obj)

\S4method{load_stored_object}{BIOMOD.stored.data}(obj)
}
\arguments{
\item{obj}{a \code{\link{BIOMOD.stored.data}} object}
}
\description{
This functions allow the user to load \code{\link{BIOMOD.stored.data}} objects 
into memory.
}
\seealso{
\code{\link{BIOMOD.stored.data}}

Other Toolbox functions: 
\code{\link{getters.bm}},
\code{\link{getters.out}},
\code{\link{predict.bm}},
\code{\link{predict.em}},
\code{\link{predict2.bm}},
\code{\link{predict2.em}}
}
\author{
Damien Georges
}
\concept{Toolbox functions}
