% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomod2_classes_4.R
\name{predict.bm}
\alias{predict.bm}
\alias{.predict.ANN_biomod2_model.RasterStack}
\alias{.predict.ANN_biomod2_model.data.frame}
\alias{.predict.CTA_biomod2_model.RasterStack}
\alias{.predict.CTA_biomod2_model.data.frame}
\alias{.predict.FDA_biomod2_model.RasterStack}
\alias{.predict.FDA_biomod2_model.data.frame}
\alias{.predict.GAM_biomod2_model.RasterStack}
\alias{.predict.GAM_biomod2_model.data.frame}
\alias{.predict.GBM_biomod2_model.RasterStack}
\alias{.predict.GBM_biomod2_model.data.frame}
\alias{.predict.GLM_biomod2_model.RasterStack}
\alias{.predict.GLM_biomod2_model.data.frame}
\alias{.predict.MARS_biomod2_model.RasterStack}
\alias{.predict.MARS_biomod2_model.data.frame}
\alias{.predict.MAXENT.Phillips_biomod2_model.data.frame}
\alias{.predict.MAXENT.Phillips.2_biomod2_model.RasterStack}
\alias{.predict.MAXENT.Phillips.2_biomod2_model.data.frame}
\alias{.predict.RF_biomod2_model.RasterStack}
\alias{.predict.RF_biomod2_model.data.frame}
\alias{.predict.SRE_biomod2_model.RasterStack}
\alias{.predict.SRE_biomod2_model.data.frame}
\alias{predict,ANN_biomod2_model-method}
\alias{predict,CTA_biomod2_model-method}
\alias{predict,FDA_biomod2_model-method}
\alias{predict,GAM_biomod2_model-method}
\alias{predict,GBM_biomod2_model-method}
\alias{predict,GLM_biomod2_model-method}
\alias{predict,MARS_biomod2_model-method}
\alias{predict,MAXENT.Phillips_biomod2_model-method}
\alias{predict,MAXENT.Phillips.2_biomod2_model-method}
\alias{predict,RF_biomod2_model-method}
\alias{predict,SRE_biomod2_model-method}
\title{Functions to get predictions from \code{\link{biomod2_model}} objects}
\usage{
\S4method{predict}{ANN_biomod2_model}(object, newdata, ...)

\S4method{predict}{CTA_biomod2_model}(object, newdata, ...)

\S4method{predict}{FDA_biomod2_model}(object, newdata, ...)

\S4method{predict}{GAM_biomod2_model}(object, newdata, ...)

\S4method{predict}{GBM_biomod2_model}(object, newdata, ...)

\S4method{predict}{GLM_biomod2_model}(object, newdata, ...)

\S4method{predict}{MARS_biomod2_model}(object, newdata, ...)

\S4method{predict}{MAXENT.Phillips_biomod2_model}(object, newdata, ...)

\S4method{predict}{MAXENT.Phillips.2_biomod2_model}(object, newdata, ...)

\S4method{predict}{RF_biomod2_model}(object, newdata, ...)

\S4method{predict}{SRE_biomod2_model}(object, newdata, ...)
}
\arguments{
\item{object}{a \code{\link{biomod2_model}} object}

\item{newdata}{a \code{data.frame} or \code{\link[raster:stack]{RasterStack}} object 
containing data for new predictions}

\item{\ldots}{(\emph{optional)})}
}
\description{
This function allows the user to predict single models from 
\code{\link{biomod2_model}} on (new) explanatory variables.
}
\seealso{
\code{\link{biomod2_model}}

Other Toolbox functions: 
\code{\link{getters.bm}},
\code{\link{getters.out}},
\code{\link{load_stored_object}()},
\code{\link{predict.em}}
}
\author{
Damien Georges
}
\concept{Toolbox functions}
