% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomod2_classes_4.R
\name{predict.bm}
\alias{predict.bm}
\alias{predict.biomod2_model}
\alias{predict,biomod2_model-method}
\title{Functions to get predictions from \code{\link{biomod2_model}} objects}
\usage{
\S4method{predict}{biomod2_model}(object, newdata, ...)
}
\arguments{
\item{object}{a \code{\link{biomod2_model}} object}

\item{newdata}{a \code{data.frame} or
\code{\link[terra:SpatRaster]{SpatRaster}} object 
containing data for new predictions}

\item{\ldots}{(\emph{optional})}
}
\description{
This function allows the user to predict single models from 
\code{\link{biomod2_model}} on (new) explanatory variables.
}
\seealso{
\code{\link{biomod2_model}}

Other Toolbox functions: 
\code{\link{getters.bm}},
\code{\link{getters.out}},
\code{\link{load_stored_object}()},
\code{\link{predict.em}},
\code{\link{predict2.bm}},
\code{\link{predict2.em}}
}
\author{
Damien Georges
}
\concept{Toolbox functions}
