% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomod2_classes_5.R
\name{predict.em}
\alias{predict.em}
\alias{predict}
\alias{.predict.EMmean_biomod2_model.RasterStack}
\alias{.predict.EMmean_biomod2_model.data.frame}
\alias{.predict.EMmedian_biomod2_model.RasterStack}
\alias{.predict.EMmedian_biomod2_model.data.frame}
\alias{.predict.EMcv_biomod2_model.RasterStack}
\alias{.predict.EMcv_biomod2_model.data.frame}
\alias{.predict.EMci_biomod2_model.RasterStack}
\alias{.predict.EMci_biomod2_model.data.frame}
\alias{.predict.EMca_biomod2_model.RasterStack}
\alias{.predict.EMca_biomod2_model.data.frame}
\alias{.predict.EMwmean_biomod2_model.RasterStack}
\alias{.predict.EMwmean_biomod2_model.data.frame}
\alias{predict,EMmean_biomod2_model-method}
\alias{predict,EMmedian_biomod2_model-method}
\alias{predict,EMcv_biomod2_model-method}
\alias{predict,EMci_biomod2_model-method}
\alias{predict,EMca_biomod2_model-method}
\alias{predict,EMwmean_biomod2_model-method}
\title{Functions to get predictions from \code{\link{biomod2_ensemble_model}} objects}
\usage{
\S4method{predict}{EMmean_biomod2_model}(object, newdata = NULL, formal_predictions = NULL, ...)

\S4method{predict}{EMmedian_biomod2_model}(object, newdata = NULL, formal_predictions = NULL, ...)

\S4method{predict}{EMcv_biomod2_model}(object, newdata = NULL, formal_predictions = NULL, ...)

\S4method{predict}{EMci_biomod2_model}(object, newdata = NULL, formal_predictions = NULL, ...)

\S4method{predict}{EMca_biomod2_model}(object, newdata = NULL, formal_predictions = NULL, ...)

\S4method{predict}{EMwmean_biomod2_model}(object, newdata = NULL, formal_predictions = NULL, ...)
}
\arguments{
\item{obj}{a \code{\link{biomod2_ensemble_model}} object}
}
\description{
This function allows the user to predict single models from 
\code{\link{biomod2_ensemble_model}} on (new) explanatory variables.
}
\author{
Damien Georges
}
