% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomod2_classes_5.R
\name{predict.em}
\alias{predict.em}
\title{Functions to get predictions from \code{\link{biomod2_ensemble_model}} objects}
\arguments{
\item{object}{a \code{\link{biomod2_ensemble_model}} object}

\item{newdata}{a \code{data.frame} or \code{\link[raster:stack]{RasterStack}} object 
containing data for new predictions}

\item{\ldots}{(\emph{optional})}
}
\description{
This function allows the user to predict single models from 
\code{\link{biomod2_ensemble_model}} on (new) explanatory variables.
}
\seealso{
\code{\link{biomod2_ensemble_model}}

Other Toolbox functions: 
\code{\link{getters.bm}},
\code{\link{getters.out}},
\code{\link{load_stored_object}()},
\code{\link{predict.bm}},
\code{\link{predict2.bm}},
\code{\link{predict2.em}}
}
\author{
Damien Georges
}
\concept{Toolbox functions}
