% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomod2_classes_3.R
\name{setters}
\alias{setters}
\alias{set_newdirname}
\title{Functions to change the place of the different biomod2 objects}
\usage{
set_newdirname(obj, new.dir.name)
}
\arguments{
\item{obj}{a \code{\link{BIOMOD.formated.data}}, \code{\link{BIOMOD.formated.data.PA}}, 
\code{\link{BIOMOD.models.out}}, \code{\link{BIOMOD.projection.out}} or 
\code{\link{BIOMOD.ensemble.models.out}} object}

\item{new.dir.name}{a \code{character} corresponding to the new folder path}
}
\description{
This function allow the user to change the folder where
the modelling of biomod2 have been done.
}
\seealso{
\code{\link{BIOMOD.models.out}}, \code{\link{BIOMOD.projection.out}}, 
\code{\link{BIOMOD.ensemble.models.out}}

Other Toolbox functions: 
\code{\link{getters.bm}},
\code{\link{getters.out}},
\code{\link{load_stored_object}()},
\code{\link{predict.bm}},
\code{\link{predict.em}},
\code{\link{predict2.bm}},
\code{\link{predict2.em}}
}
\author{
Hélène Blancheteau
}
\concept{Toolbox functions}
